Crate layout

Layout. Performs layout on the DOM, builds display lists and sends them to be painted.

Modules

accessibility_tree 🔒

cell 🔒

construct_modern 🔒

Layout construction code that is shared between modern layout modes (Flexbox and CSS Grid)

context 🔒

display_list 🔒

dom 🔒

dom_traversal 🔒

flexbox 🔒

flow

Flow layout, also known as block-and-inline layout.

formatting_contexts 🔒

fragment_tree 🔒

geom

layout_box_base 🔒

layout_impl 🔒

layout_root 🔒

lists 🔒

positioned 🔒

query 🔒

Utilities for querying the layout, as needed by layout.

quotes 🔒

replaced 🔒

sizing 🔒

https://drafts.csswg.org/css-sizing/

style_ext 🔒

table

HTML Tables (╯°□°)╯︵ ┻━┻

taffy 🔒

traversal 🔒

Structs

ArcRefCell

ConstraintSpace 🔒

Represents the set of constraints that we use when computing the min-content and max-content inline sizes of an element.

ContainingBlock 🔒

ContainingBlockSize 🔒

DefiniteContainingBlock 🔒

IndefiniteContainingBlock 🔒

A variant of ContainingBlock that allows an indefinite inline size. Useful for code that is shared for both layout (where we know the inline size of the containing block) and intrinsic sizing (where we don’t know it).

LayoutFactoryImpl

PropagatedBoxTreeData 🔒

Data that is propagated from ancestors to descendants during crate::flow::BoxTree construction. This allows data to flow in the reverse direction of the typical layout propoagation, but only during BoxTree construction.

Type Aliases

SharedStyle 🔒

At times, a style is “owned” by more than one layout object. For example, text fragments need a handle on their parent inline box’s style. In order to make incremental layout easier to implement, another layer of shared ownership is added via SharedStyle. This allows updating the style in originating layout object and having all “depdendent” objects update automatically.