layout::accessibility_tree

Struct AccessibilityNode

struct AccessibilityNode {
    id: NodeId,
    accesskit_node: Node,
    parent_node: Option<WeakRefCell<AccessibilityNode>>,
    child_nodes: Vec<ArcRefCell<AccessibilityNode>>,
    opaque_node: Option<OpaqueNode>,
    scroll_offset: Option<LayoutVector2D>,
    dirty_state: DirtyState,
}

Fields
id: NodeId

The unique ID for the node. This is used both as a key in AccessibilityTree’s cache of nodes, and as an identifier in accesskit datastructures: accesskit::Nodes, accesskit::TreeUpdates and accesskit::ActionRequests.

accesskit_node: Node

The computed accesskit::Node data. This will be copied and serialized into a accesskit::TreeUpdate whenever it is changed during an update.

parent_node: Option<WeakRefCell<AccessibilityNode>>

This node’s parent, if any.

child_nodes: Vec<ArcRefCell<AccessibilityNode>>

All this node’s children.

opaque_node: Option<OpaqueNode>

The OpaqueNode for the DOM node which corresponds to this accessibility node, if any. An accessibility node may not correspond to a DOM node if it corresponds to a pseudo-element, or in a test.

scroll_offset: Option<LayoutVector2D>

This node’s scroll offset, if it is a scroll container which has scrolled. This is used to translate this node’s children.

dirty_state: DirtyState

Any dirty state for the current update.

Implementations

impl AccessibilityNode

fn new(id: NodeId) -> Self

fn new_with_role(id: NodeId, role: Role) -> Self

fn update_subtree<'update>( &mut self, ref_self: ArcRefCell<Self>, context: &AccessibilityContext<'_>, tree: &mut AccessibilityTree, update: &mut AccessibilityUpdate<'update>, ) -> LocalAccessibilityDamage

Update this node and its subtree based on damage from the DOM.

First, if this node has damage from the DOM to be resolved, update the node from the DOM tree, recursively populating any new children.

Next, recursively call this method for any children which are dirty, or have dirty descendants.

Finally, update any properties on this node which are may have changed due to other changes in the tree.

At the end of this method, both has_dirty_descendants and is_dirty should be false for this node and all its descendants.

fn update_ancestors( &self, local_damage: LocalAccessibilityDamage, update: &mut AccessibilityUpdate<'_>, )

Update each of this node’s ancestors based on changes which have already been applied in the tree.

fn update_properties_and_children_from_dom_node<'update>( &mut self, ref_self: ArcRefCell<Self>, dom_node: &ServoLayoutNode<'update>, dom_damage: AccessibilityDamage, tree: &mut AccessibilityTree, update: &mut AccessibilityUpdate<'update>, ) -> LocalAccessibilityDamage

Update the given AccessibilityNode from its corresponding DOM node and AccessibilityDamage. If it has new children, those will be created here, but not yet populated.

fn update_children_from_dom_node<'update>( &mut self, ref_self: ArcRefCell<AccessibilityNode>, dom_node: &ServoLayoutNode<'update>, _dom_damage: AccessibilityDamage, tree: &mut AccessibilityTree, update: &mut AccessibilityUpdate<'update>, ) -> LocalAccessibilityDamage

Update this node’s Self::children from its corresponding DOM node. If it has new children, those will be created here, but not yet populated.

fn update_properties_from_dom_node( &mut self, dom_node: &ServoLayoutNode<'_>, dom_damage: AccessibilityDamage, ) -> LocalAccessibilityDamage

Update this node’s properties from its corresponding DOM node.

fn update_bounds_from_dom_node( &mut self, dom_node: &ServoLayoutNode<'_>, context: &AccessibilityContext<'_>, update: &mut AccessibilityUpdate<'_>, )

Update this node’s bounds from the current layout geometry.

fn update_node_local( &mut self, local_damage: LocalAccessibilityDamage, update: &mut AccessibilityUpdate<'_>, ) -> LocalAccessibilityDamage

Update this node’s properties based on changes already made to the accessibility tree. For example, if there were nodes added or removed in its subtree, its computed text may have changed, so that will be recomputed here. If any changes are made, add this node to the given AccessibilityUpdate.

fn label_from_descendants(&self) -> Option<String>

fn print(&self, print_tree: &mut PrintTree)

fn parent(&self) -> Option<ArcRefCell<AccessibilityNode>>

fn children( &self, ) -> impl DoubleEndedIterator<Item = &ArcRefCell<AccessibilityNode>>

fn ancestors(&self) -> impl Iterator<Item = ArcRefCell<AccessibilityNode>>

fn child_ids(&self) -> &[NodeId]

fn set_scroll_offset( &mut self, offset: LayoutVector2D, update: &mut AccessibilityUpdate<'_>, )

fn role(&self) -> Role

fn set_role(&mut self, role: Role) -> LocalAccessibilityDamage

fn label(&self) -> Option<&str>

fn set_label(&mut self, label: &str) -> LocalAccessibilityDamage

fn clear_label(&mut self) -> LocalAccessibilityDamage

fn html_tag(&self) -> Option<&str>

fn set_html_tag(&mut self, html_tag: &str)

fn value(&self) -> Option<&str>

fn set_value(&mut self, value: &str) -> LocalAccessibilityDamage

fn bounds(&self) -> Option<Rect>

fn set_bounds(&mut self, bounds: Rect)

fn clear_bounds(&mut self)

fn set_transform(&mut self, transform: Affine)

fn clear_transform(&mut self)

fn assert_integrity( &self, expected_parent: Option<WeakRefCell<AccessibilityNode>>, )

fn compute_damage( &self, update: &mut AccessibilityUpdate<'_>, ) -> AccessibilityDamage