Module layout::animation

CSS transitions and animations.

Functions

process_new_animations
Processes any new animations that were discovered after style recalculation.

recalc_style_for_animation
Recalculates style for an animation. This does not run with the DOM lock held.

start_transitions_if_applicable
Inserts transitions into the queue of running animations as applicable for the given style difference. This is called from the layout worker threads.

tick_all_animations
Handles animation updates.