Function layout::animation::update_animation_state

pub fn update_animation_state<Impl: SelectorImplExt>(constellation_chan: &IpcSender<ConstellationMsg>, running_animations: &mut HashMap<OpaqueNode, Vec<Animation<Impl>>>, expired_animations: &mut HashMap<OpaqueNode, Vec<Animation<Impl>>>, new_animations_receiver: &Receiver<Animation<Impl>>, pipeline_id: PipelineId)

Processes any new animations that were discovered after style recalculation. Also expire any old animations that have completed, inserting them into expired_animations.