Struct layout::block::CandidateBSizeIterator

pub struct CandidateBSizeIterator {
    pub candidate_value: Au,
    // some fields omitted
}

Performs block-size calculations potentially multiple times, taking (assuming an horizontal writing mode) height, min-height, and max-height into account. After each call to next(), the caller must call .try() with the current calculated value of height.

See CSS 2.1  10.7.

Fields

candidate_value

Methods

impl CandidateBSizeIterator

fn new(fragment: &Fragment, block_container_block_size: Option<Au>) -> CandidateBSizeIterator

Creates a new candidate block-size iterator. block_container_block-size is None if the block-size of the block container has not been determined yet. It will always be Some in the case of absolutely-positioned containing blocks.