Enum layout::construct::ConstructionResult

pub enum ConstructionResult {
    None,
    Flow(FlowRef, AbsDescendants),
    ConstructionItem(ConstructionItem),
}

The results of flow construction for a DOM node.

Variants

None

This node contributes nothing at all (display: none). Alternately, this is what newly created nodes have their ConstructionResult set to.

Flow

This node contributed a flow at the proper position in the tree. Nothing more needs to be done for this node. It has bubbled up fixed and absolute descendant flows that have a containing block above it.

ConstructionItem

This node contributed some object or objects that will be needed to construct a proper flow later up the tree, but these objects have not yet found their home.

Methods

impl ConstructionResult

fn swap_out(&mut self) -> ConstructionResult

fn debug_id(&self) -> usize