Struct layout::construct::InlineFragmentsAccumulator

struct InlineFragmentsAccumulator {
    fragments: IntermediateInlineFragments,
    enclosing_node: Option<InlineFragmentNodeInfo>,
    restyle_damage: RestyleDamage,
    bidi_control_chars: Option<(&'static str, &'static str)>,
}

Holds inline fragments that we're gathering for children of an inline node.

Fields
fragments: IntermediateInlineFragments

The list of fragments.

enclosing_node: Option<InlineFragmentNodeInfo>

Whether we've created a range to enclose all the fragments. This will be Some() if the outer node is an inline and None otherwise.

restyle_damage: RestyleDamage

Restyle damage to use for fragments created in this node.

bidi_control_chars: Option<(&'static str, &'static str)>

Bidi control characters to insert before and after these fragments.

Methods

impl InlineFragmentsAccumulator

fn new() -> InlineFragmentsAccumulator

fn from_inline_node<N>(node: &N, style_context: &SharedStyleContext) -> InlineFragmentsAccumulator where N: ThreadSafeLayoutNode

fn push(&mut self, fragment: Fragment)

fn push_all(&mut self, fragments: IntermediateInlineFragments)

fn to_intermediate_inline_fragments(self) -> IntermediateInlineFragments