Struct layout::construct::Legalizer

struct Legalizer {
    stack: Vec<FlowRef>,
}

Maintains a stack of anonymous boxes needed to ensure that the flow tree is legal. The tree is legal if it follows the rules in CSS 2.1  17.2.1.

As an example, the legalizer makes sure that table row flows contain only table cells. If the flow constructor attempts to place, say, a block flow directly underneath the table row, the legalizer generates an anonymous table cell in between to hold the block.

Generally, the flow constructor should use Legalizer::add_child() instead of calling Flow::add_new_child() directly. This ensures that the flow tree remains legal at all times and centralizes the anonymous flow generation logic in one place.

Fields
stack: Vec<FlowRef>

A stack of anonymous flows that have yet to be finalized (i.e. that still could acquire new children).

Methods

impl Legalizer

fn new() -> Legalizer

Creates a new legalizer.

fn add_child(&mut self, stylist: &Stylist, parent: &mut FlowRef, child: FlowRef)

Makes the child flow a new child of parent. Anonymous flows are automatically inserted to keep the tree legal.

fn finish(self, parent: &mut FlowRef)

Flushes all flows we've been gathering up.

fn try_to_add_child(&mut self,
stylist: &Stylist,
parent: &mut FlowRef,
child: &mut FlowRef)
-> bool

Attempts to make child a child of parent. On success, this returns true. If this would make the tree illegal, this method does nothing and returns false.

This method attempts to create anonymous blocks in between parent and child if and only if those blocks will only ever have child as their sole child. At present, this is only true for anonymous block children of flex flows.

fn flush_top_of_stack(&mut self, parent: &mut FlowRef)

Finalizes the flow on the top of the stack.

fn push_next_anonymous_flow(&mut self, stylist: &Stylist, parent: &FlowRef)

Adds the anonymous flow that would be necessary to make an illegal child of parent legal to the stack.

fn push_new_anonymous_flow<F>(&mut self,
stylist: &Stylist,
reference: &FlowRef,
pseudos: &[PseudoElement],
specific_fragment_info: SpecificFragmentInfo,
constructor: fn(Fragment) -> F) where F: Flow

Creates an anonymous flow and pushes it onto the stack.

fn create_anonymous_flow<F>(stylist: &Stylist,
reference: &FlowRef,
pseudos: &[PseudoElement],
specific_fragment_info: SpecificFragmentInfo,
constructor: fn(Fragment) -> F)
-> FlowRef where F: Flow

Creates a new anonymous flow. The new flow is identical to reference except with all styles applying to every pseudo-element in pseudos applied.

This method invokes the supplied constructor function on the given specific fragment info in order to actually generate the flow.