Trait layout::construct::FlowConstructionUtils

pub trait FlowConstructionUtils {
    fn add_new_child(&mut self, new_child: FlowRef);
    fn finish(&mut self);
}

Required Methods

fn add_new_child(&mut self, new_child: FlowRef)

Adds a new flow as a child of this flow. Removes the flow from the given leaf set if it's present.

fn finish(&mut self)

Finishes a flow. Once a flow is finished, no more child flows or boxes may be added to it. This will normally run the bubble-inline-sizes (minimum and preferred -- i.e. intrinsic -- inline-size) calculation, unless the global bubble_inline-sizes_separately flag is on.

All flows must be finished at some point, or they will not have their intrinsic inline- sizes properly computed. (This is not, however, a memory safety problem.)