Trait layout::construct::NodeUtils

trait NodeUtils {
    fn is_replaced_content(&self) -> bool;
    fn construction_result_mut(self, layout_data: &mut PrivateLayoutData) -> &mut ConstructionResult;
    fn set_flow_construction_result(self, result: ConstructionResult);
    fn swap_out_construction_result(self) -> ConstructionResult;
}

A utility trait with some useful methods for node queries.

Required Methods

fn is_replaced_content(&self) -> bool

Returns true if this node doesn't render its kids and false otherwise.

fn construction_result_mut(self, layout_data: &mut PrivateLayoutData) -> &mut ConstructionResult

fn set_flow_construction_result(self, result: ConstructionResult)

Sets the construction result of a flow.

fn swap_out_construction_result(self) -> ConstructionResult

Replaces the flow construction result in a node with ConstructionResult::None and returns the old value.