Struct layout::context::SharedLayoutContext

pub struct SharedLayoutContext {
    pub image_cache: Arc<Mutex<LocalImageCache<UntrustedNodeAddress>>>,
    pub screen_size: Size2D<Au>,
    pub screen_size_changed: bool,
    pub constellation_chan: ConstellationChan,
    pub layout_chan: LayoutChan,
    pub font_cache_task: FontCacheTask,
    pub stylist: *const Stylist,
    pub reflow_root: Option<OpaqueNode>,
    pub url: Url,
    pub dirty: Rect<Au>,
    pub generation: u32,
    pub new_animations_sender: Sender<Animation>,
}

Layout information shared among all workers. This must be thread-safe.

Fields

image_cache

The local image cache.

screen_size

The current screen size.

screen_size_changed

Screen sized changed?

constellation_chan

A channel up to the constellation.

layout_chan

A channel up to the layout task.

font_cache_task

Interface to the font cache task.

stylist

The CSS selector stylist.

FIXME(#2604): Make this no longer an unsafe pointer once we have fast RWArcs.

reflow_root

The root node at which we're starting the layout.

url

The URL.

dirty

The dirty rectangle, used during display list building.

generation

Starts at zero, and increased by one every time a layout completes. This can be used to easily check for invalid stale data.

new_animations_sender

A channel on which new animations that have been triggered by style recalculation can be sent.