Enum layout::css::matching::StyleSharingResult

pub enum StyleSharingResult {
    CannotShare(bool),
    StyleWasShared(usize, RestyleDamage),
}

The results of attempting to share a style.

Variants

CannotShare

We didn't find anybody to share the style with. The boolean indicates whether the style is shareable at all.

StyleWasShared

The node's style can be shared. The integer specifies the index in the LRU cache that was hit and the damage that was done.