Trait layout::css::matching::MatchMethods

pub trait MatchMethods {
    fn insert_into_bloom_filter(&self, bf: &mut BloomFilter);
    fn remove_from_bloom_filter(&self, bf: &mut BloomFilter);
    fn match_node(&self, stylist: &Stylist, parent_bf: &Option<Box<BloomFilter>>, applicable_declarations: &mut ApplicableDeclarations, shareable: &mut bool);
    unsafe fn share_style_if_possible(&self, style_sharing_candidate_cache: &mut StyleSharingCandidateCache, parent: Option<LayoutNode>) -> StyleSharingResult;
    unsafe fn cascade_node(&self, layout_context: &SharedLayoutContext, parent: Option<LayoutNode>, applicable_declarations: &ApplicableDeclarations, applicable_declarations_cache: &mut ApplicableDeclarationsCache, new_animations_sender: &Sender<Animation>);
}

Required Methods

fn insert_into_bloom_filter(&self, bf: &mut BloomFilter)

Inserts and removes the matching Descendant selectors from a bloom filter. This is used to speed up CSS selector matching to remove unnecessary tree climbs for Descendant queries.

A bloom filter of the local names, namespaces, IDs, and classes is kept. Therefore, each node must have its matching selectors inserted after its own selector matching and before its children start.

fn remove_from_bloom_filter(&self, bf: &mut BloomFilter)

After all the children are done css selector matching, this must be called to reset the bloom filter after an insert.

fn match_node(&self, stylist: &Stylist, parent_bf: &Option<Box<BloomFilter>>, applicable_declarations: &mut ApplicableDeclarations, shareable: &mut bool)

unsafe fn share_style_if_possible(&self, style_sharing_candidate_cache: &mut StyleSharingCandidateCache, parent: Option<LayoutNode>) -> StyleSharingResult

Attempts to share a style with another node. This method is unsafe because it depends on the style_sharing_candidate_cache having only live nodes in it, and we have no way to guarantee that at the type system level yet.

unsafe fn cascade_node(&self, layout_context: &SharedLayoutContext, parent: Option<LayoutNode>, applicable_declarations: &ApplicableDeclarations, applicable_declarations_cache: &mut ApplicableDeclarationsCache, new_animations_sender: &Sender<Animation>)