Trait layout::css::node_style::StyledNode

pub trait StyledNode {
    fn style<'a>(&'a self) -> &'a Arc<ComputedValues>;
    fn has_style(&self) -> bool;
    fn unstyle(self);
}

Node mixin providing style method that returns a NodeStyle

Required Methods

fn style<'a>(&'a self) -> &'a Arc<ComputedValues>

Returns the style results for the given node. If CSS selector matching has not yet been performed, fails.

fn has_style(&self) -> bool

Does this node have a computed style yet?

fn unstyle(self)

Removes the style from this node.