Struct layout::data::PersistentLayoutData

pub struct PersistentLayoutData {
    pub base: PartialPersistentLayoutData,
    pub flow_construction_result: ConstructionResult,
    pub before_flow_construction_result: ConstructionResult,
    pub after_flow_construction_result: ConstructionResult,
    pub details_summary_flow_construction_result: ConstructionResult,
    pub details_content_flow_construction_result: ConstructionResult,
    pub flags: LayoutDataFlags,
}

Data that layout associates with a node.

Fields
base: PartialPersistentLayoutData

Data accessed by script_layout_interface. This must be first to allow casting between PersistentLayoutData and PartialPersistentLayoutData.

flow_construction_result: ConstructionResult

The current results of flow construction for this node. This is either a flow or a ConstructionItem. See comments in construct.rs for more details.

before_flow_construction_result: ConstructionResult after_flow_construction_result: ConstructionResult details_summary_flow_construction_result: ConstructionResult details_content_flow_construction_result: ConstructionResult flags: LayoutDataFlags

Various flags.

Methods

impl PersistentLayoutData

fn new() -> PersistentLayoutData

Creates new layout data.