Struct layout::data::PrivateLayoutData

pub struct PrivateLayoutData {
    pub before_style: Option<Arc<ComputedValues>>,
    pub after_style: Option<Arc<ComputedValues>>,
    pub restyle_damage: RestyleDamage,
    pub flow_construction_result: ConstructionResult,
    pub before_flow_construction_result: ConstructionResult,
    pub after_flow_construction_result: ConstructionResult,
    pub parallel: DomParallelInfo,
    pub flags: LayoutDataFlags,
}

Data that layout associates with a node.

Fields

before_style

The results of CSS styling for this node's before pseudo-element, if any.

after_style

The results of CSS styling for this node's after pseudo-element, if any.

restyle_damage

Description of how to account for recent style changes.

flow_construction_result

The current results of flow construction for this node. This is either a flow or a ConstructionItem. See comments in construct.rs for more details.

before_flow_construction_result

after_flow_construction_result

parallel

Information needed during parallel traversals.

flags

Various flags.

Methods

impl PrivateLayoutData

fn new() -> PrivateLayoutData

Creates new layout data.