layout

Module display_list

Modules

background 🔒

clip 🔒

conversions 🔒

gradient 🔒

hit_test 🔒

paint_timing_handler 🔒

paint_traversal 🔒

stacking_context 🔒

Structs

BuilderForBoxFragment 🔒

DisplayListBuilder 🔒

HighlightTraversalState 🔒

InspectorHighlight 🔒

Constants

INSERTION_POINT_LOGICAL_WIDTH 🔒

Functions

compute_margin_box_radius 🔒

https://drafts.csswg.org/css-shapes-1/#valdef-shape-box-margin-box

glyphs 🔒

normalize_radii 🔒

offset_radii 🔒

Given a set of corner radii for a rectangle, this function returns the corresponding radii for the outer rectangle resulting from expanding the original rectangle by the given offsets.

resolve_border_image_outset 🔒

Resolve the WebRender border-image outset area from the style values.

resolve_border_image_slice 🔒

Resolve the WebRender border-image slice from the style values.

resolve_border_image_width 🔒

Resolve the WebRender border-image width from the style values.

rgba 🔒