layout::display_list

Module background

Structs

BackgroundLayer 🔒

BackgroundPainter 🔒

Layout1DResult 🔒

Functions

get_cyclic 🔒

layout_1d 🔒

Abstract over the horizontal or vertical dimension Coordinates (0, 0) for the purpose of this function are the positioning area’s origin.

layout_layer 🔒