Module layout::display_list::border

Functions

corner_radius
Computes a border radius size against the containing size.

image_outset
Compute the additional border-image area.

image_slice

image_width

inner_radii
Calculates radii for the inner side.

overlapping_radii

radii
Determine the four corner radii of a border.

resolve_percentage

scaled_radii

side_image_outset

side_image_width

simple
Creates a four-sided border with square corners and uniform color and width.