Function layout::display_list::border::inner_radii

pub fn inner_radii(
    radii: BorderRadius, 
    offsets: UntypedSideOffsets2D<Au>
) -> BorderRadius

Calculates radii for the inner side.

Radii usually describe the outer side of a border but for the lines to look nice the inner radii need to be smaller depending on the line width.

This is used to determine clipping areas.