Struct layout::display_list::builder::DisplayListBuildState

pub struct DisplayListBuildState<'a> {
    pub layout_context: &'a LayoutContext<'a>,
    pub root_stacking_context: StackingContext,
    stacking_context_info: FnvHashMap<StackingContextId, StackingContextInfo>,
    pub clip_scroll_nodes: Vec<ClipScrollNode>,
    pub items: FnvHashMap<StackingContextId, Vec<DisplayItem>>,
    pub processing_scrolling_overflow_element: bool,
    pub current_stacking_context_id: StackingContextId,
    pub current_clipping_and_scrolling: ClippingAndScrolling,
    pub iframe_sizes: Vec<IFrameSize>,
    pub indexable_text: IndexableText,
}

Fields
layout_context: &'a LayoutContext<'a>

A LayoutContext reference important for creating WebRender images.

root_stacking_context: StackingContext

The root of the StackingContext tree.

stacking_context_info: FnvHashMap<StackingContextId, StackingContextInfo>

StackingContext and ClipScrollNode children for each StackingContext.

clip_scroll_nodes: Vec<ClipScrollNode>

A vector of ClipScrollNodes which will be given ids during WebRender DL conversion.

items: FnvHashMap<StackingContextId, Vec<DisplayItem>>

The items in this display list.

processing_scrolling_overflow_element: bool

Whether or not we are processing an element that establishes scrolling overflow. Used to determine what ClipScrollNode to place backgrounds and borders into.

current_stacking_context_id: StackingContextId

The current stacking context id, used to keep track of state when building. recursively building and processing the display list.

current_clipping_and_scrolling: ClippingAndScrolling

The current clip and scroll info, used to keep track of state when recursively building and processing the display list.

iframe_sizes: Vec<IFrameSize>

Vector containing iframe sizes, used to inform the constellation about new iframe sizes

indexable_text: IndexableText

Stores text runs to answer text queries used to place a cursor inside text.

Implementations

impl<'a> DisplayListBuildState<'a>

pub fn new(
layout_context: &'a LayoutContext,
state: StackingContextCollectionState
) -> DisplayListBuildState<'a>

pub fn add_display_item(&mut self, display_item: DisplayItem)

fn add_image_item(&mut self, base: BaseDisplayItem, item: ImageDisplayItem)

fn parent_clip_scroll_node_index(
&self,
index: ClipScrollNodeIndex
) -> ClipScrollNodeIndex

fn is_background_or_border_of_clip_scroll_node(
&self,
section: DisplayListSection
) -> bool

pub fn create_base_display_item(
&self,
clip_rect: Rect<Au>,
node: OpaqueNode,
cursor: Option<Cursor>,
section: DisplayListSection
) -> BaseDisplayItem

fn create_base_display_item_with_clipping_and_scrolling(
&self,
clip_rect: Rect<Au>,
node: OpaqueNode,
cursor: Option<Cursor>,
section: DisplayListSection,
clipping_and_scrolling: ClippingAndScrolling
) -> BaseDisplayItem

fn add_late_clip_node(
&mut self,
rect: LayoutRect,
radii: BorderRadius
) -> ClipScrollNodeIndex

pub fn to_display_list(self) -> DisplayList

fn to_display_list_for_stacking_context(
&mut self,
list: &mut Vec<DisplayItem>,
stacking_context: StackingContext
)

fn to_display_list_for_items(
&mut self,
list: &mut Vec<DisplayItem>,
child_items: Vec<DisplayItem>,
child_stacking_contexts: Vec<StackingContext>
)

fn clipping_and_scrolling_scope<R, F: FnOnce(&mut Self) -> R>(
&mut self,
function: F
) -> R