layout::display_list

Module clip

Structs

Clip 🔒

All the information needed to create a clip on a WebRender display list. These are created at two times: during StackingContextTree creation and during WebRender display list construction. Only the former are stored in a [ClipStore].

ClipId 🔒

An identifier for a clip used during StackingContextTree construction. This is a simple index in a [ClipStore]s vector of clips.

StackingContextTreeClipStore 🔒

A simple vector of Clip that is built during StackingContextTree construction. These are later turned into WebRender clips and clip chains during WebRender display list construction.

Functions

compute_shape_radius 🔒