layout::display_list

Function offset_radii

fn offset_radii(radii: BorderRadius, offsets: LayoutSideOffsets) -> BorderRadius

Given a set of corner radii for a rectangle, this function returns the corresponding radii for the outer rectangle resulting from expanding the original rectangle by the given offsets.