Module layout::display_list::gradient

Structs

StopRun
A helper data structure for gradients.

Functions

circle_size_keyword
Determines the radius of a circle if it was not explictly provided. https://drafts.csswg.org/css-images-3/#typedef-size

convert_gradient_stops

distance_to_corner
Returns the the distance to the nearest or farthest corner depending on the comperator.

distance_to_sides
Returns the distance to the nearest or farthest sides depending on the comparator.

ellipse_radius
Returns the radius for an ellipse with the same ratio as if it was matched to the sides.

ellipse_size_keyword
Determines the radius of an ellipse if it was not explictly provided. https://drafts.csswg.org/css-images-3/#typedef-size

extend_mode

linear

position_to_offset

radial