layout::display_list::paint_traversal

Struct TraversalState

pub(crate) struct TraversalState {
    pub spatial_id: ScrollTreeNodeId,
    pub clip_id: ClipId,
    pub origin: PhysicalPoint<Au>,
    pub text_decorations: Rc<Vec<FragmentTextDecoration>>,
    pub containing_element_tag: Option<Tag>,
}

Fields
spatial_id: ScrollTreeNodeIdclip_id: ClipIdorigin: PhysicalPoint<Au>text_decorations: Rc<Vec<FragmentTextDecoration>>containing_element_tag: Option<Tag>

The tag of the nearest ancestor box fragment that has a tag. Used for text LCP candidate grouping — all text fragments within a single element are unioned before computing effective visual size.

Implementations

impl TraversalState

pub(crate) fn push_box_fragment( &self, box_fragment: &BoxFragmentWithStyle<'_>, ) -> Self

pub(crate) fn without_text_decorations(&self) -> Self

pub(crate) fn push_positioning_fragment( &self, positioning_fragment: &PositioningFragment, ) -> Self

pub(crate) fn push_stacking_context( &self, stacking_context: &StackingContext, ) -> Self