layout::display_list

Module stacking_context

Structs

ContainingBlock 🔒

FragmentTextDecoration 🔒

The text decorations for a Fragment, collecting during StackingContextTree construction.

OverflowFrameData 🔒

ReferenceFrameData 🔒

ScrollFrameData 🔒

StackingContext

Either a stacking context or a stacking container, per the definitions in https://drafts.csswg.org/css-position-4/#painting-order.

StackingContextReferenceFrameInfo 🔒

StackingContextTree 🔒

Enums

StackingContextBuildMode 🔒

StackingContextFragments 🔒

StackingContextType 🔒

Functions

au_rect_to_length_rect 🔒

Type Aliases

ContainingBlockInfo 🔒