layout::display_list::stacking_context

Struct StackingContextTree

pub(crate) struct StackingContextTree {
    pub root_stacking_context: StackingContext,
    pub paint_info: PaintDisplayListInfo,
    pub clip_store: StackingContextTreeClipStore,
}

Fields
root_stacking_context: StackingContext

The root StackingContext of this StackingContextTree.

paint_info: PaintDisplayListInfo

The information about the WebRender display list that Paint consumes. This currently contains the out-of-band hit testing information data structure that Paint uses to map hit tests to information about the item hit.

clip_store: StackingContextTreeClipStore

All of the clips collected for this StackingContextTree. These are added for things like overflow. More clips may be created later during WebRender display list construction, but they are never added here.

Implementations

impl StackingContextTree

pub fn new( fragment_tree: &FragmentTree, viewport_details: ViewportDetails, pipeline_id: PipelineId, first_reflow: bool, debug: &DiagnosticsLogging, ) -> Self

Create a new [DisplayList] given the dimensions of the layout and the WebRender pipeline id.

fn push_reference_frame( &mut self, origin: LayoutPoint, frame_origin_for_query: LayoutPoint, parent_scroll_node_id: ScrollTreeNodeId, transform_style: TransformStyle, transform: LayoutTransform, kind: ReferenceFrameKind, ) -> ScrollTreeNodeId

fn define_scroll_frame( &mut self, parent_scroll_node_id: ScrollTreeNodeId, external_id: ExternalScrollId, content_rect: LayoutRect, clip_rect: LayoutRect, scroll_sensitivity: AxesScrollSensitivity, touch_action: TouchAction, ) -> ScrollTreeNodeId

fn define_sticky_frame( &mut self, parent_scroll_node_id: ScrollTreeNodeId, frame_rect: LayoutRect, margins: SideOffsets2D<Option<f32>, LayoutPixel>, vertical_offset_bounds: StickyOffsetBounds, horizontal_offset_bounds: StickyOffsetBounds, ) -> ScrollTreeNodeId

pub(crate) fn offset_in_fragment( &self, fragment: &Fragment, point_in_viewport: PhysicalPoint<Au>, ) -> Option<Point2D<Au, CSSPixel>>

Given a Fragment and a point in the viewport of the page, return the point in the Fragment’s content rectangle in its transformed coordinate system (untransformed CSS pixels). Note that the point may be outside the Fragment’s boundaries.

TODO: Currently, this only works for BoxFragment, but we should extend it to other types of Fragments in the future.