layout::display_list

Struct DisplayListBuilder

pub(crate) struct DisplayListBuilder<'a> {
    fragment_tree: &'a FragmentTree,
    current_reference_frame_scroll_node_id: ScrollTreeNodeId,
    pub webrender_display_list_builder: &'a mut DisplayListBuilder,
    pub paint_info: &'a mut PaintDisplayListInfo,
    inspector_highlight: Option<InspectorHighlight>,
    paint_body_background: bool,
    clip_map: Vec<ClipChainId>,
    image_resolver: Arc<ImageResolver>,
    device_pixel_ratio: Scale<f32, CSSPixel, DevicePixel>,
    paint_timing_handler: &'a mut PaintTimingHandler,
    reflow_statistics: &'a mut ReflowStatistics,
}

Fields
fragment_tree: &'a FragmentTree

The FragmentTree that we are building a display list for.

current_reference_frame_scroll_node_id: ScrollTreeNodeId

The current ScrollTreeNodeId for this DisplayListBuilder. This is necessary because some pieces of fragments as backgrounds with background-attachment: fixed need to not scroll while the rest of the fragment does.

webrender_display_list_builder: &'a mut DisplayListBuilder

The wr::DisplayListBuilder for this Servo DisplayListBuilder.

paint_info: &'a mut PaintDisplayListInfo

The PaintDisplayListInfo used to collect display list items and metadata.

inspector_highlight: Option<InspectorHighlight>

Data about the fragments that are highlighted by the inspector, if any.

This data is collected during the traversal of the fragment tree and used to paint the highlight at the very end.

paint_body_background: bool

Whether or not the <body> element should be painted. This is false if the root <html> element inherits the <body>’s background to paint the page canvas background. See https://drafts.csswg.org/css-backgrounds/#body-background.

clip_map: Vec<ClipChainId>

A mapping from ClipId To WebRender ClipChainId used when building this WebRender display list.

image_resolver: Arc<ImageResolver>

An ImageResolver to use during display list construction.

device_pixel_ratio: Scale<f32, CSSPixel, DevicePixel>

The device pixel ratio used for this Document’s display list.

paint_timing_handler: &'a mut PaintTimingHandler

Handler for all Paint Timings

reflow_statistics: &'a mut ReflowStatistics

Statistics collected about the reflow, in order to write tests for incremental layout.

Implementations

impl DisplayListBuilder<'_>

pub(crate) fn build( stacking_context_tree: &mut StackingContextTree, fragment_tree: &FragmentTree, image_resolver: Arc<ImageResolver>, device_pixel_ratio: Scale<f32, StyloCSSPixel, StyloDevicePixel>, highlighted_dom_node: Option<OpaqueNode>, debug: &DiagnosticsLogging, paint_timing_handler: &mut PaintTimingHandler, reflow_statistics: &mut ReflowStatistics, ) -> BuiltDisplayList

fn wr(&mut self) -> &mut DisplayListBuilder

fn pipeline_id(&self) -> PipelineId

fn mark_is_paintable(&mut self)

fn mark_is_contentful(&mut self)

fn spatial_id(&self, id: ScrollTreeNodeId) -> SpatialId

fn clip_chain_id(&self, id: ClipId) -> ClipChainId

pub(crate) fn add_all_spatial_nodes(&mut self)

pub(crate) fn add_clip_to_display_list(&mut self, clip: &Clip) -> ClipChainId

Add the given Clip to the WebRender display list and create a mapping from its ClipId to a WebRender ClipChainId. This happens:

When WebRender display list construction starts: All clips created during the StackingContextTree construction are added in one batch. These clips are used for things such as overflow: scroll elements.

When a clip is added during WebRender display list construction for individual items. In that case, this is called by Self::maybe_create_clip.

fn maybe_create_clip( &mut self, state: &TraversalState, radii: BorderRadius, rect: LayoutRect, force_clip_creation: bool, ) -> Option<ClipChainId>

Add a new clip to the WebRender display list being built. This only happens during WebRender display list building and these clips should be added after all clips from the StackingContextTree have already been processed.

fn push_webrender_stacking_context_if_necessary( &mut self, stacking_context: &StackingContext, ) -> bool

fn common_properties( &self, state: &TraversalState, clip_rect: LayoutRect, style: &ComputedValues, ) -> CommonItemProperties

fn paint_dom_inspector_highlight(&mut self)

Draw highlights around the node that is currently hovered in the devtools.

fn check_if_paintable( &mut self, bounds: LayoutRect, clip_rect: LayoutRect, opacity: f32, )

fn collect_image_record( &mut self, state: &TraversalState, bounds: LayoutRect, clip_rect: LayoutRect, tag: Option<Tag>, url: Option<ServoUrl>, natural_width: Option<Au>, natural_height: Option<Au>, )

fn visit_stacking_context_reference_frame_info( &mut self, stacking_context: &StackingContext, ) -> (usize, Option<ScrollTreeNodeId>)