Module layout::display_list_builder

Builds display lists from flows and fragments.

Other browser engines sometimes call this "painting", but it is more accurately called display list building, as the actual painting does not happen here—only deciding what we're going to paint.

Enums

BackgroundAndBorderLevel
Which level to place backgrounds and borders in.

DisplayListBuildingResult
The results of display list building for a single flow.

StackingLevel
"Steps" as defined by CSS 2.1  E.2.

Traits

BlockFlowDisplayListBuilding

FragmentDisplayListBuilding

InlineFlowDisplayListBuilding

ListItemFlowDisplayListBuilding

ToGfxColor
Allows a CSS color to be converted into a graphics color.