Enum layout::display_list_builder::DisplayListBuildingResult

pub enum DisplayListBuildingResult {
    None,
    StackingContext(Arc<StackingContext>),
    Normal(Box<DisplayList>),
}

The results of display list building for a single flow.

Variants

None

StackingContext

Normal

Methods

impl DisplayListBuildingResult

fn add_to(&mut self, display_list: &mut DisplayList)

Adds the display list items contained within this display list building result to the given display list, preserving stacking order. If this display list building result does not consist of an entire stacking context, it will be emptied.