Trait layout::display_list_builder::BlockFlowDisplayListBuilding

pub trait BlockFlowDisplayListBuilding {
    fn collect_stacking_contexts_for_block(&mut self,
                                       parent: &mut StackingContext,
                                       parent_scroll_root_id: ScrollRootId);
    fn build_display_list_for_block(&mut self,
                                state: &mut DisplayListBuildState,
                                border_painting_mode: BorderPaintingMode);
    fn switch_coordinate_system_if_necessary(&mut self);
}

Required Methods

fn collect_stacking_contexts_for_block(&mut self,
parent: &mut StackingContext,
parent_scroll_root_id: ScrollRootId)

fn build_display_list_for_block(&mut self,
state: &mut DisplayListBuildState,
border_painting_mode: BorderPaintingMode)

fn switch_coordinate_system_if_necessary(&mut self)

Changes this block's clipping region from its parent's coordinate system to its own coordinate system if necessary (i.e. if this block is a stacking context).

The clipping region is initially in each block's parent's coordinate system because the parent of each block does not have enough information to determine what the child's coordinate system is on its own. Specifically, if the child is absolutely positioned, the parent does not know where the child's absolute position is at the time it assigns clipping regions, because flows compute their own absolute positions.