Trait layout::display_list_builder::FragmentDisplayListBuilding

pub trait FragmentDisplayListBuilding {
    fn build_display_list_for_background_if_applicable(&self,
                                                   state: &mut DisplayListBuildState,
                                                   style: &ServoComputedValues,
                                                   display_list_section: DisplayListSection,
                                                   absolute_bounds: &Rect<Au>,
                                                   clip: &ClippingRegion);
    fn compute_background_image_size(&self,
                                 style: &ServoComputedValues,
                                 bounds: &Rect<Au>,
                                 image: &WebRenderImageInfo,
                                 index: usize)
                                 -> Size2D<Au>;
    fn build_display_list_for_background_image(&self,
                                           state: &mut DisplayListBuildState,
                                           style: &ServoComputedValues,
                                           display_list_section: DisplayListSection,
                                           absolute_bounds: &Rect<Au>,
                                           clip: &ClippingRegion,
                                           image_url: &Url,
                                           background_index: usize);
    fn build_display_list_for_background_gradient(&self,
                                              state: &mut DisplayListBuildState,
                                              display_list_section: DisplayListSection,
                                              absolute_bounds: &Rect<Au>,
                                              clip: &ClippingRegion,
                                              gradient: &Gradient,
                                              style: &ServoComputedValues);
    fn build_display_list_for_borders_if_applicable(&self,
                                                state: &mut DisplayListBuildState,
                                                style: &ServoComputedValues,
                                                border_painting_mode: BorderPaintingMode,
                                                bounds: &Rect<Au>,
                                                display_list_section: DisplayListSection,
                                                clip: &ClippingRegion);
    fn build_display_list_for_outline_if_applicable(&self,
                                                state: &mut DisplayListBuildState,
                                                style: &ServoComputedValues,
                                                bounds: &Rect<Au>,
                                                clip: &ClippingRegion);
    fn build_display_list_for_box_shadow_if_applicable(&self,
                                                   state: &mut DisplayListBuildState,
                                                   style: &ServoComputedValues,
                                                   display_list_section: DisplayListSection,
                                                   absolute_bounds: &Rect<Au>,
                                                   clip: &ClippingRegion);
    fn build_debug_borders_around_text_fragments(&self,
                                             state: &mut DisplayListBuildState,
                                             style: &ServoComputedValues,
                                             stacking_relative_border_box: &Rect<Au>,
                                             stacking_relative_content_box: &Rect<Au>,
                                             text_fragment: &ScannedTextFragmentInfo,
                                             clip: &ClippingRegion);
    fn build_debug_borders_around_fragment(&self,
                                       state: &mut DisplayListBuildState,
                                       stacking_relative_border_box: &Rect<Au>,
                                       clip: &ClippingRegion);
    fn build_display_list(&mut self,
                      state: &mut DisplayListBuildState,
                      stacking_relative_flow_origin: &Point2D<Au>,
                      relative_containing_block_size: &LogicalSize<Au>,
                      relative_containing_block_mode: WritingMode,
                      border_painting_mode: BorderPaintingMode,
                      display_list_section: DisplayListSection,
                      clip: &ClippingRegion);
    fn adjust_clipping_region_for_children(&self,
                                       current_clip: &mut ClippingRegion,
                                       stacking_relative_border_box: &Rect<Au>);
    fn adjust_clip_for_style(&self,
                         parent_clip: &mut ClippingRegion,
                         stacking_relative_border_box: &Rect<Au>);
    fn build_display_items_for_selection_if_necessary(&self,
                                                  state: &mut DisplayListBuildState,
                                                  stacking_relative_border_box: &Rect<Au>,
                                                  display_list_section: DisplayListSection,
                                                  clip: &ClippingRegion);
    fn build_display_list_for_text_fragment(&self,
                                        state: &mut DisplayListBuildState,
                                        text_fragment: &ScannedTextFragmentInfo,
                                        stacking_relative_content_box: &Rect<Au>,
                                        text_shadow: Option<&TextShadow>,
                                        clip: &ClippingRegion);
    fn build_display_list_for_text_decoration(&self,
                                          state: &mut DisplayListBuildState,
                                          color: &RGBA,
                                          stacking_relative_box: &LogicalRect<Au>,
                                          clip: &ClippingRegion,
                                          blur_radius: Au);
    fn build_fragment_type_specific_display_items(&mut self,
                                              state: &mut DisplayListBuildState,
                                              stacking_relative_border_box: &Rect<Au>,
                                              clip: &ClippingRegion);
    fn create_stacking_context(&self,
                           id: StackingContextId,
                           base_flow: &BaseFlow,
                           scroll_policy: ScrollPolicy,
                           mode: StackingContextCreationMode,
                           parent_scroll_id: ScrollRootId)
                           -> StackingContext;
    fn transform_matrix(&self,
                    stacking_relative_border_box: &Rect<Au>)
                    -> Matrix4D<f32>;
}

Required Methods

fn build_display_list_for_background_if_applicable(&self,
state: &mut DisplayListBuildState,
style: &ServoComputedValues,
display_list_section: DisplayListSection,
absolute_bounds: &Rect<Au>,
clip: &ClippingRegion)

Adds the display items necessary to paint the background of this fragment to the display list if necessary.

fn compute_background_image_size(&self,
style: &ServoComputedValues,
bounds: &Rect<Au>,
image: &WebRenderImageInfo,
index: usize)
-> Size2D<Au>

Computes the background size for an image with the given background area according to the rules in CSS-BACKGROUNDS  3.9.

fn build_display_list_for_background_image(&self,
state: &mut DisplayListBuildState,
style: &ServoComputedValues,
display_list_section: DisplayListSection,
absolute_bounds: &Rect<Au>,
clip: &ClippingRegion,
image_url: &Url,
background_index: usize)

Adds the display items necessary to paint the background image of this fragment to the appropriate section of the display list.

fn build_display_list_for_background_gradient(&self,
state: &mut DisplayListBuildState,
display_list_section: DisplayListSection,
absolute_bounds: &Rect<Au>,
clip: &ClippingRegion,
gradient: &Gradient,
style: &ServoComputedValues)

Adds the display items necessary to paint the background linear gradient of this fragment to the appropriate section of the display list.

fn build_display_list_for_borders_if_applicable(&self,
state: &mut DisplayListBuildState,
style: &ServoComputedValues,
border_painting_mode: BorderPaintingMode,
bounds: &Rect<Au>,
display_list_section: DisplayListSection,
clip: &ClippingRegion)

Adds the display items necessary to paint the borders of this fragment to a display list if necessary.

fn build_display_list_for_outline_if_applicable(&self,
state: &mut DisplayListBuildState,
style: &ServoComputedValues,
bounds: &Rect<Au>,
clip: &ClippingRegion)

Adds the display items necessary to paint the outline of this fragment to the display list if necessary.

fn build_display_list_for_box_shadow_if_applicable(&self,
state: &mut DisplayListBuildState,
style: &ServoComputedValues,
display_list_section: DisplayListSection,
absolute_bounds: &Rect<Au>,
clip: &ClippingRegion)

Adds the display items necessary to paint the box shadow of this fragment to the display list if necessary.

fn build_debug_borders_around_text_fragments(&self,
state: &mut DisplayListBuildState,
style: &ServoComputedValues,
stacking_relative_border_box: &Rect<Au>,
stacking_relative_content_box: &Rect<Au>,
text_fragment: &ScannedTextFragmentInfo,
clip: &ClippingRegion)

Adds display items necessary to draw debug boxes around a scanned text fragment.

fn build_debug_borders_around_fragment(&self,
state: &mut DisplayListBuildState,
stacking_relative_border_box: &Rect<Au>,
clip: &ClippingRegion)

Adds display items necessary to draw debug boxes around this fragment.

fn build_display_list(&mut self,
state: &mut DisplayListBuildState,
stacking_relative_flow_origin: &Point2D<Au>,
relative_containing_block_size: &LogicalSize<Au>,
relative_containing_block_mode: WritingMode,
border_painting_mode: BorderPaintingMode,
display_list_section: DisplayListSection,
clip: &ClippingRegion)

Adds the display items for this fragment to the given display list.

Arguments:

state: The display building state, including the display list currently under construction and other metadata useful for constructing it.

dirty: The dirty rectangle in the coordinate system of the owning flow.

stacking_relative_flow_origin: Position of the origin of the owning flow with respect to its nearest ancestor stacking context.

relative_containing_block_size: The size of the containing block that position: relative makes use of.

clip: The region to clip the display items to.

fn adjust_clipping_region_for_children(&self,
current_clip: &mut ClippingRegion,
stacking_relative_border_box: &Rect<Au>)

Adjusts the clipping region for all descendants of this fragment as appropriate.

fn adjust_clip_for_style(&self,
parent_clip: &mut ClippingRegion,
stacking_relative_border_box: &Rect<Au>)

Adjusts the clipping rectangle for a fragment to take the clip property into account per CSS 2.1  11.1.2.

fn build_display_items_for_selection_if_necessary(&self,
state: &mut DisplayListBuildState,
stacking_relative_border_box: &Rect<Au>,
display_list_section: DisplayListSection,
clip: &ClippingRegion)

Builds the display items necessary to paint the selection and/or caret for this fragment, if any.

fn build_display_list_for_text_fragment(&self,
state: &mut DisplayListBuildState,
text_fragment: &ScannedTextFragmentInfo,
stacking_relative_content_box: &Rect<Au>,
text_shadow: Option<&TextShadow>,
clip: &ClippingRegion)

Creates the text display item for one text fragment. This can be called multiple times for one fragment if there are text shadows.

text_shadow will be Some if this is rendering a shadow.

fn build_display_list_for_text_decoration(&self,
state: &mut DisplayListBuildState,
color: &RGBA,
stacking_relative_box: &LogicalRect<Au>,
clip: &ClippingRegion,
blur_radius: Au)

Creates the display item for a text decoration: underline, overline, or line-through.

fn build_fragment_type_specific_display_items(&mut self,
state: &mut DisplayListBuildState,
stacking_relative_border_box: &Rect<Au>,
clip: &ClippingRegion)

A helper method that build_display_list calls to create per-fragment-type display items.

fn create_stacking_context(&self,
id: StackingContextId,
base_flow: &BaseFlow,
scroll_policy: ScrollPolicy,
mode: StackingContextCreationMode,
parent_scroll_id: ScrollRootId)
-> StackingContext

Creates a stacking context for associated fragment.

fn transform_matrix(&self,
stacking_relative_border_box: &Rect<Au>)
-> Matrix4D<f32>

Returns the 4D matrix representing this fragment's transform.