layout

Module dom_traversal

Structs

NodeAndStyleInfo 🔒

A data structure used to pass and store related layout information together to avoid having to repeat the same arguments in argument lists.

Enums

BoxTreeString 🔒

A reference to a string encountered during box tree construction. This can either be a reference to a borrowed DOM string, a &str or an owned String.

Contents 🔒

NonReplacedContents 🔒

PseudoElementContentItem 🔒

Traits

TraversalHandler 🔒

Functions

generate_pseudo_element_content 🔒

https://www.w3.org/TR/CSS2/generate.html#propdef-content

get_quote_from_pair 🔒

traverse_children_of 🔒

traverse_eager_pseudo_element 🔒

traverse_element 🔒

traverse_pseudo_element_contents 🔒