Module layout::flex

Layout for elements with a CSS display property of flex.

Structs

FlexFlow

A block with the CSS display property equal to flex.

FlexItem

Represents a child in a flex container. Most fields here are used in flex size resolving, and items are sorted by the ‘order’ property.

FlexLine

A line in a flex container.

Enums

AxisSize

The size of an axis. May be a specified size, a min/max constraint, or an unlimited size

Functions

from_flex_basis

This function accepts the flex-basis and the size property in main direction from style, and the container size, then return the used value of flex basis. it can be used to help determining the flex base size and to indicate whether the main size of the item is definite after flex size resolving.