Enum layout::flex::AxisSize

enum AxisSize {
    Definite(Au),
    MinMax(MinMaxConstraint),
    Infinite,
}

The size of an axis. May be a specified size, a min/max constraint, or an unlimited size

Variants

Definite

MinMax

Infinite

Methods

impl AxisSize

fn new(size: LengthOrPercentageOrAuto, content_size: Option<Au>, min: LengthOrPercentage, max: LengthOrPercentageOrNone) -> AxisSize

Generate a new available cross or main axis size from the specified size of the container, containing block size, min constraint, and max constraint