Function layout::flex::from_flex_basis

fn from_flex_basis(
    flex_basis: &FlexBasis, 
    main_length: &Size, 
    containing_length: Au
) -> MaybeAuto

This function accepts the flex-basis and the size property in main direction from style, and the container size, then return the used value of flex basis. it can be used to help determining the flex base size and to indicate whether the main size of the item is definite after flex size resolving.