Struct layout::flex::FlexLine

struct FlexLine {
    pub range: Range<usize>,
    pub free_space: Au,
    pub auto_margin_count: i32,
    pub cross_size: Au,
}

A line in a flex container.

Fields
range: Range<usize>

Range of items belong to this line in 'self.items'.

free_space: Au

Remaining free space of this line, items will grow or shrink based on it being positive or negative.

auto_margin_count: i32

The number of auto margins of items.

cross_size: Au

Line size in the block direction.

Methods

impl FlexLine

fn new(range: Range<usize>, free_space: Au, auto_margin_count: i32) -> FlexLine

fn flex_resolve(&mut self, items: &mut [FlexItem], collapse: bool)

This method implements the flexible lengths resolving algorithm. The 'collapse' parameter is used to indicate whether items with 'visibility: collapse' is included in length resolving. The result main size is stored in 'item.main_size'. https://drafts.csswg.org/css-flexbox/#resolve-flexible-lengths