layout::flexbox

Module layout

Structs

DesiredFlexFractionAndGrowOrShrinkFactor 🔒

FinalFlexLineLayout 🔒

Once the final cross size of a line is known, the line can go through their final layout and this the return value. See InitialFlexLineLayout::finish_with_final_cross_size.

FlexContext 🔒

Layout parameters and intermediate results about a flex container, grouped to avoid passing around many parameters

FlexItem 🔒

A flex item with some intermediate results

FlexItemBoxInlineContentSizesInfo 🔒

FlexItemLayoutResult 🔒

Return type of FlexItem::layout

FlexLineItem 🔒

A data structure to hold all of the information about a flex item that has been placed into a flex line. This happens once the item is laid out and its line has been determined.

InitialFlexLineLayout 🔒

The result of splitting the flex items into lines using their intrinsic sizes and doing an initial layout of each item. A final layout still needs to happen after this is produced to handle stretching.

Enums

FlexContent 🔒

Child of a FlexContainer. Can either be absolutely positioned, or not. If not, a placeholder is used and flex content is stored outside of this enum.

Functions

allocate_free_cross_space_for_flex_line 🔒

Align all flex lines per align-content according to https://drafts.csswg.org/css-flexbox/#algo-line-align. Returns the space to add to each line or the space to add after each line.

cross_axis_is_item_block_axis 🔒

do_initial_flex_line_layout 🔒

“Collect flex items into flex lines” https://drafts.csswg.org/css-flexbox/#algo-line-break

item_with_auto_cross_size_stretches_to_line_size 🔒

Whether an item with a computed preferred cross size of auto will stretch to fill the cross size of its flex line. https://drafts.csswg.org/css-flexbox/#stretched