layout::flexbox::layout

Function item_with_auto_cross_size_stretches_to_line_size

fn item_with_auto_cross_size_stretches_to_line_size(
    align_self: T,
    margin: &FlexRelativeSides<AuOrAuto>,
) -> bool

Whether an item with a computed preferred cross size of auto will stretch to fill the cross size of its flex line. https://drafts.csswg.org/css-flexbox/#stretched