Struct layout::floats::Floats

pub struct Floats {
    pub writing_mode: WritingMode,
    // some fields omitted
}

Encapsulates information about floats. This is optimized to avoid allocation if there are no floats, and to avoid copying when translating the list of floats downward.

Fields

writing_mode

The writing mode of these floats.

Methods

impl Floats

fn new(writing_mode: WritingMode) -> Floats

Creates a new Floats object.

fn translate(&mut self, delta: LogicalSize<Au>)

Adjusts the recorded offset of the flow relative to the first float.

fn last_float_pos(&self) -> Option<LogicalPoint<Au>>

Returns the position of the last float in flow coordinates.

fn len(&self) -> usize

fn available_rect(&self, block_start: Au, block_size: Au, max_x: Au) -> Option<LogicalRect<Au>>

Returns a rectangle that encloses the region from block-start to block-start + block-size, with inline-size small enough that it doesn't collide with any floats. max_x is the inline-size beyond which floats have no effect. (Generally this is the containing block inline-size.)

fn add_float(&mut self, info: &PlacementInfo)

Adds a new float to the list.

fn place_between_floats(&self, info: &PlacementInfo) -> LogicalRect<Au>

Given placement information, finds the closest place a fragment can be positioned without colliding with any floats.

fn clearance(&self, clear: ClearType) -> Au