Struct layout::floats::PlacementInfo

pub struct PlacementInfo {
    pub size: LogicalSize<Au>,
    pub ceiling: Au,
    pub max_inline_size: Au,
    pub kind: FloatKind,
}

All the information necessary to place a float.

Fields

size

The dimensions of the float.

ceiling

The minimum block-start of the float, as determined by earlier elements.

max_inline_size

The maximum inline-end position of the float, generally determined by the containing block.

kind

The kind of float.