Struct layout::floats::SpeculatedFloatPlacement

pub struct SpeculatedFloatPlacement {
    pub left: Au,
    pub right: Au,
}

The speculated inline sizes of floats flowing through or around a flow (depending on whether the flow is a block formatting context). These speculations are always upper bounds; the actual inline sizes might be less. Note that this implies that a speculated value of zero is a guarantee that there will be no floats on that side.

This is used for two purposes: (a) determining whether we can lay out blocks in parallel; (b) guessing the inline-sizes of block formatting contexts in an effort to lay them out in parallel.

Fields
left: Au

The estimated inline size (an upper bound) of the left floats flowing through this flow.

right: Au

The estimated inline size (an upper bound) of the right floats flowing through this flow.

Methods

impl SpeculatedFloatPlacement

fn zero() -> SpeculatedFloatPlacement

Returns a SpeculatedFloatPlacement objects with both left and right speculated inline sizes initialized to zero.

fn compute_floats_in(&mut self, flow: &mut Flow)

Given the speculated inline size of the floats out for the inorder predecessor of this flow, computes the speculated inline size of the floats flowing in.

fn compute_floats_out(&mut self, flow: &mut Flow)

Given the speculated inline size of the floats out for this flow's last child, computes the speculated inline size of the floats out for this flow.

fn compute_floats_in_for_first_child(parent_flow: &mut Flow) -> SpeculatedFloatPlacement

Given a flow, computes the speculated inline size of the floats in of its first child.