layout::flow

Function solve_containing_block_padding_and_border_for_in_flow_box

fn solve_containing_block_padding_and_border_for_in_flow_box<'a>(
    containing_block: &ContainingBlock<'_>,
    layout_style: &'a LayoutStyle<'_>,
    get_inline_content_sizes: impl FnOnce(&ConstraintSpace<'_>) -> ContentSizes,
    ignore_block_margins_for_stretch: LogicalSides1D<bool>,
    context: Option<&IndependentFormattingContext>,
    has_inline_parent: bool,
) -> ContainingBlockPaddingAndBorder<'a>

Given the style for an in-flow box and its containing block, determine the containing block for its children. Note that in the presence of floats, this shouldn’t be used for a block-level box that establishes an independent formatting context (or is replaced), since the inline size could then be incorrect.