layout::flow::inline::construct

Struct InlineFormattingContextBuilder

pub(crate) struct InlineFormattingContextBuilder {

    pub shared_inline_styles_stack: Vec<SharedInlineStyles>,
    pub text_segments: Vec<String>,
    current_text_offset: usize,
    current_character_offset: usize,
    last_inline_box_ended_with_collapsible_white_space: bool,
    on_word_boundary: bool,
    pub contains_floats: bool,
    pub inline_items: Vec<InlineItem>,
    pub inline_boxes: InlineBoxes,
    inline_box_stack: Vec<InlineBoxIdentifier>,
    pub is_empty: bool,
    has_processed_first_letter: bool,
    pub has_right_to_left_content: bool,
    pub offset_map: ArcRefCell<OffsetMap>,

}

Fields
shared_inline_styles_stack: Vec<SharedInlineStyles>

A stack of SharedInlineStyles including one for the root, one for each inline box on the inline box stack, and importantly, one for every display: contents element that we are currently processing. Normally display: contents elements don’t affect the structure of the InlineFormattingContext, but the styles they provide do style their children.

text_segments: Vec<String>

The collection of text strings that make up this InlineFormattingContext under construction.

current_text_offset: usize

The current offset in the final text string of this InlineFormattingContext, used to properly set the text range of new InlineItem::TextRuns.

current_character_offset: usize

The current character offset in the final text string of this InlineFormattingContext, used to properly set the text range of new InlineItem::TextRuns. Note that this is different from the UTF-8 code point offset.

last_inline_box_ended_with_collapsible_white_space: bool

Whether the last processed node ended with whitespace. This is used to implement rule 4 of https://www.w3.org/TR/css-text-3/#collapse:

Any collapsible space immediately following another collapsible space—even one outside the boundary of the inline containing that space, provided both spaces are within the same inline formatting context—is collapsed to have zero advance width. (It is invisible, but retains its soft wrap opportunity, if any.)

on_word_boundary: bool

Whether or not the current state of the inline formatting context is on a word boundary for the purposes of text-transform: capitalize.

contains_floats: bool

Whether or not this inline formatting context will contain floats.

inline_items: Vec<InlineItem>

The current list of InlineItems in this InlineFormattingContext under construction. This is stored in a flat list to make it easy to access the last item.

inline_boxes: InlineBoxes

The current InlineBox tree of this InlineFormattingContext under construction.

inline_box_stack: Vec<InlineBoxIdentifier>

The ongoing stack of inline boxes stack of the builder.

Contains all the currently ongoing inline boxes we entered so far. The traversal is at all times as deep in the tree as this stack is, which is why the code doesn’t need to keep track of the actual container root (see handle_inline_level_element).

When an inline box ends, it’s removed from this stack.

is_empty: bool

Whether this InlineFormattingContextBuilder is empty for the purposes of ignoring during box tree construction. An IFC is empty if it only contains TextRuns with completely collapsible whitespace. When that happens it can be ignored completely.

has_processed_first_letter: bool

Whether or not the ::first-letter pseudo-element of this inline formatting context has been processed yet.

has_right_to_left_content: bool

Whether or not the inline formatting context under construction has any kind of right-to-left content such as a character with an RTL character class or a dir attribute specifying right-to-left content.

offset_map: ArcRefCell<OffsetMap>

An OffsetMap used to map selections from their offset before inline formatting context text transformation to their offsets after transformation.

Implementations

impl InlineFormattingContextBuilder

pub(crate) fn is_document_white_space(character: char) -> bool

https://drafts.csswg.org/css-text/#white-space:

Except where specified otherwise, white space processing in CSS affects only the document white space characters: spaces (U+0020), tabs (U+0009), and segment breaks.

From https://github.com/w3c/csswg-drafts/issues/5147#issuecomment-637816669:

HTML clearly treats CR, LF, and CRLF as segment breaks.

Other browsers also consider the form feed character (0x0c) to be document white space, it seems.

Taken all together, this is equivalent to the WhatWG Infra Standard’s definition of ASCII white space.

pub(crate) fn new( info: &NodeAndStyleInfo<'_>, context: &LayoutContext<'_>, ) -> Self

pub(crate) fn currently_processing_inline_box(&self) -> bool

fn push_control_character_string(&mut self, string_to_push: &str)

fn shared_inline_styles(&self) -> SharedInlineStyles

pub(crate) fn push_atomic( &mut self, independent_formatting_context_creator: impl FnOnce() -> ArcRefCell<IndependentFormattingContext>, old_layout_box: Option<LayoutBox>, ) -> InlineItem

pub(crate) fn push_absolutely_positioned_box( &mut self, absolutely_positioned_box_creator: impl FnOnce() -> ArcRefCell<AbsolutelyPositionedBox>, old_layout_box: Option<LayoutBox>, ) -> InlineItem

pub(crate) fn push_float_box( &mut self, float_box_creator: impl FnOnce() -> ArcRefCell<FloatBox>, old_layout_box: Option<LayoutBox>, ) -> InlineItem

pub(crate) fn push_block_level_box( &mut self, block_level: ArcRefCell<BlockLevelBox>, )

pub(crate) fn start_inline_box( &mut self, inline_box_creator: impl FnOnce() -> ArcRefCell<InlineBox>, old_layout_box: Option<LayoutBox>, ) -> InlineItem

pub(crate) fn end_inline_box(&mut self)

End the ongoing inline box in this InlineFormattingContextBuilder, returning shared references to all of the box tree items that were created for it. More than a single box tree items may be produced for a single inline box when that inline box is split around a block-level element.

pub(crate) fn push_text_with_possible_first_letter<'dom>( &mut self, text: BoxTreeString<'dom>, info: &NodeAndStyleInfo<'dom>, container_info: &NodeAndStyleInfo<'dom>, layout_context: &LayoutContext<'_>, ) -> bool

This is like Self::push_text, except that it might possibly add an anonymous box if

This inline formatting context has a ::first-letter style.

No anonymous box for ::first-letter has been added yet.

First letter content is detected in this text.

Note that this should only be used when processing text in block containers.

pub(crate) fn push_text<'dom>( &mut self, text: BoxTreeString<'dom>, info: &NodeAndStyleInfo<'dom>, selection: Option<RangeAny<Utf32CodeUnits>>, )

pub(crate) fn enter_display_contents( &mut self, shared_inline_styles: SharedInlineStyles, )

pub(crate) fn leave_display_contents(&mut self)

pub(crate) fn finish( self, layout_context: &LayoutContext<'_>, has_first_formatted_line: bool, is_single_line_text_input: bool, default_bidi_level: Level, ) -> Option<InlineFormattingContext>

Finish the current inline formatting context, returning None if the context was empty.