layout::flow::inline

Module line

Structs

AbsolutelyPositionedLineItem 🔒

AtomicLineItem 🔒

FloatLineItem 🔒

FragmentAndData 🔒

LineItemLayout 🔒

The second phase of super::InlineFormattingContext layout: once items are gathered for a line, we must lay them out and create fragments for them, properly positioning them according to their baselines and also handling absolutely positioned children.

LineItemLayoutInlineContainerState 🔒

The state used when laying out a collection of LineItems into a line. This state is stored per-inline container. For instance, when laying out the conents of a <span> a fresh LineItemLayoutInlineContainerState is pushed onto LineItemLayout’s stack of states.

LineLayoutInlineContainerFlags 🔒

LineMetrics 🔒

TextRunLineItem 🔒

Enums

LineItem 🔒

Functions

sort_by_indices_in_place 🔒

Sort a mutable slice by the given indices array in place, reording the slice so that final value of slice[x] is slice[indices[x]].