layout::flow::inline::line

Struct LineItemLayout

pub(super) struct LineItemLayout<'layout_data, 'layout> {
    layout: &'layout mut InlineFormattingContextLayout<'layout_data>,
    pub state_stack: Vec<LineItemLayoutInlineContainerState>,
    pub current_state: LineItemLayoutInlineContainerState,
    pub line_metrics: LineMetrics,
    pub justification_adjustment: Au,
    is_phantom_line: bool,
    for_block_level: bool,
}

The second phase of super::InlineFormattingContext layout: once items are gathered for a line, we must lay them out and create fragments for them, properly positioning them according to their baselines and also handling absolutely positioned children.

Fields
layout: &'layout mut InlineFormattingContextLayout<'layout_data>

The state of the overall super::InlineFormattingContext layout.

state_stack: Vec<LineItemLayoutInlineContainerState>

The set of LineItemLayoutInlineContainerState created while laying out items on this line. This does not include the current level of recursion.

current_state: LineItemLayoutInlineContainerState

The current LineItemLayoutInlineContainerState.

line_metrics: LineMetrics

The metrics of this line, which should remain constant throughout the layout process.

justification_adjustment: Au

The amount of space to add to each justification opportunity in order to implement text-align: justify.

is_phantom_line: bool

Whether this is a phantom line box. https://drafts.csswg.org/css-inline-3/#invisible-line-boxes

for_block_level: bool

Whether this line contains only a block-level box.

Implementations

impl LineItemLayout<'_, '_>

pub(super) fn layout_line_items( layout: &mut InlineFormattingContextLayout<'_>, line_items: Vec<LineItem>, start_position: LogicalVec2<Au>, effective_block_advance: &LineBlockSizes, justification_adjustment: Au, is_phantom_line: bool, for_block_level: bool, ) -> Vec<Fragment>

fn prepare_layout_for_inline_box( &mut self, new_inline_box: Option<InlineBoxIdentifier>, )

Start and end inline boxes in tree order, so that it reflects the given inline box.

fn reorder_line_items_for_bidi( &self, line_items: Vec<LineItem>, ) -> impl Iterator<Item = LineItem> + use<>

If the inline formatting context that this line is being laid out for had right-to-left content, reorder the line contents according to their pre-calculated BiDi levels.

Returns an iterator over the line contents.

pub(super) fn layout(&mut self, line_items: Vec<LineItem>) -> Vec<Fragment>

fn current_positioning_context_mut(&mut self) -> &mut PositioningContext

fn start_inline_box(&mut self, identifier: &InlineBoxIdentifier)

fn end_inline_box(&mut self)

fn calculate_inline_box_block_start( &self, inline_box_state: &InlineBoxContainerState, space_above_baseline: Au, ) -> Au

fn layout_text_run(&mut self, text_item: TextRunLineItem)

fn layout_atomic(&mut self, atomic: AtomicLineItem)

fn layout_absolute(&mut self, absolute: AbsolutelyPositionedLineItem)

fn layout_float(&mut self, float: FloatLineItem)

fn layout_block_level(&mut self, block_level: Arc<BoxFragment>)

fn containing_block(&self) -> &ContainingBlock<'_>

fn layout_tab(&mut self, advance: Au)