layout::flow::inline

Struct ContentSizesComputation

struct ContentSizesComputation<'layout_data> {
    layout_context: &'layout_data LayoutContext<'layout_data>,
    constraint_space: &'layout_data ConstraintSpace<'layout_data>,
    paragraph: ContentSizes,
    current_line: ContentSizes,
    pending_whitespace: ContentSizes,
    uncleared_floats: LogicalSides1D<ContentSizes>,
    cleared_floats: LogicalSides1D<ContentSizes>,
    had_content_yet_for_min_content: bool,
    had_content_yet_for_max_content: bool,
    ending_inline_pbm_stack: Vec<Au>,
    depends_on_block_constraints: bool,
}

A struct which takes care of computing ContentSizes for an InlineFormattingContext.

Fields
layout_context: &'layout_data LayoutContext<'layout_data>constraint_space: &'layout_data ConstraintSpace<'layout_data>paragraph: ContentSizescurrent_line: ContentSizespending_whitespace: ContentSizes

Size for whitespace pending to be added to this line.

uncleared_floats: LogicalSides1D<ContentSizes>

The size of the not yet cleared floats in the inline axis of the containing block.

cleared_floats: LogicalSides1D<ContentSizes>

The size of the already cleared floats in the inline axis of the containing block.

had_content_yet_for_min_content: bool

Whether or not the current line has seen any content (excluding collapsed whitespace), when sizing under a min-content constraint.

had_content_yet_for_max_content: bool

Whether or not the current line has seen any content (excluding collapsed whitespace), when sizing under a max-content constraint.

ending_inline_pbm_stack: Vec<Au>

Stack of ending padding, margin, and border to add to the length when an inline box finishes.

depends_on_block_constraints: bool

Whether the inline content size depends on block constraints.

Implementations

impl<'layout_data> ContentSizesComputation<'layout_data>

fn traverse( self, inline_formatting_context: &InlineFormattingContext, ) -> InlineContentSizesResult

fn process_item( &mut self, inline_item: &InlineItem, inline_formatting_context: &InlineFormattingContext, )

fn process_text_segment( &mut self, parent_style: &AtomicRef<'_, ServoArc<ComputedValues>>, segment: &TextRunSegment, )

fn process_preserved_tab( &mut self, parent_style: &AtomicRef<'_, ServoArc<ComputedValues>>, inline_formatting_context: &InlineFormattingContext, )

fn add_inline_size(&mut self, l: Au)

fn line_break_opportunity(&mut self)

fn forced_line_break(&mut self)

fn commit_pending_whitespace(&mut self)

fn outer_inline_content_sizes_of_float_or_atomic( &mut self, context: &IndependentFormattingContext, ) -> ContentSizes

fn clear_floats(&mut self, clear: Clear)

fn flush_floats(&mut self)

fn compute( inline_formatting_context: &InlineFormattingContext, layout_context: &'layout_data LayoutContext<'_>, constraint_space: &'layout_data ConstraintSpace<'_>, ) -> InlineContentSizesResult

Compute the ContentSizes of the given InlineFormattingContext.