layout::flow::inline

Struct InlineFormattingContext

pub(crate) struct InlineFormattingContext {
    inline_items: Vec<InlineItem>,
    inline_boxes: InlineBoxes,
    text_content: String,
    shared_inline_styles: SharedInlineStyles,
    default_font: Option<FontRef>,
    has_first_formatted_line: bool,
    pub(super) contains_floats: bool,
    is_single_line_text_input: bool,
    has_right_to_left_content: bool,
    tab_size_multiplier: OnceLock<Au>,
}

Fields
inline_items: Vec<InlineItem>

All InlineItems in this InlineFormattingContext stored in a flat array. InlineItem::StartInlineBox and InlineItem::EndInlineBox allow representing the tree of inline boxes within the formatting context, but a flat array allows easy iteration through all inline items.

inline_boxes: InlineBoxes

The tree of inline boxes in this InlineFormattingContext. These are stored in a flat array with each being given a InlineBoxIdentifier.

text_content: String

The text content of this inline formatting context.

shared_inline_styles: SharedInlineStyles

The SharedInlineStyles for the root of this InlineFormattingContext that are used to share styles with all TextRun children.

default_font: Option<FontRef>

The default font that is used for the root of this InlineFormattingContext. This is the font used when the font fallback code path is not taken. It may be None if no default font was found (this typically means that no characters can be rendered).

has_first_formatted_line: bool

Whether this IFC contains the 1st formatted line of an element: https://www.w3.org/TR/css-pseudo-4/#first-formatted-line.

contains_floats: bool

Whether or not this InlineFormattingContext contains floats.

is_single_line_text_input: bool

Whether or not this is an InlineFormattingContext for a single line text input’s inner text container.

has_right_to_left_content: bool

Whether or not this is an InlineFormattingContext has right-to-left content, which will require reordering during layout.

tab_size_multiplier: OnceLock<Au>

The cached multiplier for tab-size: <number>: https://drafts.csswg.org/css-text/#tab-size-property

the advance width of the space character (U+0020) of the nearest block container ancestor of the preserved tab, including its associated letter-spacing and word-spacing.

Implementations

impl InlineFormattingContext

fn new_with_builder( builder: InlineFormattingContextBuilder, layout_context: &LayoutContext<'_>, has_first_formatted_line: bool, is_single_line_text_input: bool, starting_bidi_level: Level, ) -> Self

pub(crate) fn repair_style( &self, context: &SharedStyleContext<'_>, node: &ServoLayoutNode<'_>, new_style: &ServoArc<ComputedValues>, )

fn inline_start_for_first_line( &self, containing_block: IndefiniteContainingBlock<'_>, ) -> Au

pub(super) fn layout( &self, layout_context: &LayoutContext<'_>, positioning_context: &mut PositioningContext, containing_block: &ContainingBlock<'_>, sequential_layout_state: Option<&mut SequentialLayoutState>, collapsible_with_parent_start_margin: CollapsibleWithParentStartMargin, ignore_block_margins_for_stretch: LogicalSides1D<bool>, ) -> IndependentFormattingContextLayoutResult

pub(crate) fn subtree_size(&self) -> usize

fn next_character_prevents_soft_wrap_opportunity(&self, index: usize) -> bool

fn previous_character_prevents_soft_wrap_opportunity( &self, index: usize, ) -> bool

pub(crate) fn find_block_margin_collapsing_with_parent( &self, layout_context: &LayoutContext<'_>, collected_margin: &mut CollapsedMargin, containing_block_for_children: &ContainingBlock<'_>, ) -> bool

pub(crate) fn attached_to_tree(&self, layout_box: WeakLayoutBox)

pub(crate) fn next_tab_stop_after_inline_advance( &self, style: &ServoArc<ComputedValues>, current_inline_advance: Au, ) -> Au