layout::flow::inline

Module text_run

Structs

CaretPlaceholder 🔒

FontAndScriptInfo 🔒

A data structure which contains information used when shaping a TextRunSegment.

FontInfo 🔒

A data structure which contains information used when shaping a TextRunSegment.

SharedTextRunData 🔒

A data structure that holds per-TextRun data used on TextFragments. This ensures that the data is not duplicated between fragments.

TextRun 🔒

A single TextRun for the box tree. These are all descendants of super::InlineBox or the root of the super::InlineFormattingContext. During box tree construction, text is split into TextRuns based on their font, script, etc. When these are created text is already shaped.

TextRunSegment 🔒

TwoCharsAtATimeIterator 🔒

Enums

SegmentStartSoftWrapPolicy 🔒

TextRunItem 🔒

A single item in a TextRun.

Functions

character_cannot_change_font 🔒

Whether or not this character should be able to change the font during segmentation. Certain character are not rendered at all, so it doesn’t matter what font we use to render them. They should just be added to the current segment.

get_font_for_first_font_for_style 🔒

is_cursive_script 🔒

From https://www.w3.org/TR/css-text-3/#cursive-script: Cursive scripts do not admit gaps between their letters for either justification or letter-spacing. The following Unicode scripts are included: Arabic, Hanifi Rohingya, Mandaic, Mongolian, N’Ko, Phags Pa, Syriac

script_is_specific 🔒