Struct layout::flow::AbsoluteDescendantInfo

pub struct AbsoluteDescendantInfo {
    flow: FlowRef,
    has_reached_containing_block: bool,
}

Information about each absolutely-positioned descendant of the given flow.

Fields
flow: FlowRef

The absolute descendant flow in question.

has_reached_containing_block: bool

Whether the absolute descendant has reached its containing block. This exists so that we can handle cases like the following:

<div>
    <span id=a style="position: absolute; ...">foo</span>
    <span style="position: relative">
        <span id=b style="position: absolute; ...">bar</span>
    </span>
</div>

When we go to create the InlineFlow for the outer div, our absolute descendants will be a and b. At this point, we need a way to distinguish between the two, because the containing block for a will be different from the containing block for b. Specifically, the latter's containing block is the inline flow itself, while the former's containing block is going to be some parent of the outer div. Hence we need this flag as a way to distinguish the two; it will be false for a and true for b.