Struct layout::flow::AbsolutePositionInfo

pub struct AbsolutePositionInfo {
    pub relative_containing_block_size: LogicalSize<Au>,
    pub relative_containing_block_mode: WritingMode,
    pub stacking_relative_position_of_absolute_containing_block: Point2D<Au>,
    pub layers_needed_for_positioned_flows: bool,
}

Information needed to compute absolute (i.e. viewport-relative) flow positions (not to be confused with absolutely-positioned flows).

Fields

relative_containing_block_size

The size of the containing block for relatively-positioned descendants.

relative_containing_block_mode

The writing mode for relative_containing_block_size.

stacking_relative_position_of_absolute_containing_block

The position of the absolute containing block relative to the nearest ancestor stacking context. If the absolute containing block establishes the stacking context for this flow, and this flow is not itself absolutely-positioned, then this is (0, 0).

layers_needed_for_positioned_flows

Whether the absolute containing block forces positioned descendants to be layerized.

FIXME(pcwalton): Move into FlowFlags.

Methods

impl AbsolutePositionInfo

fn new(writing_mode: WritingMode) -> AbsolutePositionInfo