Struct layout::flow::ContainingBlockLink

pub struct ContainingBlockLink {
    // some fields omitted
}

A link to a flow's containing block.

This cannot safely be a Flow pointer because this is a pointer up the tree, not down the tree. A pointer up the tree is unsafe during layout because it can be used to access a node with an immutable reference while that same node is being laid out, causing possible iterator invalidation and use-after-free.

FIXME(pcwalton): I think this would be better with a borrow flag instead of unsafe.

Methods

impl ContainingBlockLink

unsafe fn get<'a>(&'a mut self) -> &'a mut Option<WeakFlowRef>

fn generated_containing_block_rect(&mut self) -> LogicalRect<Au>