Struct layout::flow::Descendants

pub struct Descendants {
    pub static_block_offsets: Vec<Au>,
    // some fields omitted
}

The Descendants of a flow.

Also, details about their position wrt this flow.

Fields

static_block_offsets

Static block-direction offsets of all descendants from the start of this flow box.

Methods

impl Descendants

fn new() -> Descendants

fn len(&self) -> usize

fn is_empty(&self) -> bool

fn push(&mut self, given_descendant: FlowRef)

fn push_descendants(&mut self, given_descendants: Descendants)

Push the given descendants on to the existing descendants.

Ignore any static y offsets, because they are None before layout.

fn iter<'a>(&'a mut self) -> DescendantIter<'a>

Return an iterator over the descendant flows.

fn iter_with_offset<'a>(&'a mut self) -> DescendantOffsetIter<'a>

Return an iterator over (descendant, static y offset).