Struct layout::flow::LateAbsolutePositionInfo

pub struct LateAbsolutePositionInfo {
    pub stacking_relative_position_of_absolute_containing_block: Point2D<Au>,
}

Information needed to compute absolute (i.e. viewport-relative) flow positions (not to be confused with absolutely-positioned flows) that is computed during final position assignment.

Fields

stacking_relative_position_of_absolute_containing_block

The position of the absolute containing block relative to the nearest ancestor stacking context. If the absolute containing block establishes the stacking context for this flow, and this flow is not itself absolutely-positioned, then this is (0, 0).

Methods

impl LateAbsolutePositionInfo

fn new() -> LateAbsolutePositionInfo