Trait layout::flow::MutableFlowUtils

pub trait MutableFlowUtils {
    fn traverse_preorder<T: PreorderFlowTraversal>(self, traversal: &T);
    fn traverse_postorder<T: PostorderFlowTraversal>(self, traversal: &T);
    fn store_overflow(self, _: &LayoutContext);
    fn collect_static_block_offsets_from_children(self);
}

Required Methods

fn traverse_preorder<T: PreorderFlowTraversal>(self, traversal: &T)

Traverses the tree in preorder.

fn traverse_postorder<T: PostorderFlowTraversal>(self, traversal: &T)

Traverses the tree in postorder.

fn store_overflow(self, _: &LayoutContext)

Computes the overflow region for this flow.

fn collect_static_block_offsets_from_children(self)

Gathers static block-offsets bubbled up by kids.

This essentially gives us offsets of all absolutely positioned direct descendants and all fixed descendants, in tree order.

This is called in a bottom-up traversal (specifically, the assign-block-size traversal). So, kids have their flow origin already set. In the case of absolute flow kids, they have their hypothetical box position already set.