Trait layout::flow::MutableOwnedFlowUtils

pub trait MutableOwnedFlowUtils {
    fn set_absolute_descendants(&mut self, abs_descendants: AbsoluteDescendants);

    fn take_applicable_absolute_descendants(
        &mut self, 
        absolute_descendants: &mut AbsoluteDescendants
    );
}

Required methods

fn set_absolute_descendants(&mut self, abs_descendants: AbsoluteDescendants)

Set absolute descendants for this flow.

Set this flow as the Containing Block for all the absolute descendants.

fn take_applicable_absolute_descendants(
&mut self,
absolute_descendants: &mut AbsoluteDescendants
)

Sets the flow as the containing block for all absolute descendants that have been marked as having reached their containing block. This is needed in order to handle cases like:

<div>
    <span style="position: relative">
        <span style="position: absolute; ..."></span>
    </span>
</div>

Loading content...