Trait layout::flow::PreorderFlowTraversal

pub trait PreorderFlowTraversal {
    fn process(&self, flow: &mut Flow);

    fn should_process(&self, _flow: &mut Flow) -> bool { ... }
}

A top-down traversal.

Required Methods

fn process(&self, flow: &mut Flow)

The operation to perform. Return true to continue or false to stop.

Provided Methods

fn should_process(&self, _flow: &mut Flow) -> bool

Returns true if this node must be processed in-order. If this returns false, we skip the operation for this node, but continue processing the descendants. This is called after parent nodes are visited.