Struct layout::flow_list::FlowList

pub struct FlowList {
    flows: LinkedList<FlowRef>,
}

This needs to be reworked now that we have dynamically-sized types in Rust. Until then, it's just a wrapper around LinkedList.

SECURITY-NOTE(pcwalton): It is very important that FlowRef values not leak directly to layout. Layout code must only interact with &Flow or &mut Flow values. Otherwise, layout could stash FlowRef values in random places unknown to the system and thereby cause data races. Those data races can lead to memory safety problems, potentially including arbitrary remote code execution! In general, do not add new methods to this file (e.g. new ways of iterating over flows) unless you are very sure of what you are doing.

Fields
flows: LinkedList<FlowRef>
Methods

impl FlowList

fn push_back(&mut self, new_tail: FlowRef)

Add an element last in the list

O(1)

fn push_back_arc(&mut self, new_head: Arc<Flow>)

fn back(&self) -> Option<&Flow>

fn push_front(&mut self, new_head: FlowRef)

Add an element first in the list

O(1)

fn push_front_arc(&mut self, new_head: Arc<Flow>)

fn pop_front_arc(&mut self) -> Option<Arc<Flow>>

fn front(&self) -> Option<&Flow>

fn new() -> FlowList

Create an empty list

fn iter<'a>(&'a self) -> impl DoubleEndedIterator<Item=&'a Flow>

Provide a forward iterator.

SECURITY-NOTE(pcwalton): This does not hand out FlowRefs by design. Do not add a method to do so! See the comment above in FlowList.

fn iter_mut(&mut self) -> MutFlowListIterator

Provide a forward iterator with mutable references

SECURITY-NOTE(pcwalton): This does not hand out FlowRefs by design. Do not add a method to do so! See the comment above in FlowList.

fn random_access_mut(&mut self) -> FlowListRandomAccessMut

Provides a caching random-access iterator that yields mutable references. This is guaranteed to perform no more than O(n) pointer chases.

SECURITY-NOTE(pcwalton): This does not hand out FlowRefs by design. Do not add a method to do so! See the comment above in FlowList.

fn is_empty(&self) -> bool

O(1)

fn len(&self) -> usize

O(1)

fn split_off(&mut self, i: usize) -> Self