Module layout::flow_ref

Reference-counted pointers to flows.

Eventually, with dynamically sized types in Rust, much of this code will be superfluous. This design is largely duplicating logic of Arc and Weak; please see comments there for details.

Functions

deref_mut
WARNING: This should only be used when there is no aliasing: when the traversal ensures that no other threads accesses the same flow at the same time. See https://github.com/servo/servo/issues/6503 Use Arc::get_mut instead when possible (e.g. on an Arc that was just created).

Type Definitions

FlowRef

WeakFlowRef