Struct layout::flow_ref::FlowRef

pub struct FlowRef(Arc<Flow>);

Methods

impl FlowRef

fn new(r: Arc<Flow>) -> Self

FlowRefs can only be made available to the traversal code. See https://github.com/servo/servo/issues/14014 for more details.

fn get_mut(this: &mut FlowRef) -> Option<&mut Flow>

fn downgrade(this: &FlowRef) -> WeakFlowRef

fn into_arc(this: FlowRef) -> Arc<Flow>

fn deref_mut(this: &mut FlowRef) -> &mut Flow

WARNING: This should only be used when there is no aliasing: when the traversal ensures that no other threads accesses the same flow at the same time. See https://github.com/servo/servo/issues/6503 Use Arc::get_mut instead when possible (e.g. on an Arc that was just created).