layout::formatting_contexts

Struct IndependentFormattingContext

pub(crate) struct IndependentFormattingContext {
    pub base: LayoutBoxBase,
    contents: IndependentFormattingContextContents,
    pub propagated_data: PropagatedBoxTreeData,
    pub layout_root_layout_inputs: AtomicRefCell<Option<Box<LayoutRootLayoutInputs>>>,
}

https://drafts.csswg.org/css-display/#independent-formatting-context

Fields
base: LayoutBoxBasecontents: IndependentFormattingContextContentspropagated_data: PropagatedBoxTreeData

Data that was originally propagated down to this IndependentFormattingContext during creation. This is used during incremental layout.

layout_root_layout_inputs: AtomicRefCell<Option<Box<LayoutRootLayoutInputs>>>

If this IndependentFormattingContext was a layout root, this stores the data necessary to lay it out again.

Implementations

impl IndependentFormattingContext

fn layout_into_line_items( &self, layout: &mut InlineFormattingContextLayout<'_>, offset_in_text: usize, bidi_level: Level, )

fn pick_baseline(&self, baselines: &Baselines) -> Option<Au>

Picks either the first or the last baseline, depending on baseline-source. TODO: clarify that this is not to be used for box alignment in flex/grid https://drafts.csswg.org/css-inline/#baseline-source

fn get_block_sizes_and_baseline_offset( &self, ifc: &InlineFormattingContextLayout<'_>, block_size: Au, baseline_offset_in_content_area: Au, ) -> (LineBlockSizes, Au)

impl IndependentFormattingContext

pub(crate) fn layout_in_flow_block_level( &self, layout_context: &LayoutContext<'_>, positioning_context: &mut PositioningContext, containing_block: &ContainingBlock<'_>, sequential_layout_state: Option<&mut SequentialLayoutState>, ignore_block_margins_for_stretch: LogicalSides1D<bool>, has_inline_parent: bool, ) -> BoxFragment

Lay out an in-flow block-level box that establishes an independent formatting context in its containing formatting context.

https://drafts.csswg.org/css2/visudet.html#blockwidth

https://drafts.csswg.org/css2/visudet.html#block-replaced-width

https://drafts.csswg.org/css2/visudet.html#normal-block

https://drafts.csswg.org/css2/visudet.html#inline-replaced-height

fn layout_in_flow_block_level_sequentially( &self, layout_context: &LayoutContext<'_>, positioning_context: &mut PositioningContext, containing_block: &ContainingBlock<'_>, sequential_layout_state: &mut SequentialLayoutState, ignore_block_margins_for_stretch: LogicalSides1D<bool>, has_inline_parent: bool, ) -> BoxFragment

Lay out a normal in flow non-replaced block that establishes an independent formatting context in its containing formatting context but handling sequential layout concerns, such clearing and placing the content next to floats.

impl IndependentFormattingContext

pub(crate) fn layout_float_or_atomic_inline( &self, layout_context: &LayoutContext<'_>, child_positioning_context: &mut PositioningContext, containing_block: &ContainingBlock<'_>, ) -> IndependentFloatOrAtomicLayoutResult

impl IndependentFormattingContext

pub(crate) fn new( base: LayoutBoxBase, contents: IndependentFormattingContextContents, propagated_data: PropagatedBoxTreeData, ) -> Self

pub(crate) fn rebuild( &mut self, layout_context: &LayoutContext<'_>, node_and_style_info: &NodeAndStyleInfo<'_>, )

pub(crate) fn construct( context: &LayoutContext<'_>, node_and_style_info: &NodeAndStyleInfo<'_>, display_inside: DisplayInside, contents: Contents, propagated_data: PropagatedBoxTreeData, ) -> Self

fn construct_contents( context: &LayoutContext<'_>, node_and_style_info: &NodeAndStyleInfo<'_>, base_fragment_info: &mut BaseFragmentInfo, display_inside: DisplayInside, contents: Contents, propagated_data: PropagatedBoxTreeData, ) -> IndependentFormattingContextContents

pub fn style(&self) -> &Arc<ComputedValues>

pub fn base_fragment_info(&self) -> BaseFragmentInfo

pub(crate) fn inline_content_sizes( &self, layout_context: &LayoutContext<'_>, constraint_space: &ConstraintSpace<'_>, ) -> InlineContentSizesResult

pub(crate) fn tentative_block_content_size( &self, preferred_aspect_ratio: Option<AspectRatio>, inline_stretch_size: Au, ) -> Option<ContentSizes>

Computes the tentative intrinsic block sizes that may be needed while computing the intrinsic inline sizes. Therefore, this ignores the values of the sizing properties in both axes. A return value of None indicates that there is no suitable tentative intrinsic block size, so intrinsic keywords in the block sizing properties will be ignored, possibly resulting in an indefinite [SizeConstraint] for computing the intrinsic inline sizes and laying out the contents. A return value of Some indicates that intrinsic keywords in the block sizing properties will be resolved as the contained value, guaranteeing a definite amount for computing the intrinsic inline sizes and laying out the contents.

pub(crate) fn tentative_block_content_size_with_dependency( &self, preferred_aspect_ratio: Option<AspectRatio>, inline_stretch_size: Au, ) -> Option<(ContentSizes, bool)>

Same as Self::tentative_block_content_size(), but if there is a tentative intrinsic block size, it also includes a bool which will be true if the former depends on the provided inline_stretch_size.

pub(crate) fn outer_inline_content_sizes( &self, layout_context: &LayoutContext<'_>, containing_block: &IndefiniteContainingBlock<'_>, auto_minimum: &LogicalVec2<Au>, auto_block_size_stretches_to_containing_block: bool, ) -> InlineContentSizesResult

pub(crate) fn repair_style( &mut self, context: &SharedStyleContext<'_>, node: &ServoLayoutNode<'_>, new_style: &Arc<ComputedValues>, )

pub(crate) fn is_block_container(&self) -> bool

pub(crate) fn is_replaced(&self) -> bool

pub(crate) fn is_table(&self) -> bool

pub(crate) fn is_grid(&self) -> bool

fn layout_without_caching( &self, layout_context: &LayoutContext<'_>, positioning_context: &mut PositioningContext, containing_block_for_children: &ContainingBlock<'_>, containing_block: &ContainingBlock<'_>, preferred_aspect_ratio: Option<AspectRatio>, lazy_block_size: &LazySize<'_>, ) -> IndependentFormattingContextLayoutResult

pub(crate) fn layout_and_is_cached( &self, layout_context: &LayoutContext<'_>, positioning_context: &mut PositioningContext, containing_block_for_children: &ContainingBlock<'_>, containing_block: &ContainingBlock<'_>, preferred_aspect_ratio: Option<AspectRatio>, lazy_block_size: &LazySize<'_>, ) -> (IndependentFormattingContextLayoutResult, bool)

pub(crate) fn layout( &self, layout_context: &LayoutContext<'_>, positioning_context: &mut PositioningContext, containing_block_for_children: &ContainingBlock<'_>, containing_block: &ContainingBlock<'_>, preferred_aspect_ratio: Option<AspectRatio>, lazy_block_size: &LazySize<'_>, ) -> IndependentFormattingContextLayoutResult

pub(crate) fn layout_style(&self) -> LayoutStyle<'_>

pub(crate) fn preferred_aspect_ratio( &self, padding_border_sums: &LogicalVec2<Au>, ) -> Option<AspectRatio>

pub(crate) fn attached_to_tree(&self, layout_box: WeakLayoutBox)

pub(crate) fn subtree_size(&self) -> usize

impl IndependentFormattingContext

pub(crate) fn layout_as_absolute( &self, layout_context: &LayoutContext<'_>, static_position_rect: &LogicalRect<Au>, containing_block: &DefiniteContainingBlock<'_>, resolved_alignment: LogicalVec2<AlignFlags>, original_parent_writing_mode: WritingMode, ) -> (Arc<BoxFragment>, PositioningContext)