Enum layout::fragment::SpecificFragmentInfo

pub enum SpecificFragmentInfo {
    Generic,
    GeneratedContent(Box<GeneratedContentInfo>),
    Iframe(IframeFragmentInfo),
    Image(Box<ImageFragmentInfo>),
    Canvas(Box<CanvasFragmentInfo>),
    InlineAbsoluteHypothetical(InlineAbsoluteHypotheticalFragmentInfo),
    InlineBlock(InlineBlockFragmentInfo),
    InlineAbsolute(InlineAbsoluteFragmentInfo),
    ScannedText(Box<ScannedTextFragmentInfo>),
    Table,
    TableCell,
    TableColumn(TableColumnFragmentInfo),
    TableRow,
    TableWrapper,
    Multicol,
    MulticolColumn,
    UnscannedText(Box<UnscannedTextFragmentInfo>),
}

Info specific to the kind of fragment.

Keep this enum small. As in, no more than one word. Or pcwalton will yell at you.

Variants
GenericGeneratedContent(Box<GeneratedContentInfo>)

A piece of generated content that cannot be resolved into ScannedText until the generated content resolution phase (e.g. an ordered list item marker).

Iframe(IframeFragmentInfo)Image(Box<ImageFragmentInfo>)Canvas(Box<CanvasFragmentInfo>)InlineAbsoluteHypothetical(InlineAbsoluteHypotheticalFragmentInfo)

A hypothetical box (see CSS 2.1  10.3.7) for an absolutely-positioned block that was declared with display: inline;.

InlineBlock(InlineBlockFragmentInfo)InlineAbsolute(InlineAbsoluteFragmentInfo)

An inline fragment that establishes an absolute containing block for its descendants (i.e. a positioned inline fragment).

ScannedText(Box<ScannedTextFragmentInfo>)TableTableCellTableColumn(TableColumnFragmentInfo)TableRowTableWrapperMulticolMulticolColumnUnscannedText(Box<UnscannedTextFragmentInfo>)
Methods

impl SpecificFragmentInfo

fn restyle_damage(&self) -> RestyleDamage

fn get_type(&self) -> &'static str