Struct layout::fragment::Fragment

pub struct Fragment {
    pub node: OpaqueNode,
    pub style: Arc<ComputedValues>,
    pub border_box: LogicalRect<Au>,
    pub border_padding: LogicalMargin<Au>,
    pub margin: LogicalMargin<Au>,
    pub specific: SpecificFragmentInfo,
    pub inline_context: Option<InlineFragmentContext>,
    pub restyle_damage: RestyleDamage,
    pub debug_id: u16,
}

Fragments (struct Fragment) are the leaves of the layout tree. They cannot position themselves. In general, fragments do not have a simple correspondence with CSS fragments in the specification:

Several fragments may correspond to the same CSS box or DOM node. For example, a CSS text box broken across two lines is represented by two fragments.

Some CSS fragments are not created at all, such as some anonymous block fragments induced by inline fragments with block-level sibling fragments. In that case, Servo uses an InlineFlow with BlockFlow siblings; the InlineFlow is block-level, but not a block container. It is positioned as if it were a block fragment, but its children are positioned according to inline flow.

A SpecificFragmentInfo::Generic is an empty fragment that contributes only borders, margins, padding, and backgrounds. It is analogous to a CSS nonreplaced content box.

A fragment's type influences how its styles are interpreted during layout. For example, replaced content such as images are resized differently from tables, text, or other content. Different types of fragments may also contain custom data; for example, text fragments contain text.

Do not add fields to this structure unless they're really really mega necessary! Fragments get moved around a lot and thus their size impacts performance of layout quite a bit.

FIXME(#2260, pcwalton): This can be slimmed down some by (at least) moving inline_context to be on InlineFlow only.

Fields

node

An opaque reference to the DOM node that this Fragment originates from.

style

The CSS style of this fragment.

border_box

The position of this fragment relative to its owning flow. The size includes padding and border, but not margin.

NB: This does not account for relative positioning.

border_padding

The sum of border and padding; i.e. the distance from the edge of the border box to the content edge of the fragment.

margin

The margin of the content box.

specific

Info specific to the kind of fragment. Keep this enum small.

inline_context

Holds the style context information for fragments that are part of an inline formatting context.

restyle_damage

How damaged this fragment is since last reflow.

debug_id

A debug ID that is consistent for the life of this fragment (via transform etc).

Methods

impl Fragment

fn new(node: &ThreadSafeLayoutNode, specific: SpecificFragmentInfo) -> Fragment

Constructs a new Fragment instance.

fn new_anonymous_from_specific_info(node: &ThreadSafeLayoutNode, specific: SpecificFragmentInfo) -> Fragment

Constructs a new Fragment instance for an anonymous table object.

fn from_opaque_node_and_style(node: OpaqueNode, style: Arc<ComputedValues>, restyle_damage: RestyleDamage, specific: SpecificFragmentInfo) -> Fragment

Constructs a new Fragment instance from an opaque node.

fn reset_inline_sizes(&mut self)

fn save_new_line_pos(&mut self)

Saves the new_line_pos vector into a SpecificFragmentInfo::ScannedText. This will fail if called on any other type of fragment.

fn restore_new_line_pos(&mut self)

fn debug_id(&self) -> u16

Returns a debug ID of this fragment. This ID should not be considered stable across multiple layouts or fragment manipulations.

fn transform(&self, size: LogicalSize<Au>, info: SpecificFragmentInfo) -> Fragment

Transforms this fragment into another fragment of the given type, with the given size, preserving all the other data.

fn transform_with_split_info(&self, split: &SplitInfo, text_run: Arc<Box<TextRun>>) -> Fragment

Transforms this fragment using the given SplitInfo, preserving all the other data.

fn transform_into_ellipsis(&self, layout_context: &LayoutContext) -> Fragment

Transforms this fragment into an ellipsis fragment, preserving all the other data.

fn restyle_damage(&self) -> RestyleDamage

fn add_inline_context_style(&mut self, style: Arc<ComputedValues>, first_frag: bool, last_frag: bool)

Adds a style to the inline context for this fragment. If the inline context doesn't exist yet, it will be created.

fn surrounding_intrinsic_inline_size(&self) -> Au

Returns the portion of the intrinsic inline-size that consists of borders, padding, and/or margins.

FIXME(#2261, pcwalton): This won't work well for inlines: is this OK?

fn calculate_line_height(&self, layout_context: &LayoutContext) -> Au

fn border_width(&self) -> LogicalMargin<Au>

Returns the sum of the inline-sizes of all the borders of this fragment. Note that this can be expensive to compute, so if possible use the border_padding field instead.

fn compute_inline_direction_margins(&mut self, containing_block_inline_size: Au)

Computes the margins in the inline direction from the containing block inline-size and the style. After this call, the inline direction of the margin field will be correct.

Do not use this method if the inline direction margins are to be computed some other way (for example, via constraint solving for blocks).

fn compute_block_direction_margins(&mut self, containing_block_inline_size: Au)

Computes the margins in the block direction from the containing block inline-size and the style. After this call, the block direction of the margin field will be correct.

Do not use this method if the block direction margins are to be computed some other way (for example, via constraint solving for absolutely-positioned flows).

fn compute_border_and_padding(&mut self, containing_block_inline_size: Au)

Computes the border and padding in both inline and block directions from the containing block inline-size and the style. After this call, the border_padding field will be correct.

fn relative_position(&self, containing_block_size: &LogicalSize<Au>) -> LogicalSize<Au>

fn clear(&self) -> Option<ClearType>

Always inline for SCCP.

FIXME(pcwalton): Just replace with the clear type from the style module for speed?

fn style<'a>(&'a self) -> &'a ComputedValues

fn text_align(&self) -> T

Returns the text alignment of the computed style of the nearest ancestor-or-self Element node.

fn vertical_align(&self) -> T

fn white_space(&self) -> T

fn text_decoration(&self) -> T

Returns the text decoration of this fragment, according to the style of the nearest ancestor element.

NB: This may not be the actual text decoration, because of the override rules specified in CSS 2.1  16.3.1. Unfortunately, computing this properly doesn't really fit into Servo's model. Therefore, this is a best lower bound approximation, but the end result may actually have the various decoration flags turned on afterward.

fn inline_start_offset(&self) -> Au

Returns the inline-start offset from margin edge to content edge.

FIXME(#2262, pcwalton): I think this method is pretty bogus, because it won't work for inlines.

fn can_split(&self) -> bool

Returns true if this element can be split. This is true for text fragments.

fn newline_positions(&self) -> Option<&Vec<CharIndex>>

Returns the newline positions of this fragment, if it's a scanned text fragment.

fn newline_positions_mut(&mut self) -> Option<&mut Vec<CharIndex>>

Returns the newline positions of this fragment, if it's a scanned text fragment.

fn is_generated_content(&self) -> bool

Returns true if and only if this fragment is a generated content fragment.

fn is_scanned_text_fragment(&self) -> bool

Returns true if and only if this is a scanned text fragment.

fn compute_intrinsic_inline_sizes(&mut self) -> IntrinsicISizesContribution

Computes the intrinsic inline-sizes of this fragment.

fn content_inline_size(&self) -> Au

TODO: What exactly does this function return? Why is it Au(0) for SpecificFragmentInfo::Generic?

fn content_block_size(&self, layout_context: &LayoutContext) -> Au

Returns, and computes, the block-size of this fragment.

fn content_box(&self) -> LogicalRect<Au>

Returns the dimensions of the content box.

This is marked #[inline] because it is frequently called when only one or two of the values are needed and that will save computation.

fn find_split_info_by_new_line(&self) -> Option<(SplitInfo, Option<SplitInfo>, Arc<Box<TextRun>>)>

Find the split of a fragment that includes a new-line character.

A return value of None indicates that the fragment is not splittable. Otherwise the split information is returned. The right information is optional due to the possibility of it being whitespace.

fn calculate_split_position(&self, max_inline_size: Au, starts_line: bool) -> Option<SplitResult>

Attempts to find the split positions of a text fragment so that its inline-size is no more than max_inline_size.

A return value of None indicates that the fragment could not be split. Otherwise the information pertaining to the split is returned. The inline-start and inline-end split information are both optional due to the possibility of them being whitespace.

fn truncate_to_inline_size(&self, max_inline_size: Au) -> Option<TruncationResult>

Truncates this fragment to the given max_inline_size, using a character-based breaking strategy. If no characters could fit, returns None.

fn strip_trailing_whitespace_if_necessary(&mut self) -> bool

Attempts to strip trailing whitespace from this fragment by adjusting the text run range. Returns true if any modifications were made.

fn is_ignorable_whitespace(&self) -> bool

Returns true if this fragment is an unscanned text fragment that consists entirely of whitespace that should be stripped.

fn assign_replaced_inline_size_if_necessary<'a>(&'a mut self, container_inline_size: Au)

Assigns replaced inline-size, padding, and margins for this fragment only if it is replaced content per CSS 2.1  10.3.2.

fn assign_replaced_block_size_if_necessary(&mut self, containing_block_block_size: Au)

Assign block-size for this fragment if it is replaced content. The inline-size must have been assigned first.

Ideally, this should follow CSS 2.1  10.6.2.

fn inline_metrics(&self, layout_context: &LayoutContext) -> InlineMetrics

Calculates block-size above baseline, depth below baseline, and ascent for this fragment when used in an inline formatting context. See CSS 2.1  10.8.1.

fn is_hypothetical(&self) -> bool

Returns true if this fragment is a hypothetical box. See CSS 2.1  10.3.7.

fn can_merge_with_fragment(&self, other: &Fragment) -> bool

Returns true if this fragment can merge with another adjacent fragment or false otherwise.

fn is_primary_fragment(&self) -> bool

Returns true if and only if this is the primary fragment for the fragment's style object (conceptually, though style sharing makes this not really true, of course). The primary fragment is the one that draws backgrounds, borders, etc., and takes borders, padding and margins into account. Every style object has at most one primary fragment.

At present, all fragments are primary fragments except for inline-block and table wrapper fragments. Inline-block fragments are not primary fragments because the corresponding block flow is the primary fragment, while table wrapper fragments are not primary fragments because the corresponding table flow is the primary fragment.

fn update_late_computed_inline_position_if_necessary(&mut self)

fn update_late_computed_block_position_if_necessary(&mut self)

fn repair_style(&mut self, new_style: &Arc<ComputedValues>)

fn stacking_relative_border_box(&self, stacking_relative_flow_origin: &Point2D<Au>, relative_containing_block_size: &LogicalSize<Au>, relative_containing_block_mode: WritingMode, coordinate_system: CoordinateSystem) -> Rect<Au>

Given the stacking-context-relative position of the containing flow, returns the border box of this fragment relative to the parent stacking context. This takes position: relative into account.

If coordinate_system is Parent, this returns the border box in the parent stacking context's coordinate system. Otherwise, if coordinate_system is Own and this fragment establishes a stacking context itself, this returns a border box anchored at (0, 0). (If this fragment does not establish a stacking context, then it always belongs to its parent stacking context and thus coordinate_system is ignored.)

This is the method you should use for display list construction as well as getBoundingClientRect() and so forth.

fn stacking_relative_content_box(&self, stacking_relative_border_box: &Rect<Au>) -> Rect<Au>

Given the stacking-context-relative border box, returns the stacking-context-relative content box.

fn establishes_stacking_context(&self) -> bool

Returns true if this fragment establishes a new stacking context and false otherwise.

fn compute_overflow(&self) -> Rect<Au>

Computes the overflow rect of this fragment relative to the start of the flow.

fn remove_compositor_layers(&self, constellation_chan: ConstellationChan)

Remove any compositor layers associated with this fragment - it is being removed from the tree or had its display property set to none. TODO(gw): This just hides the compositor layer for now. In the future it probably makes sense to provide a hint to the compositor whether the layers should be destroyed to free memory.