Struct layout::fragment::IframeFragmentInfo

pub struct IframeFragmentInfo {
    pub pipeline_id: PipelineId,
    pub subpage_id: SubpageId,
}

A fragment that represents an inline frame (iframe). This stores the pipeline ID so that the size of this iframe can be communicated via the constellation to the iframe's own layout task.

Fields

pipeline_id

The pipeline ID of this iframe.

subpage_id

The subpage ID of this iframe.

Methods

impl IframeFragmentInfo

fn new(node: &ThreadSafeLayoutNode) -> IframeFragmentInfo

Creates the information specific to an iframe fragment.

fn calculate_replaced_inline_size(style: &ComputedValues, containing_size: Au) -> Au

fn calculate_replaced_block_size(style: &ComputedValues, containing_size: Au) -> Au