Struct layout::fragment::ImageFragmentInfo

pub struct ImageFragmentInfo {
    pub replaced_image_fragment_info: ReplacedImageFragmentInfo,
    pub image: Option<Arc<Image>>,
    pub metadata: Option<ImageMetadata>,
}

A fragment that represents a replaced content image and its accompanying borders, shadows, etc.

Fields
replaced_image_fragment_info: ReplacedImageFragmentInfo

The image held within this fragment.

image: Option<Arc<Image>> metadata: Option<ImageMetadata>
Methods

impl ImageFragmentInfo

fn new<N: ThreadSafeLayoutNode>(node: &N,
url: Option<Url>,
shared_layout_context: &SharedLayoutContext)
-> ImageFragmentInfo

Creates a new image fragment from the given URL and local image cache.

FIXME(pcwalton): The fact that image fragments store the cache in the fragment makes little sense to me.

fn image_inline_size(&mut self) -> Au

Returns the original inline-size of the image.

fn image_block_size(&mut self) -> Au

Returns the original block-size of the image.

fn tile_image_round(position: &mut Au,
size: &mut Au,
absolute_anchor_origin: Au,
image_size: &mut Au)

fn tile_image_spaced(position: &mut Au,
size: &mut Au,
tile_spacing: &mut Au,
absolute_anchor_origin: Au,
image_size: Au)

fn tile_image(position: &mut Au,
size: &mut Au,
absolute_anchor_origin: Au,
image_size: Au)

Tile an image