Struct layout::fragment::ScannedTextFragmentInfo

pub struct ScannedTextFragmentInfo {
    pub run: Arc<Box<TextRun>>,
    pub range: Range<CharIndex>,
    pub new_line_pos: Vec<CharIndex>,
    pub original_new_line_pos: Option<Vec<CharIndex>>,
    pub content_size: LogicalSize<Au>,
}

A scanned text fragment represents a single run of text with a distinct style. A TextFragment may be split into two or more fragments across line breaks. Several TextFragments may correspond to a single DOM text node. Split text fragments are implemented by referring to subsets of a single TextRun object.

Fields

run

The text run that this represents.

range

The range within the above text run that this represents.

new_line_pos

The positions of newlines within this scanned text fragment.

FIXME(#2260, pcwalton): Can't this go somewhere else, like in the text run or something? Or can we just remove it?

original_new_line_pos

The new_line_pos is eaten during line breaking. If we need to re-merge fragments, it will have to be restored.

content_size

The intrinsic size of the text fragment.

Methods

impl ScannedTextFragmentInfo

fn new(run: Arc<Box<TextRun>>, range: Range<CharIndex>, new_line_positions: Vec<CharIndex>, content_size: LogicalSize<Au>) -> ScannedTextFragmentInfo

Creates the information specific to a scanned text fragment from a range and a text run.