Struct layout::fragment::SplitResult

pub struct SplitResult {
    pub inline_start: Option<SplitInfo>,
    pub inline_end: Option<SplitInfo>,
    pub text_run: Arc<Box<TextRun>>,
}

Describes how to split a fragment into two. This contains up to two SplitInfos.

Fields

inline_start

The part of the fragment that goes on the first line.

inline_end

The part of the fragment that goes on the second line.

text_run

The text run which is being split.