Struct layout::fragment::UnscannedTextFragmentInfo

pub struct UnscannedTextFragmentInfo {
    pub text: Box<String>,
}

Data for an unscanned text fragment. Unscanned text fragments are the results of flow construction that have not yet had their inline-size determined.

Fields

text

The text inside the fragment.

FIXME(pcwalton): Is there something more clever we can do here that avoids the double indirection while not penalizing all fragments?

Methods

impl UnscannedTextFragmentInfo

fn from_text(text: String) -> UnscannedTextFragmentInfo

Creates a new instance of UnscannedTextFragmentInfo from the given text.