layout::fragment_tree

Module box_fragment

Structs

BlockLevelLayoutInfo 🔒

BoxFragment 🔒

BoxFragmentRareData 🔒

BoxFragmentWithStyle 🔒

Contains &Arc<BoxFragment> and dereferences to it so it can mostly be used in the same ways, except the .style() method is shadowed to use an existing atomic_refcell::AtomicRef that lives as long as BoxFragmentWithStyle.

ExtraBackground 🔒

Enums

BackgroundMode 🔒

Describes how a BoxFragment paints its background.

SpecificLayoutInfo 🔒