layout::fragment_tree::box_fragment

Struct BoxFragment

pub(crate) struct BoxFragment {

    pub base: BaseFragment,
    pub style: ArcRefCell<Arc<ComputedValues>>,
    pub children: Vec<Fragment>,
    pub cumulative_containing_block_rect: SyncPhysicalRectAu,
    pub padding: PhysicalSides<Au>,
    pub border: PhysicalSides<Au>,
    pub margin: PhysicalSides<Au>,
    baselines: Baselines,
    scrollable_overflow: SyncPhysicalRectAu,
    scrollable_overflow_is_up_to_date: AtomicBool,
    pub background_mode: BackgroundMode,
    pub rare_data: OnceBox<AtomicRefCell<BoxFragmentRareData>>,
    pub block_level_layout_info: Option<Box<BlockLevelLayoutInfo>>,
    pub spatial_tree_node: AtomicOptionScrollTreeNodeId,

}

Fields
base: BaseFragmentstyle: ArcRefCell<Arc<ComputedValues>>

The style for this BoxFragment.

children: Vec<Fragment>cumulative_containing_block_rect: SyncPhysicalRectAu

This BoxFragment’s containing block rectangle in coordinates relative to the initial containing block, but not taking into account any transforms.

padding: PhysicalSides<Au>border: PhysicalSides<Au>margin: PhysicalSides<Au>baselines: Baselines

When this BoxFragment is for content that has a baseline, this tracks the first and last baselines of that content. This is used to propagate baselines to things such as tables and inline formatting contexts.

scrollable_overflow: SyncPhysicalRectAu

The scrollable overflow of this box fragment in the same coordiante system as Self::content_rect ie a rectangle within the parent fragment’s content rectangle. This does not take into account any transforms this fragment applies. This is handled when calling [Self::scrollable_overflow_for_parent].

scrollable_overflow_is_up_to_date: AtomicBoolbackground_mode: BackgroundModerare_data: OnceBox<AtomicRefCell<BoxFragmentRareData>>

Rare data that not all kinds of BoxFragment would have.

block_level_layout_info: Option<Box<BlockLevelLayoutInfo>>

Additional information for block-level boxes.

spatial_tree_node: AtomicOptionScrollTreeNodeId

The containing spatial tree node of this BoxFragment. This is assigned during StackingContextTree construction, so isn’t available before that time. This is used to for determining final viewport size and position of this node and will also be used in the future for hit testing.

Implementations

impl BoxFragment

pub(crate) fn stacking_context_type(&self) -> Option<StackingContextType>

fn build_stacking_context_tree( self: &Arc<Self>, fragment: Fragment, stacking_context_tree: &mut StackingContextTree, containing_block: &ContainingBlock, containing_block_info: &ContainingBlockManager<'_, ContainingBlock>, parent_stacking_context: &mut StackingContext, text_decorations: &Rc<Vec<FragmentTextDecoration>>, )

fn build_stacking_context_tree_maybe_creating_reference_frame( self: &Arc<Self>, fragment: Fragment, stacking_context_tree: &mut StackingContextTree, containing_block: &ContainingBlock, containing_block_info: &ContainingBlockManager<'_, ContainingBlock>, parent_stacking_context: &mut StackingContext, text_decorations: &Rc<Vec<FragmentTextDecoration>>, )

fn build_stacking_context_tree_maybe_creating_stacking_context( self: &Arc<Self>, fragment: Fragment, stacking_context_tree: &mut StackingContextTree, containing_block: &ContainingBlock, containing_block_info: &ContainingBlockManager<'_, ContainingBlock>, parent_stacking_context: &mut StackingContext, text_decorations: &Rc<Vec<FragmentTextDecoration>>, reference_frame_info: Option<StackingContextReferenceFrameInfo>, )

impl BoxFragment

fn build_sticky_frame_if_necessary( &self, stacking_context_tree: &mut StackingContextTree, parent_scroll_node_id: ScrollTreeNodeId, containing_block_rect: &PhysicalRect<Au>, scroll_frame_size: &Option<LayoutSize>, established_scroll_frame: bool, ) -> Option<ScrollTreeNodeId>

fn reference_frame_data_if_necessary( &self, containing_block_rect: &PhysicalRect<Au>, ) -> Option<ReferenceFrameData>

Optionally returns the data for building a reference frame, without yet building it.

pub fn calculate_transform_matrix( &self, border_rect: &Rect<Au, CSSPixel>, ) -> Option<LayoutTransform>

Returns the 4D matrix representing this fragment’s transform.

pub fn calculate_perspective_matrix( &self, border_rect: &Rect<Au, CSSPixel>, ) -> Option<LayoutTransform>

Returns the 4D matrix representing this fragment’s perspective.

fn clear_stacking_context_tree_traversal_data_recursively(&self)

impl BoxFragment

fn border_radius(&self) -> BorderRadius

impl BoxFragment

pub(crate) fn new( base_fragment_info: BaseFragmentInfo, style: ServoArc<ComputedValues>, children: Vec<Fragment>, content_rect: PhysicalRect<Au>, padding: PhysicalSides<Au>, border: PhysicalSides<Au>, margin: PhysicalSides<Au>, specific_layout_info: Option<SpecificLayoutInfo>, ) -> Self

pub(crate) fn with_baselines(self, baselines: Baselines) -> Self

pub(crate) fn style<'a>(&'a self) -> AtomicRef<'a, ServoArc<ComputedValues>>

pub(crate) fn with_style(self: &Arc<Self>) -> BoxFragmentWithStyle<'_>

pub(crate) fn baselines(&self, writing_mode: WritingMode) -> Baselines

Get the baselines for this BoxFragment if they are compatible with the given WritingMode. If they are not compatible, Baselines::default() is returned.

pub(crate) fn add_extra_background(&mut self, extra_background: ExtraBackground)

pub(crate) fn set_does_not_paint_background(&mut self)

pub(crate) fn ensure_rare_data(&self) -> AtomicRefMut<'_, BoxFragmentRareData>

pub(crate) fn specific_layout_info( &self, ) -> Option<AtomicRef<'_, SpecificLayoutInfo>>

pub(crate) fn clear_stacking_context_tree_traversal_data(&self)

pub(crate) fn resolved_sticky_insets( &self, ) -> Option<AtomicRef<'_, Box<PhysicalSides<AuOrAuto>>>>

pub(crate) fn set_resolved_sticky_insets( &self, sticky_insets: PhysicalSides<AuOrAuto>, )

pub(crate) fn generated_clip_id(&self) -> Option<ClipId>

pub(crate) fn set_generated_clip_id(&self, generated_clip_id: ClipId)

pub(crate) fn generated_scroll_tree_node_id(&self) -> Option<ScrollTreeNodeId>

pub(crate) fn set_generated_scroll_tree_node_id( &self, generated_scroll_tree_node_id: ScrollTreeNodeId, )

pub(crate) fn with_block_level_layout_info( self, block_margins_collapsed_with_children: CollapsedBlockMargins, clearance: Option<Au>, ) -> Self

pub(crate) fn clear_scrollable_overflow(&self)

Clear the scrollable overflow on this BoxFragment. This is called during damage propagation when a fragment is preserved, itself or one of its descendants has scrollable overflow damage.

pub(crate) fn set_containing_block(&self, containing_block: &PhysicalRect<Au>)

pub(crate) fn offset_by_containing_block( &self, rect: &PhysicalRect<Au>, containing_block_computation: ContainingBlockCalculation<'_>, ) -> PhysicalRect<Au>

pub(crate) fn cumulative_content_box_rect( &self, containing_block_computation: ContainingBlockCalculation<'_>, ) -> PhysicalRect<Au>

pub(crate) fn cumulative_padding_box_rect( &self, containing_block_computation: ContainingBlockCalculation<'_>, ) -> PhysicalRect<Au>

pub(crate) fn cumulative_border_box_rect( &self, containing_block_computation: ContainingBlockCalculation<'_>, ) -> PhysicalRect<Au>

pub(crate) fn content_rect(&self) -> PhysicalRect<Au>

pub(crate) fn padding_rect(&self) -> PhysicalRect<Au>

pub(crate) fn border_rect(&self) -> PhysicalRect<Au>

pub(crate) fn margin_rect(&self) -> PhysicalRect<Au>

pub(crate) fn padding_border_margin(&self) -> PhysicalSides<Au>

pub(crate) fn is_root_element(&self) -> bool

pub(crate) fn is_body_element_of_html_element_root(&self) -> bool

pub(crate) fn print(self: &Arc<Self>, tree: &mut PrintTree)

pub(crate) fn calculate_resolved_insets_if_positioned( &self, containing_block_computation: ContainingBlockCalculation<'_>, ) -> PhysicalSides<AuOrAuto>

pub(crate) fn is_flex_or_grid_item(&self) -> bool

Whether or this is a flex or grid item.

pub(crate) fn is_replaced(&self) -> bool

Whether or this box is for replaced content.

pub(crate) fn is_table_wrapper(&self) -> bool

Whether this is a table wrapper box. https://www.w3.org/TR/css-tables-3/#table-wrapper-box

pub(crate) fn is_table_grid_with_collapsed_borders(&self) -> bool

Whether or not this is the BoxFragment for a table grid with collapsed borders.

pub(crate) fn spatial_tree_node(&self) -> Option<ScrollTreeNodeId>