layout::fragment_tree::box_fragment

Struct BoxFragmentWithStyle

pub(crate) struct BoxFragmentWithStyle<'a> {
    pub(crate) box_fragment: &'a Arc<BoxFragment>,
    pub(crate) style: AtomicRef<'a, Arc<ComputedValues>>,
}

Contains &Arc<BoxFragment> and dereferences to it so it can mostly be used in the same ways, except the .style() method is shadowed to use an existing atomic_refcell::AtomicRef that lives as long as BoxFragmentWithStyle.

Compared to calling BoxFragment::style() repeatedly, this reduce the number of atomic increments and decrements on ArcRefCell’s borrow counter.

Fields
box_fragment: &'a Arc<BoxFragment>style: AtomicRef<'a, Arc<ComputedValues>>
Implementations

impl BoxFragmentWithStyle<'_>

fn build_stacking_context_tree_for_children( &self, stacking_context_tree: &mut StackingContextTree, containing_block: &ContainingBlock, containing_block_info: &ContainingBlockManager<'_, ContainingBlock>, stacking_context: &mut StackingContext, text_decorations: &Rc<Vec<FragmentTextDecoration>>, )

fn build_clip_frame_if_necessary( &self, stacking_context_tree: &mut StackingContextTree, parent_scroll_node_id: ScrollTreeNodeId, parent_clip_id: ClipId, containing_block_rect: &PhysicalRect<Au>, ) -> Option<ClipId>

fn build_overflow_frame_if_necessary( &self, stacking_context_tree: &mut StackingContextTree, parent_scroll_node_id: ScrollTreeNodeId, parent_clip_id: ClipId, containing_block_rect: &PhysicalRect<Au>, ) -> Option<OverflowFrameData>

impl<'a> BoxFragmentWithStyle<'a>

pub(crate) fn style(&self) -> &ServoArc<ComputedValues>

pub(crate) fn clip_wholly_unreachable_scrollable_overflow( &self, scrollable_overflow_from_child: PhysicalRect<Au>, clipping_rect: PhysicalRect<Au>, ) -> PhysicalRect<Au>

Return the clipped scrollable overflow based on its scroll origin, determined by overflow direction. Return None if the scrollable overflow from child is wholly unreachable. For an element, the clip rect is the padding rect and for viewport, it is the initial containing block.

pub(crate) fn scrollable_overflow_for_parent(&self) -> PhysicalRect<Au>

pub(crate) fn scrollable_overflow(&self) -> PhysicalRect<Au>

Get the scrollable overflow for this BoxFragment relative to its containing block, recalculating scrollable overflow when necessary, for instance after a style change.

fn calculate_scrollable_overflow(&self) -> PhysicalRect<Au>

This is an implementation of:

https://drafts.csswg.org/css-overflow-3/#scrollable.

https://drafts.csswg.org/cssom-view/#scrolling-area

pub(crate) fn is_inline_box(&self) -> bool

Whether this is a non-replaced inline-level box whose inner display type is flow. https://drafts.csswg.org/css-display-3/#inline-box

pub(crate) fn is_atomic_inline_level(&self) -> bool

Whether this is an atomic inline-level box. https://drafts.csswg.org/css-display-3/#atomic-inline

pub(crate) fn has_collapsed_borders(&self) -> bool

pub(crate) fn has_outline(&self) -> bool

Whether or not this BoxFragment has outlines.