layout::fragment_tree

Module fragment

Structs

CollapsedBlockMargins 🔒

CollapsedMargin 🔒

IFrameFragment 🔒

ImageFragment 🔒

LayoutRootFragment 🔒

TextFragment 🔒

Enums

ContainingBlockCalculation 🔒

A token which ensures the calculation and assignment of cumulative containing blocks to fragments in the fragment tree. This is used because these cumulative containing block offsets are set during stacking context tree construction, but some queries might need them beforehand. If the query is executed before stacking context tree construction, a quick traversal is performed to calculate them for the purpose of the query.

Fragment 🔒