layout::geom

Type Alias PhysicalSize

pub type PhysicalSize<U> = Size2D<U, CSSPixel>;

Aliased Type

pub struct PhysicalSize<U> {
    pub width: U,
    pub height: U,
}

Fields
width: U

The extent of the element in the U units along the x axis (usually horizontal).

height: U

The extent of the element in the U units along the y axis (usually vertical).