Module layout::incremental

Structs

RestyleDamage
Individual layout actions that may be necessary after restyling.

SpecialRestyleDamage

Constants

BUBBLE_ISIZES
Recompute intrinsic inline_sizes (minimum and preferred). Propagates down the flow tree because the computation is bottom-up.

RECONSTRUCT_FLOW
The entire flow needs to be reconstructed.

REFLOW
Recompute actual inline_sizes and block_sizes. Propagates up the flow tree because the computation is top-down.

REFLOW_ENTIRE_DOCUMENT
If this flag is set, we need to reflow the entire document. This is more or less a temporary hack to deal with cases that we don't handle incrementally yet.

REFLOW_OUT_OF_FLOW
Recompute actual inline-sizes and block-sizes, only taking out-of-flow children into account. Propagates up the flow tree because the computation is top-down.

REPAINT
Repaint the node itself. Currently unused; need to decide how this propagates.

RESOLVE_GENERATED_CONTENT
Re-resolve generated content. Propagates up the flow tree because the computation is inorder.

Traits

LayoutDamageComputation

Functions

compute_damage

rebuild_and_reflow
Returns a bitmask that represents a flow that needs to be rebuilt and reflowed.