Module layout::inline

Structs

FragmentIndex
The index of a fragment in a flattened vector of DOM elements.

InlineFlow
Flows for inline layout.

InlineFragmentContext

InlineFragmentNodeFlags

InlineFragmentNodeInfo

InlineFragments
Represents a list of inline fragments, including element ranges.

InlineMetrics
Ascent and space needed above and below the baseline for a fragment. See CSS 2.1  10.8.1.

Line
Lines are represented as offsets into the child list, rather than as an object that "owns" fragments. Choosing a different set of line breaks requires a new list of offsets, and possibly some splitting and merging of TextFragments.

LineBreaker
Arranges fragments into lines, splitting them up as necessary.

LineMetrics

Enums

LineFlushMode

Functions

inline_contexts_are_equal