Struct layout::inline::InlineFragments

pub struct InlineFragments {
    pub fragments: Vec<Fragment>,
}

Represents a list of inline fragments, including element ranges.

Fields

fragments

The fragments themselves.

Methods

impl InlineFragments

fn new() -> InlineFragments

Creates an empty set of inline fragments.

fn len(&self) -> usize

Returns the number of inline fragments.

fn is_empty(&self) -> bool

Returns true if this list contains no fragments and false if it contains at least one fragment.

fn push(&mut self, fragment: &mut Fragment)

Pushes a new inline fragment.

fn push_all(&mut self, other: InlineFragments)

Merges another set of inline fragments with this one.

fn get<'a>(&'a self, index: usize) -> &'a Fragment

A convenience function to return the fragment at a given index.

fn get_mut<'a>(&'a mut self, index: usize) -> &'a mut Fragment

A convenience function to return a mutable reference to the fragment at a given index.