Struct layout::inline::InlineMetrics

pub struct InlineMetrics {
    pub space_above_baseline: Au,
    pub space_below_baseline: Au,
    pub ascent: Au,
}

Ascent and space needed above and below the baseline for a fragment. See CSS 2.1  10.8.1.

Descent is not included in this structure because it can be computed from the fragment's border/content box and the ascent.

Fields
space_above_baseline: Au

The amount of space above the baseline needed for this fragment.

space_below_baseline: Au

The amount of space below the baseline needed for this fragment.

ascent: Au

The distance from the baseline to the top of this fragment. This can differ from block_size_above_baseline if the fragment needs some empty space above it due to line-height, etc.

Methods

impl InlineMetrics

fn new(space_above_baseline: Au,
space_below_baseline: Au,
ascent: Au)
-> InlineMetrics

Creates a new set of inline metrics.

fn from_font_metrics(font_metrics: &FontMetrics,
line_height: Au)
-> InlineMetrics

Calculates inline metrics from font metrics and line block-size per CSS 2.1  10.8.1.

fn space_needed(&self) -> Au

Returns the sum of the space needed above and below the baseline.