Module layout::layout_debug

Supports writing a trace file created during each layout scope that can be viewed by an external tool to make layout debugging easier.

Structs

Scope

ScopeData

State

Constants

STATE_KEY

Statics

DEBUG_ID_COUNTER

Functions

begin_trace
Begin a layout debug trace. If this has not been called, creating debug scopes has no effect.

end_trace
End the debug layout trace. This will write the layout trace to disk in the current directory. The output file can then be viewed with an external tool.

generate_unique_debug_id
Generate a unique ID. This is used for items such as Fragment which are often reallocated but represent essentially the same data.