Struct layout::layout_debug::Scope

pub struct Scope;

Methods

impl Scope

A layout debugging scope. The entire state of the flow tree will be output at the beginning and end of this scope.

fn new(name: String) -> Scope