layout::layout_impl

Struct LayoutThread

pub struct LayoutThread {

    id: PipelineId,
    webview_id: WebViewId,
    url: ServoUrl,
    stylist: Stylist,
    is_iframe: bool,
    script_chan: GenericSender<ScriptThreadMessage>,
    time_profiler_chan: ProfilerChan,
    image_cache: Arc<dyn ImageCache>,
    font_context: Arc<FontContext>,
    have_added_user_agent_stylesheets: bool,
    device_has_changed: bool,
    have_ever_generated_display_list: Cell<bool>,
    need_overflow_calculation: Cell<bool>,
    need_new_display_list: Cell<bool>,
    need_new_stacking_context_tree: Cell<bool>,
    box_tree: RefCell<Option<Arc<BoxTree>>>,
    fragment_tree: RefCell<Option<Rc<FragmentTree>>>,
    stacking_context_tree: RefCell<Option<StackingContextTree>>,
    resolved_images_cache: Arc<RwLock<HashMap<ServoUrl, Result<Image, ResolveImageError>>>>,
    registered_painters: RegisteredPaintersImpl,
    paint_api: CrossProcessPaintApi,
    debug: DiagnosticsLogging,
    previously_highlighted_dom_node: Cell<Option<OpaqueNode>>,
    lcp_candidate_collector: RefCell<Option<LargestContentfulPaintCandidateCollector>>,

}

Information needed by layout.

Fields
id: PipelineId

The ID of the pipeline that we belong to.

webview_id: WebViewId

The webview that contains the pipeline we belong to.

url: ServoUrl

The URL of the pipeline that we belong to.

stylist: Stylist

Performs CSS selector matching and style resolution.

is_iframe: bool

Is the current reflow of an iframe, as opposed to a root window?

script_chan: GenericSender<ScriptThreadMessage>

The channel on which messages can be sent to the script thread.

time_profiler_chan: ProfilerChan

The channel on which messages can be sent to the time profiler.

image_cache: Arc<dyn ImageCache>

Reference to the script thread image cache.

font_context: Arc<FontContext>

A FontContext to be used during layout.

have_added_user_agent_stylesheets: bool

Whether or not user agent stylesheets have been added to the Stylist or not.

device_has_changed: bool

Whether or not this [LayoutImpl]’s Device has changed since the last restyle. If it has, a restyle is pending.

have_ever_generated_display_list: Cell<bool>

Is this the first reflow in this LayoutThread?

need_overflow_calculation: Cell<bool>

Whether a new overflow calculation needs to happen due to changes to the fragment tree. This is set to true every time a restyle requests overflow calculation.

need_new_display_list: Cell<bool>

Whether a new display list is necessary due to changes to layout or stacking contexts. This is set to true every time layout changes, even when a display list isn’t requested for this layout, such as for layout queries. The next time a layout requests a display list, it is produced unconditionally, even when the layout trees remain the same.

need_new_stacking_context_tree: Cell<bool>

Whether or not the existing stacking context tree is dirty and needs to be rebuilt. This happens after a relayout or overflow update. The reason that we don’t simply clear the stacking context tree when it becomes dirty is that we need to preserve scroll offsets from the old tree to the new one.

box_tree: RefCell<Option<Arc<BoxTree>>>

The box tree.

fragment_tree: RefCell<Option<Rc<FragmentTree>>>

The fragment tree.

stacking_context_tree: RefCell<Option<StackingContextTree>>

The StackingContextTree cached from previous layouts.

resolved_images_cache: Arc<RwLock<HashMap<ServoUrl, Result<Image, ResolveImageError>>>>registered_painters: RegisteredPaintersImpl

The executors for paint worklets.

paint_api: CrossProcessPaintApi

Cross-process access to the Paint API.

debug: DiagnosticsLogging

Debug options, copied from configuration to this LayoutThread in order to avoid having to constantly access the thread-safe global options.

previously_highlighted_dom_node: Cell<Option<OpaqueNode>>

Tracks the node that was highlighted by the devtools during the last reflow.

If this changed, then we need to create a new display list.

lcp_candidate_collector: RefCell<Option<LargestContentfulPaintCandidateCollector>>

The collector for calculating Largest Contentful Paint

Implementations

impl LayoutThread

fn new(config: LayoutConfig) -> LayoutThread

fn build_shared_style_context<'a>( &'a self, guards: StylesheetGuards<'a>, snapshot_map: &'a SnapshotMap, animation_timeline_value: f64, animations: &DocumentAnimationSet, traversal_flags: TraversalFlags, ) -> SharedStyleContext<'a>

fn load_all_web_fonts_from_stylesheet_with_guard( &self, stylesheet: &DocumentStyleSheet, guard: &SharedRwLockReadGuard<'_>, document_context: &WebFontDocumentContext, )

fn can_skip_reflow_request_entirely( &self, reflow_request: &ReflowRequest, ) -> bool

In some cases, if a restyle isn’t necessary we can skip doing any work for layout entirely. This check allows us to return early from layout without doing any work at all.

fn maybe_print_reflow_event(&self, reflow_request: &ReflowRequest)

fn handle_update_scroll_node_request( &self, reflow_request: &ReflowRequest, ) -> bool

Checks whether we need to update the scroll node, and report whether the node is scrolled. We need to update the scroll node whenever it is requested.

fn handle_reflow( &mut self, reflow_request: ReflowRequest, ) -> Option<ReflowResult>

The high-level routine that performs layout.

fn prepare_stylist_for_reflow<'dom>( &mut self, reflow_request: &ReflowRequest, document: ServoLayoutDocument<'dom>, root_element: ServoLayoutElement<'dom>, guards: &StylesheetGuards<'_>, ua_stylesheets: &UserAgentStylesheets, snapshot_map: &SnapshotMap, )

fn restyle_and_build_trees( &mut self, reflow_request: &mut ReflowRequest, document: ServoLayoutDocument<'_>, root_element: ServoLayoutElement<'_>, image_resolver: &Arc<ImageResolver>, ) -> (ReflowPhasesRun, IFrameSizes)

fn calculate_overflow(&self) -> bool

fn build_stacking_context_tree_for_reflow( &self, reflow_request: &ReflowRequest, ) -> bool

fn build_stacking_context_tree(&self, viewport_details: ViewportDetails) -> bool

fn build_display_list( &self, reflow_request: &ReflowRequest, image_resolver: &Arc<ImageResolver>, ) -> bool

Build the display list for the current layout and send it to the renderer. If no display list is built, returns false.

fn set_scroll_offset_from_script( &self, external_scroll_id: ExternalScrollId, offset: LayoutVector2D, ) -> bool

fn profiler_metadata(&self) -> Option<TimerMetadata>

Returns profiling information which is passed to the time profiler.