Struct layout::layout_task::LayoutTask

pub struct LayoutTask {
    pub id: PipelineId,
    pub url: Url,
    pub port: Receiver<Msg>,
    pub pipeline_port: Receiver<LayoutControlMsg>,
    pub chan: LayoutChan,
    pub constellation_chan: ConstellationChan,
    pub script_chan: ScriptControlChan,
    pub paint_chan: PaintChan,
    pub time_profiler_chan: ProfilerChan,
    pub mem_profiler_chan: ProfilerChan,
    pub reporter_name: String,
    pub resource_task: ResourceTask,
    pub image_cache_task: ImageCacheTask,
    pub font_cache_task: FontCacheTask,
    pub first_reflow: Cell<bool>,
    pub rw_data: Arc<Mutex<LayoutTaskData>>,
}

Information needed by the layout task.

Fields

id

The ID of the pipeline that we belong to.

url

The URL of the pipeline that we belong to.

port

The port on which we receive messages from the script task.

pipeline_port

The port on which we receive messages from the constellation

chan

The channel on which we or others can send messages to ourselves.

constellation_chan

The channel on which messages can be sent to the constellation.

script_chan

The channel on which messages can be sent to the script task.

paint_chan

The channel on which messages can be sent to the painting task.

time_profiler_chan

The channel on which messages can be sent to the time profiler.

mem_profiler_chan

The channel on which messages can be sent to the memory profiler.

reporter_name

The name used for the task's memory reporter.

resource_task

The channel on which messages can be sent to the resource task.

image_cache_task

The channel on which messages can be sent to the image cache.

font_cache_task

Public interface to the font cache task.

first_reflow

Is this the first reflow in this LayoutTask?

rw_data

A mutex to allow for fast, read-only RPC of layout's internal data structures, while still letting the LayoutTask modify them.

All the other elements of this struct are read-only.

Methods

impl LayoutTask

fn tick_animation<'a>(&'a self, animation: Animation, rw_data: &mut LayoutTaskData)