Struct layout::layout_task::LayoutTaskData

pub struct LayoutTaskData {
    pub root_flow: Option<FlowRef>,
    pub local_image_cache: Arc<Mutex<LocalImageCache<UntrustedNodeAddress>>>,
    pub constellation_chan: ConstellationChan,
    pub screen_size: Size2D<Au>,
    pub stacking_context: Option<Arc<StackingContext>>,
    pub stylist: Box<Stylist>,
    pub parallel_traversal: Option<WorkQueue<SharedLayoutContextWrapper, UnsafeFlow>>,
    pub dirty: Rect<Au>,
    pub generation: u32,
    pub content_box_response: Rect<Au>,
    pub content_boxes_response: Vec<Rect<Au>>,
    pub running_animations: Vec<Animation>,
    pub new_animations_receiver: Receiver<Animation>,
    pub new_animations_sender: Sender<Animation>,
}

Mutable data belonging to the LayoutTask.

This needs to be protected by a mutex so we can do fast RPCs.

Fields

root_flow

The root of the flow tree.

local_image_cache

The local image cache.

constellation_chan

The channel on which messages can be sent to the constellation.

screen_size

The size of the viewport.

stacking_context

The root stacking context.

stylist

Performs CSS selector matching and style resolution.

parallel_traversal

The workers that we use for parallel operation.

dirty

The dirty rect. Used during display list construction.

generation

Starts at zero, and increased by one every time a layout completes. This can be used to easily check for invalid stale data.

content_box_response

A queued response for the union of the content boxes of a node.

content_boxes_response

A queued response for the content boxes of a node.

running_animations

The list of currently-running animations.

new_animations_receiver

Receives newly-discovered animations.

new_animations_sender

A channel on which new animations that have been triggered by style recalculation can be sent.