Module layout::layout_thread

The layout thread. Performs layout on the DOM, builds display lists and sends them to be painted.

Structs

LayoutThread
Information needed by the layout thread.

LayoutThreadData
Mutable data belonging to the LayoutThread.

RwData

Enums

RWGuard
The LayoutThread rw_data lock must remain locked until the first reflow, as RPC calls don't make sense until then. Use this in combination with LayoutThread::lock_rw_data and LayoutThread::return_rw_data.

Constants

DISPLAY_PORT_SIZE_FACTOR
The number of screens of data we're allowed to generate display lists for in each direction.

DISPLAY_PORT_THRESHOLD_SIZE_FACTOR
The number of screens we have to traverse before we decide to generate new display lists.

Functions

add_font_face_rules

get_root_flow_background_color