Struct layout::layout_thread::LayoutThread

pub struct LayoutThread {
    id: PipelineId,
    url: Url,
    is_iframe: bool,
    port: Receiver<Msg>,
    pipeline_port: Receiver<LayoutControlMsg>,
    image_cache_receiver: Receiver<ImageCacheResult>,
    image_cache_sender: ImageCacheChan,
    font_cache_receiver: Receiver<()>,
    font_cache_sender: IpcSender<()>,
    constellation_chan: IpcSender<ConstellationMsg>,
    script_chan: IpcSender<ConstellationControlMsg>,
    paint_chan: OptionalIpcSender<LayoutToPaintMsg>,
    time_profiler_chan: ProfilerChan,
    mem_profiler_chan: ProfilerChan,
    image_cache_thread: ImageCacheThread,
    font_cache_thread: FontCacheThread,
    first_reflow: bool,
    parallel_traversal: Option<WorkQueue<SharedLayoutContext, WorkQueueData>>,
    generation: u32,
    new_animations_sender: Sender<Animation>,
    new_animations_receiver: Receiver<Animation>,
    outstanding_web_fonts: Arc<AtomicUsize>,
    root_flow: Option<FlowRef>,
    visible_rects: Arc<HashMap<LayerId, Rect<Au>, BuildHasherDefault<FnvHasher>>>,
    running_animations: Arc<RwLock<HashMap<OpaqueNode, Vec<Animation>>>>,
    expired_animations: Arc<RwLock<HashMap<OpaqueNode, Vec<Animation>>>>,
    epoch: Epoch,
    viewport_size: Size2D<Au>,
    rw_data: Arc<Mutex<LayoutThreadData>>,
    error_reporter: CSSErrorReporter,
    webrender_image_cache: Arc<RwLock<HashMap<(Url, UsePlaceholder), WebRenderImageInfo, BuildHasherDefault<FnvHasher>>>>,
    webrender_api: Option<RenderApi>,
}

Information needed by the layout thread.

Fields

id

The ID of the pipeline that we belong to.

url

The URL of the pipeline that we belong to.

is_iframe

Is the current reflow of an iframe, as opposed to a root window?

port

The port on which we receive messages from the script thread.

pipeline_port

The port on which we receive messages from the constellation.

image_cache_receiver

The port on which we receive messages from the image cache

image_cache_sender

The channel on which the image cache can send messages to ourself.

font_cache_receiver

The port on which we receive messages from the font cache thread.

font_cache_sender

The channel on which the font cache can send messages to us.

constellation_chan

The channel on which messages can be sent to the constellation.

script_chan

The channel on which messages can be sent to the script thread.

paint_chan

The channel on which messages can be sent to the painting thread.

time_profiler_chan

The channel on which messages can be sent to the time profiler.

mem_profiler_chan

The channel on which messages can be sent to the memory profiler.

image_cache_thread

The channel on which messages can be sent to the image cache.

font_cache_thread

Public interface to the font cache thread.

first_reflow

Is this the first reflow in this LayoutThread?

parallel_traversal

The workers that we use for parallel operation.

generation

Starts at zero, and increased by one every time a layout completes. This can be used to easily check for invalid stale data.

new_animations_sender

A channel on which new animations that have been triggered by style recalculation can be sent.

new_animations_receiver

Receives newly-discovered animations.

outstanding_web_fonts

The number of Web fonts that have been requested but not yet loaded.

root_flow

The root of the flow tree.

visible_rects

The position and size of the visible rect for each layer. We do not build display lists for any areas more than DISPLAY_PORT_SIZE_FACTOR screens away from this area.

running_animations

The list of currently-running animations.

expired_animations

The list of animations that have expired since the last style recalculation.

epoch

A counter for epoch messages

viewport_size

The size of the viewport. This may be different from the size of the screen due to viewport constraints.

rw_data

A mutex to allow for fast, read-only RPC of layout's internal data structures, while still letting the LayoutThread modify them.

All the other elements of this struct are read-only.

error_reporter

The CSS error reporter for all CSS loaded in this layout thread

webrender_image_cache

webrender_api

Methods

impl LayoutThread

fn new(id: PipelineId, url: Url, is_iframe: bool, port: Receiver<Msg>, pipeline_port: IpcReceiver<LayoutControlMsg>, constellation_chan: IpcSender<ConstellationMsg>, script_chan: IpcSender<ConstellationControlMsg>, paint_chan: OptionalIpcSender<LayoutToPaintMsg>, image_cache_thread: ImageCacheThread, font_cache_thread: FontCacheThread, time_profiler_chan: ProfilerChan, mem_profiler_chan: ProfilerChan, webrender_api_sender: Option<RenderApiSender>) -> LayoutThread

Creates a new LayoutThread structure.

fn start(self)

Starts listening on the port.

fn build_shared_layout_context(&self, rw_data: &LayoutThreadData, screen_size_changed: bool, goal: ReflowGoal) -> SharedLayoutContext

fn handle_request<'a, 'b>(&mut self, possibly_locked_rw_data: &mut RwData<'a, 'b>) -> bool

Receives and dispatches messages from the script and constellation threads

fn repaint<'a, 'b>(&mut self, possibly_locked_rw_data: &mut RwData<'a, 'b>) -> bool

Repaint the scene, without performing style matching. This is typically used when an image arrives asynchronously and triggers a relayout and repaint. TODO: In the future we could detect if the image size hasn't changed since last time and avoid performing a complete layout pass.

fn handle_request_helper<'a, 'b>(&mut self, request: Msg, possibly_locked_rw_data: &mut RwData<'a, 'b>) -> bool

Receives and dispatches messages from other threads.

fn collect_reports<'a, 'b>(&self, reports_chan: ReportsChan, possibly_locked_rw_data: &mut RwData<'a, 'b>)

fn create_layout_thread(&self, info: NewLayoutThreadInfo)

fn prepare_to_exit(&mut self, response_chan: Sender<()>)

Enters a quiescent state in which no new messages will be processed until an ExitNow is received. A pong is immediately sent on the given response channel.

fn exit_now(&mut self)

Shuts down the layout thread now. If there are any DOM nodes left, layout will now (safely) crash.

fn handle_add_stylesheet<'a, 'b>(&self, stylesheet: Arc<Stylesheet>, possibly_locked_rw_data: &mut RwData<'a, 'b>)

fn handle_set_quirks_mode<'a, 'b>(&self, possibly_locked_rw_data: &mut RwData<'a, 'b>)

Sets quirks mode for the document, causing the quirks mode stylesheet to be used.

fn try_get_layout_root<N: LayoutNode>(&self, node: N) -> Option<FlowRef>

fn solve_constraints(layout_root: &mut FlowRef, shared_layout_context: &SharedLayoutContext)

Performs layout constraint solving.

This corresponds to Reflow() in Gecko and layout() in WebKit/Blink and should be benchmarked against those two. It is marked #[inline(never)] to aid profiling.

fn solve_constraints_parallel(traversal: &mut WorkQueue<SharedLayoutContext, WorkQueueData>, layout_root: &mut FlowRef, profiler_metadata: Option<TimerMetadata>, time_profiler_chan: ProfilerChan, shared_layout_context: &SharedLayoutContext)

Performs layout constraint solving in parallel.

This corresponds to Reflow() in Gecko and layout() in WebKit/Blink and should be benchmarked against those two. It is marked #[inline(never)] to aid profiling.

fn compute_abs_pos_and_build_display_list(&mut self, data: &Reflow, layout_root: &mut FlowRef, shared_layout_context: &mut SharedLayoutContext, rw_data: &mut LayoutThreadData)

fn handle_reflow<'a, 'b>(&mut self, data: &ScriptReflow, possibly_locked_rw_data: &mut RwData<'a, 'b>)

The high-level routine that performs layout threads.

fn set_visible_rects<'a, 'b>(&mut self, new_visible_rects: Vec<(LayerId, Rect<Au>)>, possibly_locked_rw_data: &mut RwData<'a, 'b>) -> bool

fn set_stacking_context_scroll_states<'a, 'b>(&mut self, new_scroll_states: Vec<StackingContextScrollState>, _: &mut RwData<'a, 'b>)

fn tick_all_animations<'a, 'b>(&mut self, possibly_locked_rw_data: &mut RwData<'a, 'b>)

fn tick_animations(&mut self, rw_data: &mut LayoutThreadData)

fn reflow_with_newly_loaded_web_font<'a, 'b>(&mut self, possibly_locked_rw_data: &mut RwData<'a, 'b>)

fn perform_post_style_recalc_layout_passes(&mut self, data: &Reflow, rw_data: &mut LayoutThreadData, layout_context: &mut SharedLayoutContext)

fn perform_post_main_layout_passes(&mut self, data: &Reflow, rw_data: &mut LayoutThreadData, layout_context: &mut SharedLayoutContext)

unsafe fn dirty_all_nodes<N: LayoutNode>(node: N)

fn reflow_all_nodes(flow: &mut Flow)

unsafe fn handle_reap_style_and_layout_data(&self, data: OpaqueStyleAndLayoutData)

Handles a message to destroy layout data. Layout data must be destroyed on this thread because the struct type is transmuted to a different type on the script side.

fn profiler_metadata(&self) -> Option<TimerMetadata>

Returns profiling information which is passed to the time profiler.