Struct layout::layout_thread::LayoutThreadData

pub struct LayoutThreadData {
    pub constellation_chan: IpcSender<ConstellationMsg>,
    pub display_list: Option<Arc<DisplayList>>,
    pub stylist: Arc<Stylist>,
    pub content_box_response: Rect<Au>,
    pub content_boxes_response: Vec<Rect<Au>>,
    pub client_rect_response: Rect<i32>,
    pub layer_id_response: Option<LayerId>,
    pub hit_test_response: (Option<DisplayItemMetadata>, bool),
    pub overflow_response: NodeOverflowResponse,
    pub scroll_area_response: Rect<i32>,
    pub resolved_style_response: Option<String>,
    pub offset_parent_response: OffsetParentResponse,
    pub margin_style_response: MarginStyleResponse,
}

Mutable data belonging to the LayoutThread.

This needs to be protected by a mutex so we can do fast RPCs.

Fields

constellation_chan

The channel on which messages can be sent to the constellation.

display_list

The root stacking context.

stylist

Performs CSS selector matching and style resolution.

content_box_response

A queued response for the union of the content boxes of a node.

content_boxes_response

A queued response for the content boxes of a node.

client_rect_response

A queued response for the client {top, left, width, height} of a node in pixels.

layer_id_response

hit_test_response

A queued response for the node at a given point

overflow_response

A pair of overflow property in x and y

scroll_area_response

A queued response for the scroll {top, left, width, height} of a node in pixels.

resolved_style_response

A queued response for the resolved style property of an element.

offset_parent_response

A queued response for the offset parent/rect of a node.

margin_style_response

A queued response for the offset parent/rect of a node.