Struct layout::layout_thread::RwData

struct RwData<'a, 'b> {
    rw_data: &'b Arc<Mutex<LayoutThreadData>>,
    possibly_locked_rw_data: &'a mut Option<MutexGuard<'b, LayoutThreadData>>,
}

Fields

rw_data

possibly_locked_rw_data

Methods

impl<'a, 'b> RwData<'a, 'b>

fn lock(&mut self) -> RWGuard<'b>

If no reflow has happened yet, this will just return the lock in possibly_locked_rw_data. Otherwise, it will acquire the rw_data lock.

If you do not wish RPCs to remain blocked, just drop the RWGuard returned from this function. If you do wish for them to remain blocked, use block.

fn block(&mut self, rw_data: RWGuard<'b>)

If no reflow has ever been triggered, this will keep the lock, locked (and saved in possibly_locked_rw_data). If it has been, the lock will be unlocked.