Module layout::model

Borders, padding, and margins.

Structs

AdjoiningMargins
A collapsible margin. See CSS 2.1  8.3.1.

IntrinsicISizes
Intrinsic inline-sizes, which consist of minimum and preferred.

IntrinsicISizesContribution
The temporary result of the computation of intrinsic inline-sizes.

MarginCollapseInfo

SizeConstraint
A min-size and max-size constraint. The constructor has a optional border parameter, and when it is present the constraint will be subtracted. This is used to adjust the constraint for box-sizing: border-box, and when you do so make sure the size you want to clamp is intended to be used for content box.

Enums

CollapsibleMargins
Represents the block-start and block-end margins of a flow with collapsible margins. See CSS 2.1  8.3.1.

FinalMarginState

MarginCollapseState

MaybeAuto
Useful helper data type when computing values for blocks and positioned elements.

Functions

padding_from_style

specified_margin_from_style
Returns the explicitly-specified margin lengths from the given style. Percentage and auto margins are returned as zero.

style_length
Receive an optional container size and return used value for width or height.