Struct layout::model::AdjoiningMargins

pub struct AdjoiningMargins {
    pub most_positive: Au,
    pub most_negative: Au,
}

A collapsible margin. See CSS 2.1  8.3.1.

Fields

most_positive

The value of the greatest positive margin.

most_negative

The actual value (not the absolute value) of the negative margin with the largest absolute value. Since this is not the absolute value, this is always zero or negative.

Methods

impl AdjoiningMargins

fn new() -> AdjoiningMargins

fn from_margin(margin_value: Au) -> AdjoiningMargins

fn union(&mut self, other: AdjoiningMargins)

fn collapse(&self) -> Au