Struct layout::model::MarginCollapseInfo

pub struct MarginCollapseInfo {
    pub state: MarginCollapseState,
    pub block_start_margin: AdjoiningMargins,
    pub margin_in: AdjoiningMargins,
}

Fields

state

block_start_margin

margin_in

Methods

impl MarginCollapseInfo

fn new() -> MarginCollapseInfo

TODO(#2012, pcwalton): Remove this method once fragment is not an Option.

fn initialize_block_start_margin(&mut self, fragment: &Fragment, can_collapse_block_start_margin_with_kids: bool)

fn finish_and_compute_collapsible_margins(self, fragment: &Fragment, can_collapse_block_end_margin_with_kids: bool) -> (CollapsibleMargins, Au)

fn current_float_ceiling(&mut self) -> Au

fn advance_block_start_margin(&mut self, child_collapsible_margins: &CollapsibleMargins) -> Au

Adds the child's potentially collapsible block-start margin to the current margin state and advances the Y offset by the appropriate amount to handle that margin. Returns the amount that should be added to the Y offset during block layout.

fn advance_block_end_margin(&mut self, child_collapsible_margins: &CollapsibleMargins) -> Au

Adds the child's potentially collapsible block-end margin to the current margin state and advances the Y offset by the appropriate amount to handle that margin. Returns the amount that should be added to the Y offset during block layout.