Struct layout::model::MinMaxConstraint

pub struct MinMaxConstraint {
    min: Au,
    max: Option<Au>,
}

A min or max constraint

A max of None is equivalent to no limmit for the size in the given dimension. The min is >= 0, as negative values are illegal and by default min is 0.

Fields
min: Au max: Option<Au>
Methods

impl MinMaxConstraint

fn new(content_size: Option<Au>, min: LengthOrPercentage, max: LengthOrPercentageOrNone) -> MinMaxConstraint

Create a MinMaxConstraint for a dimension given the min, max, and content box size for an axis

fn clamp(&self, other: Au) -> Au

Clamp the given size by the given min and max constraints.