layout

Module positioned

Structs

AbsoluteAxisSolver 🔒

AbsolutelyPositionedBox 🔒

HoistedAbsolutelyPositionedBox 🔒

LayoutRootLayoutInputs 🔒

These are the recorded layout inputs that were used when laying out an absolutely-positioned element. They can be re-used when the absolutely-positioned element is a viable layout root (no escaping fixed position elements, currently). The information here is enough to re-run layout for an absolute.

PositioningContext 🔒

PositioningContextLength 🔒

A data structure which stores the size of a positioning context.

RectAxis 🔒

Functions

relative_adjustement 🔒

https://drafts.csswg.org/css2/visuren.html#relative-positioning