layout

Module query

Utilities for querying the layout, as needed by layout.

Structs

OffsetParentFragments 🔒

RenderedTextCollectionState 🔒

Enums

InnerOrOuterTextItem 🔒

Functions

containing_block_for_node 🔒

find_character_offset_in_fragment_descendants

get_the_text_steps

https://html.spec.whatwg.org/multipage/#get-the-text-steps

is_containing_block_for_position 🔒

offset_parent_fragments 🔒

https://www.w3.org/TR/2016/WD-cssom-view-1-20160317/#dom-htmlelement-offsetparent

process_box_area_request 🔒

process_box_areas_request 🔒

process_client_rect_request

process_containing_block_descendant_query

process_containing_block_query

process_current_css_zoom_query

Process a query for the current CSS zoom of an element. https://drafts.csswg.org/cssom-view/#dom-element-currentcsszoom

process_effective_overflow_query 🔒

process_node_scroll_area_request

https://drafts.csswg.org/cssom-view/#scrolling-area

process_offset_parent_query

process_padding_request 🔒

process_resolved_font_style_query

process_resolved_style_request

Return the resolved value of property for a given (pseudo)element. https://drafts.csswg.org/cssom/#resolved-value

process_resolved_style_request_for_unstyled_node

process_scroll_container_query 🔒

An implementation of scrollParent that can also be used to for scrollIntoView: https://drafts.csswg.org/cssom-view/#dom-htmlelement-scrollparent.

rendered_text_collection_steps 🔒

https://html.spec.whatwg.org/multipage/#rendered-text-collection-steps

resolve_grid_template 🔒

resolved_size_should_be_used_value 🔒

root_transform_for_fragments 🔒

Calculate the cumulative transform from the root scroll node for fragments.

shorthand_to_css_string 🔒

should_honor_min_size_auto 🔒

style_and_flags_for_node 🔒

transform_au_rectangle 🔒

transform_f32_rectangle 🔒