layout::query

Function process_current_css_zoom_query

pub fn process_current_css_zoom_query(node: ServoLayoutNode<'_>) -> f32

Process a query for the current CSS zoom of an element. https://drafts.csswg.org/cssom-view/#dom-element-currentcsszoom

Returns the effective zoom of the element, which is the product of all zoom values from the element up to the root. Returns 1.0 if the element is not being rendered (has no associated box).