Module layout::sequential

Implements sequential traversals over the DOM and flow trees.

Functions

build_display_list_for_subtree

guess_float_placement
Guesses how much inline size will be taken up by floats on the left and right sides of the given flow. This is needed to speculatively calculate the inline sizes of block formatting contexts. The speculation typically succeeds, but if it doesn't we have to lay it out again.

iterate_through_flow_tree_fragment_border_boxes

resolve_generated_content

store_overflow

traverse_dom

traverse_flow_tree_preorder