layout::sizing

Enum Size

pub(crate) enum Size<T> {
    Initial,
    MinContent,
    MaxContent,
    FitContent,
    FitContentFunction(T),
    Stretch,
    Numeric(T),
}

The possible values accepted by the sizing properties. https://drafts.csswg.org/css-sizing/#sizing-properties

Variants

Initial

Represents an auto value for the preferred and minimum size properties, or none for the maximum size properties. https://drafts.csswg.org/css-sizing/#valdef-width-auto https://drafts.csswg.org/css-sizing/#valdef-max-width-none

MinContent

https://drafts.csswg.org/css-sizing/#valdef-width-min-content

MaxContent

https://drafts.csswg.org/css-sizing/#valdef-width-max-content

FitContent

https://drafts.csswg.org/css-sizing-4/#valdef-width-fit-content

FitContentFunction(T)

https://drafts.csswg.org/css-sizing-3/#funcdef-width-fit-content

Stretch

https://drafts.csswg.org/css-sizing-4/#valdef-width-stretch

Numeric(T)

Represents a numeric <length-percentage>, but resolved as a T. https://drafts.csswg.org/css-sizing/#valdef-width-length-percentage-0

Implementations

impl<T> Size<T>

pub(crate) fn is_initial(&self) -> bool

impl<T: Clone> Size<T>

pub(crate) fn to_numeric(&self) -> Option<T>

pub(crate) fn map<U>(&self, f: impl FnOnce(T) -> U) -> Size<U>

impl Size<LengthPercentage>

pub(crate) fn to_percentage(&self) -> Option<Percentage>

pub(crate) fn resolve_percentages_for_preferred( &self, basis: Option<Au>, ) -> Size<Au>

Resolves percentages in a preferred size, against the provided basis. If the basis is missing, percentages are considered cyclic. https://www.w3.org/TR/css-sizing-3/#preferred-size-properties https://www.w3.org/TR/css-sizing-3/#cyclic-percentage-size

pub(crate) fn resolve_percentages_for_max(&self, basis: Option<Au>) -> Size<Au>

Resolves percentages in a maximum size, against the provided basis. If the basis is missing, percentages are considered cyclic. https://www.w3.org/TR/css-sizing-3/#preferred-size-properties https://www.w3.org/TR/css-sizing-3/#cyclic-percentage-size

impl Size<Au>

pub(crate) fn resolve_for_preferred<F: FnOnce() -> ContentSizes>( &self, automatic_size: Size<Au>, stretch_size: Option<Au>, content_size: &LazyCell<ContentSizes, F>, ) -> Au

Resolves a preferred size into a numerical value. https://www.w3.org/TR/css-sizing-3/#preferred-size-properties

pub(crate) fn resolve_for_min<F: FnOnce() -> ContentSizes>( &self, get_automatic_minimum_size: impl FnOnce() -> Au, stretch_size: Option<Au>, content_size: &LazyCell<ContentSizes, F>, is_table: bool, ) -> Au

Resolves a minimum size into a numerical value. https://www.w3.org/TR/css-sizing-3/#min-size-properties

pub(crate) fn resolve_for_max<F: FnOnce() -> ContentSizes>( &self, stretch_size: Option<Au>, content_size: &LazyCell<ContentSizes, F>, ) -> Option<Au>

Resolves a maximum size into a numerical value. https://www.w3.org/TR/css-sizing-3/#max-size-properties

pub(crate) fn maybe_resolve_extrinsic( &self, stretch_size: Option<Au>, ) -> Option<Au>

Tries to resolve an extrinsic size into a numerical value. Extrinsic sizes are those based on the context of an element, without regard for its contents. https://drafts.csswg.org/css-sizing-3/#extrinsic

Returns None if either:

The size is intrinsic.

The size is the initial one. TODO: should we allow it to behave as stretch instead of assuming it’s intrinsic?

The provided stretch_size is None but we need its value.