Struct layout::ServoThreadSafeLayoutNode

pub struct ServoThreadSafeLayoutNode<'ln> {
    node: ServoLayoutNode<'ln>,
    pseudo: PseudoElementType<Option<T>>,
}

Fields
node: ServoLayoutNode<'ln>

The wrapped node.

pseudo: PseudoElementType<Option<T>>

The pseudo-element type, with (optionally), an specified display value to override the stylesheet.

Methods

impl<'ln> ServoThreadSafeLayoutNode<'ln>

unsafe fn new_with_this_lifetime(&self, node: &LayoutJS<Node>) -> ServoThreadSafeLayoutNode<'ln>

Creates a new layout node with the same lifetime as this layout node.

fn new<'a>(node: &ServoLayoutNode<'a>) -> ServoThreadSafeLayoutNode<'a>

Creates a new ServoThreadSafeLayoutNode from the given ServoLayoutNode.

unsafe fn get_jsmanaged(&self) -> &LayoutJS<Node>

Returns the interior of this node as a LayoutJS. This is highly unsafe for layout to call and as such is marked unsafe.