Struct layout::table::ChildInlineSizeInfo

pub struct ChildInlineSizeInfo<'a> {
    pub spacing: T,
    pub column_computed_inline_sizes: &'a [ColumnComputedInlineSize],
}

Inline-size information that we need to push down to table children.

Fields

spacing

The spacing of the table.

column_computed_inline_sizes

The computed inline sizes for each column.

Methods

impl<'a> ChildInlineSizeInfo<'a>

fn propagate_to_child(&self, kid: &mut Flow, child_index: usize, content_inline_size: Au, writing_mode: WritingMode, inline_start_margin_edge: &mut Au)

Propagates information computed during inline size assignment to a child of a table, and lays out that child in the inline direction.