layout::table

Struct Table

pub struct Table {
    style: Arc<ComputedValues>,
    grid_style: Arc<ComputedValues>,
    grid_base_fragment_info: BaseFragmentInfo,
    pub captions: Vec<ArcRefCell<TableCaption>>,
    pub column_groups: Vec<ArcRefCell<TableTrackGroup>>,
    pub columns: Vec<ArcRefCell<TableTrack>>,
    pub row_groups: Vec<ArcRefCell<TableTrackGroup>>,
    pub rows: Vec<ArcRefCell<TableTrack>>,
    pub slots: Vec<Vec<TableSlot>>,
    pub size: TableSize,
    anonymous: bool,
    percentage_columns_allowed_for_inline_content_sizes: bool,
}

Fields
style: Arc<ComputedValues>

The style of this table. These are the properties that apply to the “wrapper” ie the element that contains both the grid and the captions. Not all properties are actually used on the wrapper though, such as background and borders, which apply to the grid.

grid_style: Arc<ComputedValues>

The style of this table’s grid. This is an anonymous style based on the table’s style, but eliminating all the properties handled by the “wrapper.”

grid_base_fragment_info: BaseFragmentInfo

The BaseFragmentInfo for this table’s grid. This is necessary so that when the grid has a background image, it can be associated with the table’s node.

captions: Vec<ArcRefCell<TableCaption>>

The captions for this table.

column_groups: Vec<ArcRefCell<TableTrackGroup>>

The column groups for this table.

columns: Vec<ArcRefCell<TableTrack>>

The columns of this table defined by <colgroup> | display: table-column-group and <col> | display: table-column elements as well as display: table-column.

row_groups: Vec<ArcRefCell<TableTrackGroup>>

The rows groups for this table defined by <tbody>, <thead>, and <tfoot>.

rows: Vec<ArcRefCell<TableTrack>>

The rows of this table defined by <tr> or display: table-row elements.

slots: Vec<Vec<TableSlot>>

The content of the slots of this table.

size: TableSize

The size of this table.

anonymous: bool

Whether or not this Table is anonymous.

percentage_columns_allowed_for_inline_content_sizes: bool

Whether percentage columns are taken into account during inline content sizes calculation.

Implementations

impl Table

pub(crate) fn construct( context: &LayoutContext<'_>, info: &NodeAndStyleInfo<'_>, grid_style: Arc<ComputedValues>, contents: NonReplacedContents, propagated_data: PropagatedBoxTreeData, ) -> Self

pub(crate) fn construct_anonymous<'dom>( context: &LayoutContext<'_>, parent: &mut impl TraversalHandler<'dom>, parent_info: &NodeAndStyleInfo<'dom>, contents: Vec<AnonymousTableContent<'dom>>, propagated_data: PropagatedBoxTreeData, ) -> (NodeAndStyleInfo<'dom>, IndependentFormattingContext)

fn push_new_slot_to_last_row(&mut self, slot: TableSlot)

Push a new slot into the last row of this table.

pub(super) fn resolve_slot_at( &self, coords: Point2D<usize, UnknownUnit>, ) -> Vec<ResolvedSlotAndLocation<'_>>

Find ResolvedSlotAndLocation of all the slots that cover the slot at the given coordinates. This recursively resolves all of the TableSlotCells that cover the target and returns a ResolvedSlotAndLocation for each of them. If there is no slot at the given coordinates or that slot is an empty space, an empty vector is returned.

impl Table

fn border_spacing(&self) -> LogicalVec2<Au>

fn total_border_spacing(&self) -> LogicalVec2<Au>

fn get_column_measure_for_column_at_index( &self, writing_mode: WritingMode, column_index: usize, is_in_fixed_mode: bool, ) -> CellOrTrackMeasure

fn get_row_measure_for_row_at_index( &self, writing_mode: WritingMode, row_index: usize, ) -> CellOrTrackMeasure

pub(crate) fn layout( &self, layout_context: &LayoutContext<'_>, positioning_context: &mut PositioningContext, containing_block_for_children: &ContainingBlock<'_>, containing_block_for_table: &ContainingBlock<'_>, ) -> IndependentFormattingContextLayoutResult

pub(crate) fn attached_to_tree(&self, layout_box: WeakLayoutBox)

impl Table

pub(crate) fn layout_style<'a>( &'a self, layout: Option<&'a TableLayout<'a>>, ) -> LayoutStyle<'a>

pub(crate) fn layout_style_for_grid(&self) -> LayoutStyle<'_>

impl Table

pub(crate) fn new( style: Arc<ComputedValues>, grid_style: Arc<ComputedValues>, base_fragment_info: BaseFragmentInfo, percentage_columns_allowed_for_inline_content_sizes: bool, ) -> Self

fn get_slot(&self, coords: Point2D<usize, UnknownUnit>) -> Option<&TableSlot>

Return the slot at the given coordinates, if it exists in the table, otherwise return None.

fn resolve_first_cell_coords( &self, coords: Point2D<usize, UnknownUnit>, ) -> Option<Point2D<usize, UnknownUnit>>

fn resolve_first_cell( &self, coords: Point2D<usize, UnknownUnit>, ) -> Option<AtomicRef<'_, TableSlotCell>>

pub(crate) fn repair_style( &mut self, context: &SharedStyleContext<'_>, new_style: &Arc<ComputedValues>, )

pub(crate) fn subtree_size(&self) -> usize