Struct layout::table::TableFlow

#[repr(C)]
pub struct TableFlow {
    pub block_flow: BlockFlow,
    pub column_intrinsic_inline_sizes: Vec<ColumnIntrinsicInlineSize>,
    pub column_computed_inline_sizes: Vec<ColumnComputedInlineSize>,
    pub collapsed_inline_direction_border_widths_for_table: Vec<Au>,
    pub collapsed_block_direction_border_widths_for_table: Vec<Au>,
    pub table_layout: TableLayout,
}

A table flow corresponded to the table’s internal table fragment under a table wrapper flow. The properties position, float, and margin-* are used on the table wrapper fragment, not table fragment per CSS 2.1  10.5.

Fields
block_flow: BlockFlowcolumn_intrinsic_inline_sizes: Vec<ColumnIntrinsicInlineSize>

Information about the intrinsic inline-sizes of each column, computed bottom-up during intrinsic inline-size bubbling.

column_computed_inline_sizes: Vec<ColumnComputedInlineSize>

Information about the actual inline sizes of each column, computed top-down during actual inline-size bubbling.

collapsed_inline_direction_border_widths_for_table: Vec<Au>

The final width of the borders in the inline direction for each cell, computed by the entire table and pushed down into each row during inline size computation.

collapsed_block_direction_border_widths_for_table: Vec<Au>

The final width of the borders in the block direction for each cell, computed by the entire table and pushed down into each row during inline size computation.

table_layout: TableLayout

Table-layout property

Implementations

impl TableFlow

pub fn from_fragment(fragment: Fragment) -> TableFlow

fn update_automatic_column_inline_sizes(
parent_inline_sizes: &mut Vec<ColumnIntrinsicInlineSize>,
child_cell_inline_sizes: &[CellIntrinsicInlineSize],
surrounding_size: Au
) -> IntrinsicISizes

Update the corresponding value of self_inline_sizes if a value of kid_inline_sizes has a larger value than one of self_inline_sizes. Returns the minimum and preferred inline sizes.

fn update_column_inline_sizes_for_row(
row: &TableRowFlow,
column_inline_sizes: &mut Vec<ColumnIntrinsicInlineSize>,
computation: &mut IntrinsicISizesContribution,
first_row: bool,
table_layout: TableLayout,
surrounding_inline_size: Au
)

Updates the minimum and preferred inline-size calculation for a single row. This is factored out into a separate function because we process children of rowgroups too.

pub fn spacing(&self) -> T

Returns the effective spacing per cell, taking the value of border-collapse into account.

pub fn total_horizontal_spacing(&self) -> Au

fn column_styles(&self) -> Vec<ColumnStyle<'_>>