Function layout::table_row::perform_inline_direction_border_collapse_for_row

fn perform_inline_direction_border_collapse_for_row(
    row_style: &ComputedValues, 
    children_count: usize, 
    child_index: usize, 
    child_table_cell: &mut TableCellFlow, 
    iterator: &mut Peekable<Enumerate<MutFlowListIterator>>, 
    preliminary_collapsed_borders: &mut CollapsedBordersForRow
)

Performs border-collapse in the inline direction for all the cells' inside borders in the inline-direction cells and propagates the outside borders (the far left and right) up to the table row. This is done eagerly here so that at least the inline inside border collapse computations can be parallelized across all the rows of the table.