Struct layout::table_row::CellIntrinsicInlineSize

pub struct CellIntrinsicInlineSize {
    pub column_size: ColumnIntrinsicInlineSize,
    pub column_span: u32,
}

Information about the column inline size and span for each cell.

Fields

column_size

Inline sizes that this cell contributes to the column.

column_span

The column span of this cell.