Struct layout::table_row::CollapsedBordersForRow

pub struct CollapsedBordersForRow {
    pub inline: Vec<CollapsedBorder>,
    pub block_start: Vec<CollapsedBorder>,
    pub block_end: Vec<CollapsedBorder>,
}

Fields
inline: Vec<CollapsedBorder>

The size of this vector should be equal to the number of cells plus one.

block_start: Vec<CollapsedBorder>

The size of this vector should be equal to the number of cells.

block_end: Vec<CollapsedBorder>

The size of this vector should be equal to the number of cells.

Methods

impl CollapsedBordersForRow

fn new() -> CollapsedBordersForRow

fn reset(&mut self, first_inline_border: CollapsedBorder)