Struct layout::table_wrapper::AutoLayoutCandidateGuess

struct AutoLayoutCandidateGuess {
    minimum_guess: Au,
    minimum_percentage_guess: Au,
    minimum_specified_guess: Au,
    preferred_guess: Au,
}

The layout "guesses" defined in INTRINSIC  4.3.

Fields

minimum_guess

The column inline-size assignment where each column is assigned its intrinsic minimum inline-size.

minimum_percentage_guess

The column inline-size assignment where: * A column with an intrinsic percentage inline-size greater than 0% is assigned the larger of: - Its intrinsic percentage inline-size times the assignable inline-size; - Its intrinsic minimum inline-size; * Other columns receive their intrinsic minimum inline-size.

minimum_specified_guess

The column inline-size assignment where: * Each column with an intrinsic percentage inline-size greater than 0% is assigned the larger of: - Its intrinsic percentage inline-size times the assignable inline-size; - Its intrinsic minimum inline-size; * Any other column that is constrained is assigned its intrinsic preferred inline-size; * Other columns are assigned their intrinsic minimum inline-size.

preferred_guess

The column inline-size assignment where: * Each column with an intrinsic percentage inline-size greater than 0% is assigned the larger of: - Its intrinsic percentage inline-size times the assignable inline-size; - Its intrinsic minimum inline-size; * Other columns are assigned their intrinsic preferred inline-size.

Methods

impl AutoLayoutCandidateGuess

fn new() -> AutoLayoutCandidateGuess

Creates a guess with all elements initialized to zero.

fn from_column_intrinsic_inline_size(column_intrinsic_inline_size: &ColumnIntrinsicInlineSize, assignable_inline_size: Au) -> AutoLayoutCandidateGuess

Fills in the inline-size guesses for this column per INTRINSIC  4.3.

fn calculate(&self, selection: SelectedAutoLayoutCandidateGuess) -> Au

Calculates the inline-size, interpolating appropriately based on the value of selection.

This does not distribute excess inline-size. That must be done later if necessary.