Struct layout::table_wrapper::TableWrapperFlow

#[repr(C)]
pub struct TableWrapperFlow {
    pub block_flow: BlockFlow,
    pub column_intrinsic_inline_sizes: Vec<ColumnIntrinsicInlineSize>,
    pub table_layout: TableLayout,
}

A table wrapper flow based on a block formatting context.

Fields
block_flow: BlockFlowcolumn_intrinsic_inline_sizes: Vec<ColumnIntrinsicInlineSize>

Intrinsic column inline sizes according to INTRINSIC  4.1

table_layout: TableLayout

Table-layout property

Implementations

impl TableWrapperFlow

pub fn from_fragment(fragment: Fragment) -> TableWrapperFlow

pub fn from_fragment_and_float_kind(
fragment: Fragment,
float_kind: Option<FloatKind>
) -> TableWrapperFlow

fn border_padding_and_spacing(&mut self) -> (Au, Au)

fn compute_border_and_padding_of_table(&mut self)

fn calculate_table_column_sizes_for_automatic_layout(
&mut self,
intermediate_column_inline_sizes: &mut [IntermediateColumnInlineSize]
)

Calculates table column sizes for automatic layout per INTRINSIC  4.3.

fn available_inline_size(&mut self) -> Au

fn set_inline_size(&mut self, total_used_inline_size: Au)

fn compute_used_inline_size(
&mut self,
shared_context: &SharedStyleContext<'_>,
parent_flow_inline_size: Au,
intermediate_column_inline_sizes: &[IntermediateColumnInlineSize]
)