layout::taffy::stylo_taffy

Module stylo

Private module of type aliases so we can refer to stylo types with nicer names

Structs

AlignFlags 🔒

Constants shared by multiple CSS Box Alignment properties

CalcLengthPercentage 🔒

The representation of a calc() function with mixed lengths and percentages.

ContentDistribution 🔒

Shared value for the align-content and justify-content properties.

Display 🔒

GridAutoFlow 🔒

Controls how the auto-placement algorithm works specifying exactly how auto-placed items get flowed into the grid: [ row | column ] || dense https://drafts.csswg.org/css-grid-2/#grid-auto-flow-property

LengthPercentage 🔒

A <length-percentage> value. This can be either a <length>, a <percentage>, or a combination of both via calc().

NonNegative 🔒

A wrapper of Non-negative values.

Percentage 🔒

A computed percentage.

Enums

BoxSizing 🔒

Direction 🔒

DisplayInside 🔒

DisplayOutside 🔒

Defines an element’s display type, which consists of the two basic qualities of how an element generates boxes https://drafts.csswg.org/css-display/#propdef-display

GenericGridTemplateComponent 🔒

Variants for <grid-template-rows> | <grid-template-columns>

GenericInset 🔒

Specified type for inset properties, which allows the use of the anchor() function. Note(dshin): LengthPercentageOrAuto is not used here because having LengthPercentageOrAuto and AnchorFunction in the enum pays the price of the discriminator for LengthPercentage | Auto as well as LengthPercentageOrAuto | AnchorFunction. This increases the size of the style struct, which would not be great. On the other hand, we trade for code duplication, so… :(

GenericLengthPercentageOrNormal 🔒

A generic <length-percentage> | normal` value.

GenericMargin 🔒

Specified type for margin properties, which allows the use of the anchor-size() function.

GenericMaxSize 🔒

A generic value for the max-width or max-height property.

GenericSize 🔒

A generic value for the width, height, min-width, or min-height property.

Overflow 🔒

The value for the overflow-x / overflow-y properties.

Position 🔒

PreferredRatio 🔒

Ratio or None.

RepeatCount 🔒

The initial argument of the repeat function.

TrackBreadth 🔒

A track breadth for explicit grid track sizing. It’s generic solely to avoid re-implementing it for the computed type.

TrackListValue 🔒

Track list values. Can be or

TrackSize 🔒

A <track-size> type for explicit grid track sizing. Like <track-breadth>, this is generic only to avoid code bloat. It only takes <length-percentage>

UnpackedLengthPercentage 🔒

An unpacked <length-percentage> that borrows the calc() variant.

Type Aliases

AspectRatio 🔒

A computed value for the aspect-ratio property.

Gap 🔒

GridLine 🔒

The computed value of a <grid-line>.

InsetVal 🔒

MarginVal 🔒

MaxSize 🔒

Size 🔒