layout::taffy::stylo_taffy::stylo

Enum DisplayOutside

#[repr(u8)]
pub(crate) enum DisplayOutside {
    None = 0,
    Inline = 1,
    Block = 2,
    TableCaption = 3,
    InternalTable = 4,
}

Defines an element’s display type, which consists of the two basic qualities of how an element generates boxes https://drafts.csswg.org/css-display/#propdef-display

Variants

None = 0

Inline = 1

Block = 2

TableCaption = 3

InternalTable = 4