layout::taffy::stylo_taffy::stylo

Enum GenericInset

#[repr(C)]
pub(crate) enum GenericInset<P, LP> {
    LengthPercentage(LP),
    Auto,
    AnchorFunction(Box<GenericAnchorFunction<P, GenericInset<P, LP>>>),
    AnchorSizeFunction(Box<GenericAnchorSizeFunction<GenericInset<P, LP>>>),
    AnchorContainingCalcFunction(LP),
}

Specified type for inset properties, which allows the use of the anchor() function. Note(dshin): LengthPercentageOrAuto is not used here because having LengthPercentageOrAuto and AnchorFunction in the enum pays the price of the discriminator for LengthPercentage | Auto as well as LengthPercentageOrAuto | AnchorFunction. This increases the size of the style struct, which would not be great. On the other hand, we trade for code duplication, so… :(

Variants

LengthPercentage(LP)

A <length-percentage> value.

Auto

An auto value.

AnchorFunction(Box<GenericAnchorFunction<P, GenericInset<P, LP>>>)

Inset defined by the anchor element.

https://drafts.csswg.org/css-anchor-position-1/#anchor-pos

AnchorSizeFunction(Box<GenericAnchorSizeFunction<GenericInset<P, LP>>>)

Inset defined by the size of the anchor element.

https://drafts.csswg.org/css-anchor-position-1/#anchor-pos

AnchorContainingCalcFunction(LP)

A <length-percentage> value, guaranteed to contain calc(), which then is guaranteed to contain anchor() or anchor-size().

Implementations

impl<P, LP> GenericInset<P, LP>

pub fn auto() -> GenericInset<P, LP>

auto value.

pub fn is_auto(&self) -> bool

Return true if it is ‘auto’.

impl GenericInset<Percentage, LengthPercentage>

pub fn parse_quirky<'i, 't>( context: &ParserContext<'_>, input: &mut Parser<'i, 't>, allow_quirks: AllowQuirks, ) -> Result<GenericInset<Percentage, LengthPercentage>, ParseError<'i, StyleParseErrorKind<'i>>>

Parses an inset type, allowing the unitless length quirk. https://quirks.spec.whatwg.org/#the-unitless-length-quirk