layout::taffy::stylo_taffy::stylo

Enum Overflow

#[repr(u8)]
pub(crate) enum Overflow {
    Visible = 0,
    Hidden = 1,
    Scroll = 2,
    Auto = 3,
    Clip = 4,
}

The value for the overflow-x / overflow-y properties.

Variants

Visible = 0

Hidden = 1

Scroll = 2

Auto = 3

Clip = 4

Implementations

impl Overflow

pub fn is_scrollable(&self) -> bool

Return true if the value will create a scrollable box.

pub fn to_scrollable(&self) -> Overflow

Convert the value to a scrollable value if it’s not already scrollable. This maps visible to auto and clip to hidden.