layout::taffy::stylo_taffy::stylo

Struct LengthPercentage

#[repr(C)]
pub(crate) struct LengthPercentage(NumericUnion<LengthPercentageTag, f32, CalcLengthPercentage>);

A <length-percentage> value. This can be either a <length>, a <percentage>, or a combination of both via calc().

https://drafts.csswg.org/css-values-4/#typedef-length-percentage

cbindgen:derive-eq=false cbindgen:derive-neq=false

Tuple Fields
0: NumericUnion<LengthPercentageTag, f32, CalcLengthPercentage>
Implementations

impl LengthPercentage

pub fn one() -> LengthPercentage

1px length value for SVG defaults

pub fn zero_percent() -> LengthPercentage

0%

pub fn hundred_percent() -> LengthPercentage

100%

pub fn new_length(length: CSSPixelLength) -> LengthPercentage

Constructs a length value.

pub fn new_percent(percentage: Percentage) -> LengthPercentage

Constructs a percentage value.

pub fn hundred_percent_minus( v: LengthPercentage, clamping_mode: AllowedNumericType, ) -> LengthPercentage

Given a LengthPercentage value v, construct the value representing calc(100% - v).

pub fn hundred_percent_minus_list( list: &[&LengthPercentage], clamping_mode: AllowedNumericType, ) -> LengthPercentage

Given a list of LengthPercentage values, construct the value representing calc(100% - the sum of the list).

pub fn new_calc( node: GenericCalcNode<ComputedLeaf>, clamping_mode: AllowedNumericType, ) -> LengthPercentage

Constructs a calc() value.

pub fn unpack<'a>(&'a self) -> Unpacked<'a>

Unpack the tagged pointer representation of a length-percentage into an enum representation with separate tag and value.

pub fn resolve(&self, basis: CSSPixelLength) -> CSSPixelLength

Resolves the percentage.

pub fn percentage_relative_to(&self, basis: CSSPixelLength) -> CSSPixelLength

Resolves the percentage. Just an alias of resolve().

pub fn has_percentage(&self) -> bool

Return whether there’s any percentage in this value.

pub fn to_length(&self) -> Option<CSSPixelLength>

Converts to a <length> if possible.

pub fn to_percentage(&self) -> Option<Percentage>

Converts to a <percentage> if possible.

pub fn to_percentage_of(&self, basis: CSSPixelLength) -> Option<Percentage>

Converts to a <percentage> with given basis. Returns None if the basis is 0.

pub fn to_used_value(&self, containing_length: Au) -> Au

Returns the used value.

pub fn to_pixel_length(&self, containing_length: Au) -> CSSPixelLength

Returns the used value as CSSPixelLength.

pub fn maybe_to_used_value(&self, container_len: Option<Au>) -> Option<Au>

Convert the computed value into used value.

pub fn maybe_percentage_relative_to( &self, container_len: Option<CSSPixelLength>, ) -> Option<CSSPixelLength>

If there are special rules for computing percentages in a value (e.g. the height property), they apply whenever a calc() expression contains percentages.

impl LengthPercentage

pub fn normal() -> LengthPercentage

Return the normal computed value, which is just zero.