layout::taffy::stylo_taffy::stylo

Struct NonNegative

#[repr(transparent)]
pub(crate) struct NonNegative<T>(pub T);

A wrapper of Non-negative values.

Tuple Fields
0: T
Implementations

impl NonNegative<CSSPixelLength>

pub fn new(px: f32) -> NonNegative<CSSPixelLength>

Create a NonNegativeLength.

pub fn px(&self) -> f32

Return the pixel value of |NonNegativeLength|.

pub fn clamp(self) -> NonNegative<CSSPixelLength>

Ensures it is non negative

impl NonNegative<LengthPercentage>

pub fn to_used_value(&self, containing_length: Au) -> Au

Returns the used value.

pub fn maybe_to_used_value(&self, containing_length: Option<Au>) -> Option<Au>

Convert the computed value into used value.

impl NonNegative<Percentage>

pub fn hundred() -> NonNegative<Percentage>

100%

impl NonNegative<NumberOrPercentage>

pub fn hundred_percent() -> NonNegative<NumberOrPercentage>

Returns the 100% value.

impl NonNegative<Length>

pub fn from_px(px_value: f32) -> NonNegative<Length>

Get an absolute length from a px value.

pub fn parse_quirky<'i, 't>( context: &ParserContext<'_>, input: &mut Parser<'i, 't>, allow_quirks: AllowQuirks, ) -> Result<NonNegative<Length>, ParseError<'i, StyleParseErrorKind<'i>>>

Parses a non-negative length, optionally with quirks.

impl NonNegative<LengthPercentage>

pub fn zero_percent() -> NonNegative<LengthPercentage>

Returns a 0% value.

pub fn parse_quirky<'i, 't>( context: &ParserContext<'_>, input: &mut Parser<'i, 't>, allow_quirks: AllowQuirks, ) -> Result<NonNegative<LengthPercentage>, ParseError<'i, StyleParseErrorKind<'i>>>

Parses a length or a percentage, allowing the unitless length quirk. https://quirks.spec.whatwg.org/#the-unitless-length-quirk

pub fn parse_non_negative_with_anchor_size<'i, 't>( context: &ParserContext<'_>, input: &mut Parser<'i, 't>, allow_quirks: AllowQuirks, ) -> Result<NonNegative<LengthPercentage>, ParseError<'i, StyleParseErrorKind<'i>>>

Parses a length or a percentage, allowing the unitless length quirk, as well as allowing anchor-size().

impl NonNegative<Number>

pub fn new(val: f32) -> NonNegative<Number>

Returns a new non-negative number with the value val.

pub fn get(&self) -> Option<f32>

Returns the numeric value.

impl NonNegative<Percentage>

pub fn compute(&self) -> Option<Percentage>

Convert to ComputedPercentage, for FontFaceRule size-adjust getter. Returns None if the value is a calc expression that cannot be resolved at parse time.

impl NonNegative<NumberOrPercentage>

pub fn hundred_percent() -> NonNegative<NumberOrPercentage>

Returns the 100% value.

pub fn new_number(n: f32) -> NonNegative<NumberOrPercentage>

Return a particular number.