Module layout::text

Text layout.

Structs

RunInfo
Information about a text run that we're about to create. This is used in scan_for_runs.

RunMapping
A mapping from a portion of an unscanned text fragment to the text run we're going to create for it.

ScannedTextRun

TextRunScanner
A stack-allocated object for scanning an inline flow into TextRun-containing TextFragments.

Functions

apply_style_transform_if_necessary
Accounts for text-transform.

bounding_box_for_run_metrics

font_metrics_for_style
Returns the metrics of the font represented by the given FontStyleStruct.

is_compatible
Can a character with script b continue a text run with script a?

is_specific
Returns true if the script is not invalid or inherited.

line_height_from_style
Returns the line block-size needed by the given computed style and font size.

split_first_fragment_at_newline_if_necessary

text
Returns the concatenated text of a list of unscanned text fragments.