Function layout::text::font_metrics_for_style

pub fn font_metrics_for_style(
    font_context: &mut LayoutFontContext, 
    style: ServoArc<FontStyleStruct>
) -> FontMetrics

Returns the metrics of the font represented by the given FontStyleStruct.

#[inline] because often the caller only needs a few fields from the font metrics.

Panics

Panics if no font can be found for the given font style.