Struct layout::text::RunMapping

struct RunMapping {
    byte_range: Range<usize>,
    old_fragment_index: usize,
    text_run_index: usize,
    selected: bool,
}

A mapping from a portion of an unscanned text fragment to the text run we're going to create for it.

Fields

byte_range

The range of byte indices within the text fragment.

old_fragment_index

The index of the unscanned text fragment that this mapping corresponds to.

text_run_index

The index of the text run we're going to create.

selected

Is the text in this fragment selected?

Methods

impl RunMapping

fn new(run_info_list: &[RunInfo], fragment_index: usize) -> RunMapping

Given the current set of text runs, creates a run mapping for the next fragment. run_info_list describes the set of runs we've seen already.

fn flush(self, mappings: &mut Vec<RunMapping>, run_info: &mut RunInfo, text: &str, compression: CompressionMode, text_transform: T, last_whitespace: &mut bool, start_position: &mut usize, end_position: usize)

Flushes this run mapping to the list. run_info describes the text run that we're currently working on. text refers to the text of this fragment.

fn contains_insertion_point(&self, insertion_point: Option<ByteIndex>) -> bool

Is the insertion point for this text run within this mapping?

NOTE: We treat the range as inclusive at both ends, since the insertion point can lie before the first character or after the last character, and should be drawn even if the text is empty.