Module layout::traversal

Traversals over the DOM and flow trees, running the layout computations.

Structs

AssignBSizesAndStoreOverflow
The assign-block-sizes-and-store-overflow traversal, the last (and most expensive) part of layout computation. Determines the final block-sizes for all layout objects, computes positions, and computes overflow regions. In Gecko this corresponds to Reflow and FinishAndStoreOverflow.

AssignISizes
The assign-inline-sizes traversal. In Gecko this corresponds to Reflow.

BubbleISizes
The bubble-inline-sizes traversal, the first part of layout computation. This computes preferred and intrinsic inline-sizes and bubbles them up the tree.

BuildDisplayList

ComputeAbsolutePositions

ConstructFlows
The flow construction traversal, which builds flows for styled nodes.

RecalcStyleForNode
The recalc-style-for-node traversal, which styles each node and must run before layout computation. This computes the styles applied to each node.