Struct layout::traversal::RecalcStyleAndConstructFlows

pub struct RecalcStyleAndConstructFlows {
    shared: SharedLayoutContext,
}

Fields
shared: SharedLayoutContext
Methods

impl RecalcStyleAndConstructFlows

fn shared_layout_context(&self) -> &SharedLayoutContext

impl RecalcStyleAndConstructFlows

fn new(shared: SharedLayoutContext) -> Self

Creates a traversal context, taking ownership of the shared layout context.

fn destroy(self) -> SharedLayoutContext

Consumes this traversal context, returning ownership of the shared layout context to the caller.