Trait layout::traversal::PostorderFlowTraversal

pub trait PostorderFlowTraversal {
    fn process(&self, flow: &mut dyn Flow);

    fn should_process(&self, _flow: &mut dyn Flow) -> bool { ... }

    fn traverse(&self, flow: &mut dyn Flow) { ... }
}

A bottom-up traversal, with a optional in-order pass.

Required methods

fn process(&self, flow: &mut dyn Flow)

The operation to perform. Return true to continue or false to stop.

Provided methods

fn should_process(&self, _flow: &mut dyn Flow) -> bool

Returns false if this node must be processed in-order. If this returns false, we skip the operation for this node, but continue processing the ancestors. This is called after child nodes are visited.

fn traverse(&self, flow: &mut dyn Flow)

Traverses the tree in postorder.