Trait layout::traversal::PreorderFlowTraversal

pub trait PreorderFlowTraversal {
    fn process(&self, flow: &mut dyn Flow);

    fn should_process_subtree(&self, _flow: &mut dyn Flow) -> bool { ... }

    fn should_process(&self, _flow: &mut dyn Flow) -> bool { ... }

    fn traverse(&self, flow: &mut dyn Flow) { ... }

    fn traverse_absolute_flows(&self, flow: &mut dyn Flow) { ... }
}

A top-down traversal.

Required methods

fn process(&self, flow: &mut dyn Flow)

The operation to perform. Return true to continue or false to stop.

Loading content...
Provided methods

fn should_process_subtree(&self, _flow: &mut dyn Flow) -> bool

Returns true if this node should be processed and false if neither this node nor its descendants should be processed.

fn should_process(&self, _flow: &mut dyn Flow) -> bool

Returns true if this node must be processed in-order. If this returns false, we skip the operation for this node, but continue processing the descendants. This is called after parent nodes are visited.

fn traverse(&self, flow: &mut dyn Flow)

Traverses the tree in preorder.

fn traverse_absolute_flows(&self, flow: &mut dyn Flow)

Traverse the Absolute flow tree in preorder.

Traverse all your direct absolute descendants, who will then traverse their direct absolute descendants.

Return true if the traversal is to continue or false to stop.

Loading content...