Module layout::wrapper

A safe wrapper for DOM nodes that prevents layout from mutating the DOM, from letting DOM nodes escape, and from generally doing anything that it isn't supposed to. This is accomplished via a simple whitelist of allowed operations, along with some lifetime magic to prevent nodes from escaping.

As a security wrapper is only as good as its whitelist, be careful when adding operations to this list. The cardinal rules are:

Layout is not allowed to mutate the DOM.

Layout is not allowed to see anything with LayoutJS in the name, because it could hang onto these objects and cause use-after-free.

When implementing wrapper functions, be careful that you do not touch the borrow flags, or you will race and cause spurious task failure. (Note that I do not believe these races are exploitable, but they'll result in brokenness nonetheless.)

Rules of the road for this file:

You must not use .get(); instead, use .unsafe_get().

Do not call any methods on DOM nodes without checking to see whether they use borrow flags.

o Instead of get_attr(), use .get_attr_val_for_layout().

o Instead of html_element_in_html_document(), use html_element_in_html_document_for_layout().

Structs

LayoutElement
A wrapper around elements that ensures layout can only ever access safe properties.

LayoutNode
A wrapper so that layout can access only the methods that it should have access to. Layout must only ever see these and must never see instances of LayoutJS.

LayoutNodeChildrenIterator

LayoutNodeReverseChildrenIterator

LayoutTreeIterator

ThreadSafeLayoutElement
A wrapper around elements that ensures layout can only ever access safe properties and cannot race on elements.

ThreadSafeLayoutNode
A thread-safe version of LayoutNode, used during flow construction. This type of layout node does not allow any parents or siblings of nodes to be accessed, to avoid races.

ThreadSafeLayoutNodeChildrenIterator

Enums

PseudoElementType

Traits

PostorderDomTraversal
A bottom-up traversal, with a optional in-order pass.

PostorderNodeMutTraversal
A bottom-up, parallelizable traversal.

PreorderDomTraversal
A top-down traversal.

TLayoutNode
Allows some convenience methods on generic layout nodes.

Functions

layout_node_from_unsafe_layout_node

layout_node_to_unsafe_layout_node

Type Definitions

UnsafeLayoutNode
Opaque type stored in type-unsafe work queues for parallel layout. Must be transmutable to and from LayoutNode/ThreadSafeLayoutNode.