Struct layout::wrapper::LayoutNode

pub struct LayoutNode<'a> {
    pub chain: PhantomData<&'a ()>,
    // some fields omitted
}

A wrapper so that layout can access only the methods that it should have access to. Layout must only ever see these and must never see instances of LayoutJS.

Fields

chain

Being chained to a PhantomData prevents LayoutNodes from escaping.

Methods

impl<'ln> LayoutNode<'ln>

fn dump(self)

fn flow_debug_id(self) -> usize

fn traverse_preorder(self) -> LayoutTreeIterator<'ln>

fn children(self) -> LayoutNodeChildrenIterator<'ln>

Returns an iterator over this node's children.

fn rev_children(self) -> LayoutNodeReverseChildrenIterator<'ln>

unsafe fn get_jsmanaged<'a>(&'a self) -> &'a LayoutJS<Node>

fn initialize_layout_data(self, chan: LayoutChan)

Resets layout data and styles for the node.

FIXME(pcwalton): Do this as part of fragment building instead of in a traversal.

fn has_children(self) -> bool

fn layout_parent_node(self, shared: &SharedLayoutContext) -> Option<LayoutNode<'ln>>

While doing a reflow, the node at the root has no parent, as far as we're concerned. This method returns None at the reflow root.

fn debug_id(self) -> usize