Struct layout::wrapper::ServoLayoutNode

pub struct ServoLayoutNode<'a> {
    node: LayoutJS<Node>,
    chain: PhantomData<&'a ()>,
}

Fields

node

The wrapped node.

chain

Being chained to a PhantomData prevents LayoutNodes from escaping.

Methods

impl<'ln> ServoLayoutNode<'ln>

fn from_layout_js(n: LayoutJS<Node>) -> ServoLayoutNode<'ln>

unsafe fn new(address: &TrustedNodeAddress) -> ServoLayoutNode

unsafe fn new_with_this_lifetime(&self, node: &LayoutJS<Node>) -> ServoLayoutNode<'ln>

Creates a new layout node with the same lifetime as this layout node.

impl<'ln> ServoLayoutNode<'ln>

fn dump_indent(self, indent: u32)

fn debug_str(self) -> String

fn flow_debug_id(self) -> usize

unsafe fn get_jsmanaged(&self) -> &LayoutJS<Node>

Returns the interior of this node as a LayoutJS. This is highly unsafe for layout to call and as such is marked unsafe.