Struct layout::wrapper::ServoThreadSafeLayoutElement

pub struct ServoThreadSafeLayoutElement<'le> {
    element: &'le Element,
}

A wrapper around elements that ensures layout can only ever access safe properties and cannot race on elements.

Fields

element