Struct layout::wrapper::ThreadSafeLayoutNode

pub struct ThreadSafeLayoutNode<'ln> {
    // some fields omitted
}

A thread-safe version of LayoutNode, used during flow construction. This type of layout node does not allow any parents or siblings of nodes to be accessed, to avoid races.

Methods

impl<'ln> ThreadSafeLayoutNode<'ln>

fn new<'a>(node: &LayoutNode<'a>) -> ThreadSafeLayoutNode<'a>

Creates a new ThreadSafeLayoutNode from the given LayoutNode.

fn debug_id(self) -> usize

fn flow_debug_id(self) -> usize

fn children(&self) -> ThreadSafeLayoutNodeChildrenIterator<'ln>

Returns an iterator over this node's children.

fn as_element(&self) -> ThreadSafeLayoutElement<'ln>

If this is an element, accesses the element data. Fails if this is not an element node.

fn get_pseudo_element_type(&self) -> PseudoElementType

fn get_normal_display(&self) -> T

fn get_before_display(&self) -> T

fn get_after_display(&self) -> T

fn has_before_pseudo(&self) -> bool

fn has_after_pseudo(&self) -> bool

fn borrow_layout_data<'a>(&'a self) -> Ref<'a, Option<LayoutDataWrapper>>

Borrows the layout data immutably. Fails on a conflicting borrow.

TODO(pcwalton): Make this private. It will let us avoid borrow flag checks in some cases.

fn mutate_layout_data<'a>(&'a self) -> RefMut<'a, Option<LayoutDataWrapper>>

Borrows the layout data mutably. Fails on a conflicting borrow.

TODO(pcwalton): Make this private. It will let us avoid borrow flag checks in some cases.

fn traverse_postorder_mut<T: PostorderNodeMutTraversal>(&mut self, traversal: &mut T) -> bool

Traverses the tree in postorder.

TODO(pcwalton): Offer a parallel version with a compatible API.

fn is_ignorable_whitespace(&self) -> bool

fn get_input_value(&self) -> String

fn get_input_size(&self) -> u32

fn get_unsigned_integer_attribute(self, attribute: UnsignedIntegerAttribute) -> Option<u32>

fn restyle_damage(self) -> RestyleDamage

Get the description of how to account for recent style changes. This is a simple bitfield and fine to copy by value.

fn set_restyle_damage(self, damage: RestyleDamage)

Set the restyle damage field.

fn flags(self) -> LayoutDataFlags

Returns the layout data flags for this node.

fn insert_flags(self, new_flags: LayoutDataFlags)

Adds the given flags to this node.

fn remove_flags(self, flags: LayoutDataFlags)

Removes the given flags from this node.

fn is_content(&self) -> bool

Returns true if this node contributes content. This is used in the implementation of empty_cells per CSS 2.1  17.6.1.1.