Trait layout::wrapper::LayoutNode

pub trait LayoutNode: TNode {
    type ConcreteThreadSafeLayoutNode: ThreadSafeLayoutNode;
    fn to_threadsafe(&self) -> Self::ConcreteThreadSafeLayoutNode;
    fn type_id(&self) -> NodeTypeId;
    unsafe fn borrow_layout_data_unchecked(&self) -> Option<*const PrivateLayoutData>;
    fn borrow_layout_data(&self) -> Option<Ref<PrivateLayoutData>>;
    fn mutate_layout_data(&self) -> Option<RefMut<PrivateLayoutData>>;
}

A wrapper so that layout can access only the methods that it should have access to. Layout must only ever see these and must never see instances of LayoutJS.

Associated Types

type ConcreteThreadSafeLayoutNode: ThreadSafeLayoutNode

Required Methods

fn to_threadsafe(&self) -> Self::ConcreteThreadSafeLayoutNode

fn type_id(&self) -> NodeTypeId

Returns the type ID of this node.

unsafe fn borrow_layout_data_unchecked(&self) -> Option<*const PrivateLayoutData>

Similar to borrow_data*, but returns the full PrivateLayoutData rather than only the PrivateStyleData.

fn borrow_layout_data(&self) -> Option<Ref<PrivateLayoutData>>

fn mutate_layout_data(&self) -> Option<RefMut<PrivateLayoutData>>