Trait layout::wrapper::TLayoutNode

pub trait TLayoutNode {
    unsafe fn new_with_this_lifetime(&self, node: &LayoutJS<Node>) -> Self;
    fn type_id(&self) -> Option<NodeTypeId>;
    unsafe fn get_jsmanaged<'a>(&'a self) -> &'a LayoutJS<Node>;
    fn text_content(&self) -> Vec<ContentItem>;
    fn first_child(&self) -> Option<Self>;

    unsafe fn get<'a>(&'a self) -> &'a Node { ... }
    fn node_is_element(&self) -> bool { ... }
    fn node_is_document(&self) -> bool { ... }
    fn image_url(&self) -> Option<Url> { ... }
    fn get_renderer(&self) -> Option<Sender<CanvasMsg>> { ... }
    fn get_canvas_width(&self) -> u32 { ... }
    fn get_canvas_height(&self) -> u32 { ... }
    fn iframe_pipeline_and_subpage_ids(&self) -> (PipelineId, SubpageId) { ... }
}

Allows some convenience methods on generic layout nodes.

Required Methods

unsafe fn new_with_this_lifetime(&self, node: &LayoutJS<Node>) -> Self

Creates a new layout node with the same lifetime as this layout node.

fn type_id(&self) -> Option<NodeTypeId>

Returns the type ID of this node. Fails if this node is borrowed mutably. Returns None if this is a pseudo-element; otherwise, returns Some.

unsafe fn get_jsmanaged<'a>(&'a self) -> &'a LayoutJS<Node>

Returns the interior of this node as a LayoutJS. This is highly unsafe for layout to call and as such is marked unsafe.

fn text_content(&self) -> Vec<ContentItem>

If this is a text node or generated content, copies out its content. If this is not a text node, fails.

FIXME(pcwalton): This might have too much copying and/or allocation. Profile this.

fn first_child(&self) -> Option<Self>

Returns the first child of this node.

Provided Methods

unsafe fn get<'a>(&'a self) -> &'a Node

Returns the interior of this node as a Node. This is highly unsafe for layout to call and as such is marked unsafe.

fn node_is_element(&self) -> bool

fn node_is_document(&self) -> bool

fn image_url(&self) -> Option<Url>

If this is an image element, returns its URL. If this is not an image element, fails.

FIXME(pcwalton): Don't copy URLs.

fn get_renderer(&self) -> Option<Sender<CanvasMsg>>

fn get_canvas_width(&self) -> u32

fn get_canvas_height(&self) -> u32

fn iframe_pipeline_and_subpage_ids(&self) -> (PipelineId, SubpageId)

If this node is an iframe element, returns its pipeline and subpage IDs. If this node is not an iframe element, fails.