Trait layout::wrapper::ThreadSafeLayoutNode

pub trait ThreadSafeLayoutNode: Clone + Copy + Sized + PartialEq {
    type ConcreteThreadSafeLayoutElement: ThreadSafeLayoutElement<ConcreteThreadSafeLayoutNode=Self> + Element<Impl=ServoSelectorImpl>;
    type ChildrenIterator: Iterator<Item=Self> + Sized;
    fn with_pseudo(&self, pseudo: PseudoElementType<Option<T>>) -> Self;
    fn opaque(&self) -> OpaqueNode;
    fn type_id(&self) -> Option<NodeTypeId>;
    fn type_id_without_excluding_pseudo_elements(&self) -> NodeTypeId;
    fn debug_id(self) -> usize;
    fn flow_debug_id(self) -> usize;
    fn children(&self) -> Self::ChildrenIterator;
    fn as_element(&self) -> Self::ConcreteThreadSafeLayoutElement;
    fn get_pseudo_element_type(&self) -> PseudoElementType<Option<T>>;
    fn borrow_layout_data(&self) -> Option<Ref<PrivateLayoutData>>;
    fn mutate_layout_data(&self) -> Option<RefMut<PrivateLayoutData>>;
    fn is_ignorable_whitespace(&self, context: &SharedStyleContext) -> bool;
    fn restyle_damage(self) -> RestyleDamage;
    fn set_restyle_damage(self, damage: RestyleDamage);
    fn flags(self) -> LayoutDataFlags;
    fn can_be_fragmented(&self) -> bool;
    fn text_content(&self) -> TextContent;
    fn selection(&self) -> Option<Range<ByteIndex>>;
    fn image_url(&self) -> Option<Url>;
    fn canvas_data(&self) -> Option<HTMLCanvasData>;
    fn iframe_pipeline_id(&self) -> PipelineId;
    fn get_colspan(&self) -> u32;

    fn is_element_or_elements_pseudo(&self) -> bool { ... }
    fn is_element(&self) -> bool { ... }
    fn get_before_pseudo(&self) -> Option<Self> { ... }
    fn get_after_pseudo(&self) -> Option<Self> { ... }
    fn get_details_summary_pseudo(&self) -> Option<Self> { ... }
    fn get_details_content_pseudo(&self) -> Option<Self> { ... }
    fn style(&self, context: &SharedStyleContext) -> Ref<Arc<ServoComputedValues>> { ... }
    fn resolved_style(&self) -> Ref<Arc<ServoComputedValues>> { ... }
    fn selected_style(&self, _context: &SharedStyleContext) -> Ref<Arc<ServoComputedValues>> { ... }
    fn unstyle(self) { ... }
    fn insert_flags(self, new_flags: LayoutDataFlags) { ... }
    fn remove_flags(self, flags: LayoutDataFlags) { ... }
    fn is_content(&self) -> bool { ... }
    fn layer_id(&self) -> LayerId { ... }
    fn layer_id_for_overflow_scroll(&self) -> LayerId { ... }
}

A thread-safe version of LayoutNode, used during flow construction. This type of layout node does not allow any parents or siblings of nodes to be accessed, to avoid races.

Associated Types

type ConcreteThreadSafeLayoutElement: ThreadSafeLayoutElement<ConcreteThreadSafeLayoutNode=Self> + Element<Impl=ServoSelectorImpl>

type ChildrenIterator: Iterator<Item=Self> + Sized

Required Methods

fn with_pseudo(&self, pseudo: PseudoElementType<Option<T>>) -> Self

Creates a new ThreadSafeLayoutNode for the same LayoutNode with a different pseudo-element type.

fn opaque(&self) -> OpaqueNode

Converts self into an OpaqueNode.

fn type_id(&self) -> Option<NodeTypeId>

Returns the type ID of this node. Returns None if this is a pseudo-element; otherwise, returns Some.

fn type_id_without_excluding_pseudo_elements(&self) -> NodeTypeId

Returns the type ID of this node, without discarding pseudo-elements as type_id does.

fn debug_id(self) -> usize

fn flow_debug_id(self) -> usize

fn children(&self) -> Self::ChildrenIterator

Returns an iterator over this node's children.

fn as_element(&self) -> Self::ConcreteThreadSafeLayoutElement

If this is an element, accesses the element data. Fails if this is not an element node.

fn get_pseudo_element_type(&self) -> PseudoElementType<Option<T>>

fn borrow_layout_data(&self) -> Option<Ref<PrivateLayoutData>>

Borrows the layout data immutably. Fails on a conflicting borrow.

TODO(pcwalton): Make this private. It will let us avoid borrow flag checks in some cases.

fn mutate_layout_data(&self) -> Option<RefMut<PrivateLayoutData>>

Borrows the layout data mutably. Fails on a conflicting borrow.

TODO(pcwalton): Make this private. It will let us avoid borrow flag checks in some cases.

fn is_ignorable_whitespace(&self, context: &SharedStyleContext) -> bool

fn restyle_damage(self) -> RestyleDamage

fn set_restyle_damage(self, damage: RestyleDamage)

fn flags(self) -> LayoutDataFlags

Returns the layout data flags for this node.

fn can_be_fragmented(&self) -> bool

fn text_content(&self) -> TextContent

If this is a text node, generated content, or a form element, copies out its content. Otherwise, panics.

FIXME(pcwalton): This might have too much copying and/or allocation. Profile this.

fn selection(&self) -> Option<Range<ByteIndex>>

If the insertion point is within this node, returns it. Otherwise, returns None.

fn image_url(&self) -> Option<Url>

If this is an image element, returns its URL. If this is not an image element, fails.

FIXME(pcwalton): Don't copy URLs.

fn canvas_data(&self) -> Option<HTMLCanvasData>

fn iframe_pipeline_id(&self) -> PipelineId

If this node is an iframe element, returns its pipeline ID. If this node is not an iframe element, fails.

fn get_colspan(&self) -> u32

Provided Methods

fn is_element_or_elements_pseudo(&self) -> bool

fn is_element(&self) -> bool

fn get_before_pseudo(&self) -> Option<Self>

fn get_after_pseudo(&self) -> Option<Self>

fn get_details_summary_pseudo(&self) -> Option<Self>

fn get_details_content_pseudo(&self) -> Option<Self>

fn style(&self, context: &SharedStyleContext) -> Ref<Arc<ServoComputedValues>>

Returns the style results for the given node. If CSS selector matching has not yet been performed, fails.

Unlike the version on TNode, this handles pseudo-elements.

fn resolved_style(&self) -> Ref<Arc<ServoComputedValues>>

Returns the already resolved style of the node.

This differs from style(ctx) in that if the pseudo-element has not yet been computed it would panic.

This should be used just for querying layout, or when we know the element style is precomputed, not from general layout itself.

fn selected_style(&self, _context: &SharedStyleContext) -> Ref<Arc<ServoComputedValues>>

fn unstyle(self)

Removes the style from this node.

Unlike the version on TNode, this handles pseudo-elements.

fn insert_flags(self, new_flags: LayoutDataFlags)

Adds the given flags to this node.

fn remove_flags(self, flags: LayoutDataFlags)

Removes the given flags from this node.

fn is_content(&self) -> bool

Returns true if this node contributes content. This is used in the implementation of empty_cells per CSS 2.1  17.6.1.1.

fn layer_id(&self) -> LayerId

fn layer_id_for_overflow_scroll(&self) -> LayerId