Module layout::block

Layout for CSS block-level elements.

As a terminology note, the term absolute positioning here refers to elements with position absolute or fixed. The term positioned element refers to elements with position relative, absolute, and fixed. The term containing block (occasionally abbreviated as CB) is the containing block for the current flow, which differs from the static containing block if the flow is absolutely-positioned.

"CSS 2.1" or "CSS 2.2" refers to the editor's draft of the W3C "Cascading Style Sheets Level 2 Revision 2 (CSS 2.2) Specification" available here:

http://dev.w3.org/csswg/css2/

"INTRINSIC" refers to L. David Baron's "More Precise Definitions of Inline Layout and Table Layout" available here:

http://dbaron.org/css/intrinsic/

"CSS-SIZING" refers to the W3C "CSS Intrinsic & Extrinsic Sizing Module Level 3" document available here:

http://dev.w3.org/csswg/css-sizing/

Structs

AbsoluteAssignBSizesTraversal
The real assign-block-sizes traversal for flows with position 'absolute'.

AbsoluteNonReplaced
The different types of Blocks.

AbsoluteReplaced

BSizeConstraintSolution
The solutions for the block-size-and-margins constraint equation.

BlockFlow

BlockFlowFlags

BlockNonReplaced

BlockReplaced

CandidateBSizeIterator
Performs block-size calculations potentially multiple times, taking (assuming an horizontal writing mode) height, min-height, and max-height into account. After each call to next(), the caller must call .try() with the current calculated value of height.

FloatNonReplaced

FloatReplaced

FloatedBlockInfo
Information specific to floated blocks.

ISizeConstraintInput
The inputs for the inline-sizes-and-margins constraint equation.

ISizeConstraintSolution
The solutions for the inline-size-and-margins constraint equation.

InlineBlockNonReplaced

InlineBlockReplaced

Enums

BlockStackingContextType
A stacking context, a pseudo-stacking context, or a non-stacking context.

BlockType

CandidateBSizeIteratorStatus

FormattingContextType

MarginsMayCollapseFlag

Constants

DISPLAY_PORT_SIZE_FACTOR
The number of screens of data we're allowed to generate display lists for in each direction.

IS_ROOT
If this is set, then this block flow is the root flow.

Traits

ISizeAndMarginsComputer

Functions

translate_including_floats