Struct layout::block::BlockFlow

pub struct BlockFlow {
    pub base: BaseFlow,
    pub fragment: Fragment,
    inline_size_of_preceding_left_floats: Au,
    inline_size_of_preceding_right_floats: Au,
    pub float: Option<Box<FloatedBlockInfo>>,
    flags: BlockFlowFlags,
}

Fields

base

Data common to all flows.

fragment

The associated fragment.

inline_size_of_preceding_left_floats

The sum of the inline-sizes of all logically left floats that precede this block. This is used to speculatively lay out block formatting contexts.

inline_size_of_preceding_right_floats

The sum of the inline-sizes of all logically right floats that precede this block. This is used to speculatively lay out block formatting contexts.

float

Additional floating flow members.

flags

Various flags.

Methods

impl BlockFlow

fn from_fragment(fragment: Fragment, float_kind: Option<FloatKind>) -> BlockFlow

fn block_type(&self) -> BlockType

Return the type of this block.

This determines the algorithm used to calculate inline-size, block-size, and the relevant margins for this Block.

fn compute_used_inline_size(&mut self, layout_context: &LayoutContext, containing_block_inline_size: Au)

Compute the actual inline size and position for this block.

fn fragment(&mut self) -> &mut Fragment

Return this flow's fragment.

fn containing_block_size(&self, viewport_size: &Size2D<Au>, descendant: OpaqueFlow) -> LogicalSize<Au>

Return the size of the containing block for the given immediate absolute descendant of this flow.

Right now, this only gets the containing block size for absolutely positioned elements. Note: We assume this is called in a top-down traversal, so it is ok to reference the CB.

fn is_replaced_content(&self) -> bool

Return true if this has a replaced fragment.

Text, Images, Inline Block and Canvas (https://html.spec.whatwg.org/multipage/#replaced-elements) fragments are considered as replaced fragments.

fn get_shrink_to_fit_inline_size(&self, available_inline_size: Au) -> Au

Return shrink-to-fit inline-size.

This is where we use the preferred inline-sizes and minimum inline-sizes calculated in the bubble-inline-sizes traversal.

fn adjust_fragments_for_collapsed_margins_if_root<'a>(&mut self, layout_context: &'a LayoutContext<'a>)

If this is the root flow, shifts all kids down and adjusts our size to account for root flow margins, which should never be collapsed according to CSS  8.3.1.

TODO(#2017, pcwalton): This is somewhat inefficient (traverses kids twice); can we do better?

fn clone_with_children(&self, new_children: FlowList) -> BlockFlow

fn assign_block_size_block_base<'a>(&mut self, layout_context: &'a LayoutContext<'a>, fragmentation_context: Option<FragmentationContext>, margins_may_collapse: MarginsMayCollapseFlag) -> Option<FlowRef>

Assign block-size for current flow.

Collapse margins for flow's children and set in-flow child flows' block offsets now that we know their block-sizes.

Calculate and set the block-size of the current flow.

Calculate block-size, vertical margins, and block offset for the flow's box using CSS  10.6.7.

For absolute flows, we store the calculated content block-size for the flow. We defer the calculation of the other values until a later traversal.

When fragmentation_context is given (not None), this should fit as much of the content as possible within the available block size. If there is more content (that doesn’t fit), this flow is fragmented with the extra content moved to another fragment (a flow like this one) which is returrned. See Flow::fragment.

The return value is always None when fragmentation_context is None.

inline(always) because this is only ever called by in-order or non-in-order top-level methods.

fn place_float(&mut self)

Add placement information about current float flow for use by the parent.

Also, use information given by parent about other floats to find out our relative position.

This does not give any information about any float descendants because they do not affect elements outside of the subtree rooted at this float.

This function is called on a kid flow by a parent. Therefore, assign_block_size_float was already called on this kid flow by the traversal function. So, the values used are well-defined.

fn explicit_block_containing_size(&self, layout_context: &LayoutContext) -> Option<Au>

fn explicit_block_size(&self, containing_block_size: Option<Au>) -> Option<Au>

fn calculate_absolute_block_size_and_margins(&mut self, layout_context: &LayoutContext)

fn propagate_and_compute_used_inline_size(&mut self, layout_context: &LayoutContext)

Compute inline size based using the block_container_inline_size set by the parent flow.

This is run in the AssignISizes traversal.

fn propagate_assigned_inline_size_to_children<F>(&mut self, layout_context: &LayoutContext, inline_start_content_edge: Au, inline_end_content_edge: Au, content_inline_size: Au, callback: F) where F: FnMut(&mut Flow, usize, Au, WritingMode, &mut Au, &mut Au)

Assigns the computed inline-start content edge and inline-size to all the children of this block flow. The given callback, if supplied, will be called once per child; it is currently used to push down column sizes for tables.

#[inline(always)] because this is called only from block or table inline-size assignment and the code for block layout is significantly simpler.

fn formatting_context_type(&self) -> FormattingContextType

Determines the type of formatting context this is. See the definition of FormattingContextType.

fn assign_inline_position_for_formatting_context(&mut self)

Per CSS 2.1  9.5, block formatting contexts' inline widths and positions are affected by the presence of floats. This is the part of the assign-heights traversal that computes the final inline position and width for such flows.

Note that this is part of the assign-block-sizes traversal, not the assign-inline-sizes traversal as one might expect. That is because, in general, float placement cannot occur until heights are assigned. To work around this unfortunate circular dependency, by the time we get here we have already estimated the width of the block formatting context based on the floats we could see at the time of inline-size assignment. The job of this function, therefore, is not only to assign the final size but also to perform the layout again for this block formatting context if our speculation was wrong.

FIXME(pcwalton): This code is not incremental-reflow-safe (i.e. not idempotent).

fn is_inline_block(&self) -> bool

fn content_intrinsic_inline_sizes(&self) -> IntrinsicISizes

Computes the content portion (only) of the intrinsic inline sizes of this flow. This is used for calculating shrink-to-fit width. Assumes that intrinsic sizes have already been computed for this flow.

fn bubble_inline_sizes_for_block(&mut self, consult_children: bool)

Computes intrinsic inline sizes for a block.

fn determine_if_layer_needed(&mut self)

fn block_stacking_context_type(&self) -> BlockStackingContextType

fn has_scrolling_overflow(&self) -> bool

fn compute_inline_sizes(&mut self, layout_context: &LayoutContext)

fn definitely_has_zero_block_size(&self) -> bool