Struct layout::block::BlockFlow

#[repr(C)]pub struct BlockFlow {
    pub base: BaseFlow,
    pub fragment: Fragment,
    pub float: Option<Box<FloatedBlockInfo>>,
    flags: BlockFlowFlags,
}

Fields
base: BaseFlow

Data common to all flows.

fragment: Fragment

The associated fragment.

float: Option<Box<FloatedBlockInfo>>

Additional floating flow members.

flags: BlockFlowFlags

Various flags.

Implementations

impl BlockFlow

pub fn from_fragment(fragment: Fragment) -> BlockFlow

pub fn from_fragment_and_float_kind(
fragment: Fragment,
float_kind: Option<FloatKind>
) -> BlockFlow

pub fn block_type(&self) -> BlockType

Return the type of this block.

This determines the algorithm used to calculate inline-size, block-size, and the relevant margins for this Block.

pub fn compute_used_inline_size(
&mut self,
shared_context: &SharedStyleContext,
containing_block_inline_size: Au
)

Compute the actual inline size and position for this block.

pub fn fragment(&mut self) -> &mut Fragment

Return this flow's fragment.

pub fn stacking_relative_border_box(&self, coor: CoordinateSystem) -> Rect<Au>

pub fn containing_block_size(
&self,
viewport_size: &Size2D<Au>,
descendant: OpaqueFlow
) -> LogicalSize<Au>

Return the size of the containing block for the given immediate absolute descendant of this flow.

Right now, this only gets the containing block size for absolutely positioned elements. Note: We assume this is called in a top-down traversal, so it is ok to reference the CB.

pub fn get_shrink_to_fit_inline_size(&self, available_inline_size: Au) -> Au

Return shrink-to-fit inline-size.

This is where we use the preferred inline-sizes and minimum inline-sizes calculated in the bubble-inline-sizes traversal.

fn adjust_fragments_for_collapsed_margins_if_root(
&mut self,
shared_context: &SharedStyleContext
)

If this is the root flow, shifts all kids down and adjusts our size to account for root flow margins, which should never be collapsed according to CSS  8.3.1.

TODO(#2017, pcwalton): This is somewhat inefficient (traverses kids twice); can we do better?

fn clone_with_children(&self, new_children: FlowList) -> BlockFlow

fn propagate_early_absolute_position_info_to_children(&mut self)

Writes in the size of the relative containing block for children. (This information is also needed to handle RTL.)

pub fn assign_block_size_block_base(
&mut self,
layout_context: &LayoutContext,
fragmentation_context: Option<FragmentationContext>,
margins_may_collapse: MarginsMayCollapseFlag
) -> Option<Arc<dyn Flow>>

Assign block-size for current flow.

Collapse margins for flow's children and set in-flow child flows' block offsets now that we know their block-sizes.

Calculate and set the block-size of the current flow.

Calculate block-size, vertical margins, and block offset for the flow's box using CSS  10.6.7.

For absolute flows, we store the calculated content block-size for the flow. We defer the calculation of the other values until a later traversal.

When fragmentation_context is given (not None), this should fit as much of the content as possible within the available block size. If there is more content (that doesn’t fit), this flow is fragmented with the extra content moved to another fragment (a flow like this one) which is returned. See Flow::fragment.

The return value is always None when fragmentation_context is None.

inline(always) because this is only ever called by in-order or non-in-order top-level methods.

pub fn place_float(&mut self)

Add placement information about current float flow for use by the parent.

Also, use information given by parent about other floats to find out our relative position.

This does not give any information about any float descendants because they do not affect elements outside of the subtree rooted at this float.

This function is called on a kid flow by a parent. Therefore, assign_block_size_float was already called on this kid flow by the traversal function. So, the values used are well-defined.

pub fn explicit_block_containing_size(
&self,
shared_context: &SharedStyleContext
) -> Option<Au>

pub fn explicit_block_size(
&self,
containing_block_size: Option<Au>
) -> Option<Au>

fn calculate_absolute_block_size_and_margins(
&mut self,
shared_context: &SharedStyleContext
)

fn propagate_and_compute_used_inline_size(
&mut self,
shared_context: &SharedStyleContext
)

Compute inline size based using the block_container_inline_size set by the parent flow.

This is run in the AssignISizes traversal.

pub fn propagate_assigned_inline_size_to_children<F>(
&mut self,
shared_context: &SharedStyleContext,
inline_start_content_edge: Au,
inline_end_content_edge: Au,
content_inline_size: Au,
callback: F
) where
F: FnMut(&mut dyn Flow, usize, Au, WritingMode, &mut Au, &mut Au),

Assigns the computed inline-start content edge and inline-size to all the children of this block flow. The given callback, if supplied, will be called once per child; it is currently used to push down column sizes for tables.

#[inline(always)] because this is called only from block or table inline-size assignment and the code for block layout is significantly simpler.

pub fn formatting_context_type(&self) -> FormattingContextType

Determines the type of formatting context this is. See the definition of FormattingContextType.

fn assign_inline_position_for_formatting_context(
&mut self,
layout_context: &LayoutContext,
content_box: LogicalRect<Au>
)

Per CSS 2.1  9.5, block formatting contexts' inline widths and positions are affected by the presence of floats. This is the part of the assign-heights traversal that computes the final inline position and width for such flows.

Note that this is part of the assign-block-sizes traversal, not the assign-inline-sizes traversal as one might expect. That is because, in general, float placement cannot occur until heights are assigned. To work around this unfortunate circular dependency, by the time we get here we have already estimated the width of the block formatting context based on the floats we could see at the time of inline-size assignment. The job of this function, therefore, is not only to assign the final size but also to perform the layout again for this block formatting context if our speculation was wrong.

fn is_inline_block_or_inline_flex(&self) -> bool

fn content_intrinsic_inline_sizes(&self) -> IntrinsicISizes

Computes the content portion (only) of the intrinsic inline sizes of this flow. This is used for calculating shrink-to-fit width. Assumes that intrinsic sizes have already been computed for this flow.

pub fn bubble_inline_sizes_for_block(&mut self, consult_children: bool)

Computes intrinsic inline sizes for a block.

pub fn overflow_style_may_require_clip_scroll_node(&self) -> bool

pub fn compute_inline_sizes(&mut self, shared_context: &SharedStyleContext)

pub fn initialize_container_size_for_root(
&mut self,
shared_context: &SharedStyleContext
)

If this is the root flow, initialize values that would normally be set by the parent.

Should be called during assign_inline_sizes for flows that may be the root.

fn guess_inline_size_for_block_formatting_context_if_necessary(&mut self)

fn definitely_has_zero_block_size(&self) -> bool

pub fn is_inline_flex_item(&self) -> bool

pub fn is_block_flex_item(&self) -> bool

pub fn mark_scrolling_overflow(&mut self, has_scrolling_overflow: bool)

pub fn has_scrolling_overflow(&self) -> bool

pub fn sticky_position(&self) -> SideOffsets2D<MaybeAuto>

pub fn background_border_section(&self) -> DisplayListSection

impl BlockFlow

fn transform_clip_to_coordinate_space(
&mut self,
state: &mut StackingContextCollectionState,
preserved_state: &mut SavedStackingContextCollectionState
)

fn is_reference_frame(&self, context_type: Option<StackingContextType>) -> bool

Returns true if this fragment may establish a reference frame and this block creates a stacking context. Both are necessary in order to establish a reference frame.

pub fn collect_stacking_contexts_for_block(
&mut self,
state: &mut StackingContextCollectionState,
flags: StackingContextCollectionFlags
)

fn setup_clipping_for_block(
&mut self,
state: &mut StackingContextCollectionState,
preserved_state: &mut SavedStackingContextCollectionState,
stacking_context_type: Option<StackingContextType>,
established_reference_frame: Option<ClipScrollNodeIndex>,
flags: StackingContextCollectionFlags
) -> ClippingAndScrolling

fn setup_clip_scroll_node_for_position(
&mut self,
state: &mut StackingContextCollectionState,
border_box: Rect<Au>
)

fn setup_clip_scroll_node_for_overflow(
&mut self,
state: &mut StackingContextCollectionState,
border_box: Rect<Au>
)

fn setup_clip_scroll_node_for_css_clip(
&mut self,
state: &mut StackingContextCollectionState,
preserved_state: &mut SavedStackingContextCollectionState,
stacking_relative_border_box: Rect<Au>
)

Adds a scroll root for a block to take the clip property into account per CSS 2.1  11.1.2.

fn create_pseudo_stacking_context_for_block(
&mut self,
stacking_context_type: StackingContextType,
parent_stacking_context_id: StackingContextId,
parent_clipping_and_scrolling: ClippingAndScrolling,
state: &mut StackingContextCollectionState
)

fn create_real_stacking_context_for_block(
&mut self,
parent_stacking_context_id: StackingContextId,
parent_clipping_and_scrolling: ClippingAndScrolling,
established_reference_frame: Option<ClipScrollNodeIndex>,
state: &mut StackingContextCollectionState
)

pub fn build_display_list_for_block_no_damage(
&self,
state: &mut DisplayListBuildState,
border_painting_mode: BorderPaintingMode
)

pub fn build_display_list_for_block(
&mut self,
state: &mut DisplayListBuildState,
border_painting_mode: BorderPaintingMode
)

pub fn build_display_list_for_background_if_applicable_with_background(
&self,
state: &mut DisplayListBuildState,
background: &Background,
background_color: RGBA
)

fn stacking_context_type(
&self,
flags: StackingContextCollectionFlags
) -> Option<StackingContextType>