Struct layout::block::BlockFlow

pub struct BlockFlow {
    pub base: BaseFlow,
    pub fragment: Fragment,
    pub static_b_offset: Au,
    pub float: Option<Box<FloatedBlockInfo>>,
    pub flags: BlockFlowFlags,
    // some fields omitted
}

Fields

base

Data common to all flows.

fragment

The associated fragment.

static_b_offset

Static y offset of an absolute flow from its CB.

float

Additional floating flow members.

flags

Various flags.

Methods

impl BlockFlow

fn from_node_and_fragment(node: &ThreadSafeLayoutNode, fragment: Fragment) -> BlockFlow

fn float_from_node_and_fragment(node: &ThreadSafeLayoutNode, fragment: Fragment, float_kind: FloatKind) -> BlockFlow

fn block_type(&self) -> BlockType

Return the type of this block.

This determines the algorithm used to calculate inline-size, block-size, and the relevant margins for this Block.

fn compute_used_inline_size(&mut self, ctx: &LayoutContext, containing_block_inline_size: Au)

Compute the actual inline size and position for this block.

fn fragment<'a>(&'a mut self) -> &'a mut Fragment

Return this flow's fragment.

fn static_i_offset(&self) -> Au

Return the static x offset from the appropriate Containing Block for this flow.

fn containing_block_size(&mut self, viewport_size: Size2D<Au>) -> LogicalSize<Au>

Return the size of the Containing Block for this flow.

Right now, this only gets the Containing Block size for absolutely positioned elements. Note: Assume this is called in a top-down traversal, so it is ok to reference the CB.

fn get_shrink_to_fit_inline_size(&self, available_inline_size: Au) -> Au

Return shrink-to-fit inline-size.

This is where we use the preferred inline-sizes and minimum inline-sizes calculated in the bubble-inline-sizes traversal.

fn assign_block_size_block_base<'a>(&mut self, layout_context: &'a LayoutContext<'a>, margins_may_collapse: MarginsMayCollapseFlag)

Assign block-size for current flow.

Collapse margins for flow's children and set in-flow child flows' block offsets now that we know their block-sizes.

Calculate and set the block-size of the current flow.

Calculate block-size, vertical margins, and block offset for the flow's box using CSS  10.6.7.

For absolute flows, we store the calculated content block-size for the flow. We defer the calculation of the other values until a later traversal.

inline(always) because this is only ever called by in-order or non-in-order top-level methods.

fn place_float(&mut self)

Add placement information about current float flow for use by the parent.

Also, use information given by parent about other floats to find out our relative position.

This does not give any information about any float descendants because they do not affect elements outside of the subtree rooted at this float.

This function is called on a kid flow by a parent. Therefore, assign_block_size_float was already called on this kid flow by the traversal function. So, the values used are well-defined.

fn propagate_and_compute_used_inline_size(&mut self, layout_context: &LayoutContext)

Compute inline size based using the block_container_inline_size set by the parent flow.

This is run in the AssignISizes traversal.

fn get_hypothetical_block_start_edge(&self) -> Au

Return the block-start outer edge of the hypothetical box for an absolute flow.

This is wrt its parent flow box.

During normal layout assign-block-size, the absolute flow's position is roughly set to its static position (the position it would have had in the normal flow).

fn propagate_assigned_inline_size_to_children(&mut self, layout_context: &LayoutContext, inline_start_content_edge: Au, inline_end_content_edge: Au, content_inline_size: Au, table_info: Option<ChildInlineSizeInfo>)

Assigns the computed inline-start content edge and inline-size to all the children of this block flow. Also computes whether each child will be impacted by floats.

#[inline(always)] because this is called only from block or table inline-size assignment and the code for block layout is significantly simpler.