Module layout::construct

Creates flows and fragments from a DOM tree via a bottom-up, incremental traversal of the DOM.

Each step of the traversal considers the node and existing flow, if there is one. If a node is not dirty and an existing flow exists, then the traversal reuses that flow. Otherwise, it proceeds to construct either a flow or a ConstructionItem. A construction item is a piece of intermediate data that goes with a DOM node and hasn't found its "home" yet-maybe it's a box, maybe it's an absolute or fixed position thing that hasn't found its containing block yet. Construction items bubble up the tree from children to parents until they find their homes.

Structs

FlowConstructor
An object that knows how to create flows.

InlineBlockSplit
Represents an {ib} split that has not yet found the containing block that it belongs to. This is somewhat tricky. An example may be helpful. For this DOM fragment:

InlineFragmentsAccumulator
Holds inline fragments that we're gathering for children of an inline node.

InlineFragmentsConstructionResult
Represents inline fragments and {ib} splits that are bubbling up from an inline.

IntermediateInlineFragments
Holds inline fragments and absolute descendants.

Enums

ConstructionItem
Represents the output of flow construction for a DOM node that has not yet resulted in a complete flow. Construction items bubble up the tree until they find a Flow to be attached to.

ConstructionResult
The results of flow construction for a DOM node.

Traits

ComputedValueUtils
Convenience methods for computed CSS values

FlowConstructionUtils

NodeUtils
A utility trait with some useful methods for node queries.

ObjectElement
Methods for interacting with HTMLObjectElement nodes

Functions

bidi_control_chars
If the 'unicode-bidi' property has a value other than 'normal', return the bidi control codes to inject before and after the text content of the element.

control_chars_to_fragment

strip_ignorable_whitespace_from_end
Strips ignorable whitespace from the end of a list of fragments.

strip_ignorable_whitespace_from_start
Strips ignorable whitespace from the start of a list of fragments.