Enum layout::construct::ConstructionItem

pub enum ConstructionItem {
    InlineFragments(InlineFragmentsConstructionResult),
    Whitespace(OpaqueNode, PseudoElementType<()>, Arc<ServoComputedValues>, RestyleDamage),
    TableColumnFragment(Fragment),
}

Represents the output of flow construction for a DOM node that has not yet resulted in a complete flow. Construction items bubble up the tree until they find a Flow to be attached to.

Variants

InlineFragments

Inline fragments and associated {ib} splits that have not yet found flows.

Whitespace

Potentially ignorable whitespace.

TableColumnFragment

TableColumn Fragment