Struct layout::construct::FlowConstructor

pub struct FlowConstructor<'a> {
    pub layout_context: &'a LayoutContext<'a>,
}

An object that knows how to create flows.

Fields

layout_context

The layout context.

Methods

impl<'a> FlowConstructor<'a>

fn new<'b>(layout_context: &'b LayoutContext<'b>) -> FlowConstructor<'b>

Creates a new flow constructor.

fn repair_if_possible(&mut self, node: &ThreadSafeLayoutNode) -> bool

Attempts to perform incremental repair to account for recent changes to this node. This can fail and return false, indicating that flows will need to be reconstructed.

TODO(pcwalton): Add some more fast paths, like toggling display: none, adding block kids to block parents with no {ib} splits, adding out-of-flow kids, etc.