Struct layout::context::SharedLayoutContext

pub struct SharedLayoutContext {
    pub style_context: SharedStyleContext,
    pub image_cache_thread: ImageCacheThread,
    pub image_cache_sender: Mutex<ImageCacheChan>,
    pub font_cache_thread: Mutex<FontCacheThread>,
    pub url: Url,
    pub visible_rects: Arc<HashMap<LayerId, Rect<Au>, BuildHasherDefault<FnvHasher>>>,
    pub webrender_image_cache: Arc<RwLock<HashMap<(Url, UsePlaceholder), WebRenderImageInfo, BuildHasherDefault<FnvHasher>>>>,
}

Layout information shared among all workers. This must be thread-safe.

Fields

style_context

Bits shared by the layout and style system.

image_cache_thread

The shared image cache thread.

image_cache_sender

A channel for the image cache to send responses to.

font_cache_thread

Interface to the font cache thread.

url

The URL.

visible_rects

The visible rects for each layer, as reported to us by the compositor.

webrender_image_cache

A cache of WebRender image info.