Module layout::display_list

Modules

background

border

builder
Builds display lists from flows and fragments.

conversions

gradient

items
Servo heavily uses display lists, which are retained-mode lists of painting commands to perform. Using a list instead of painting elements in immediate mode allows transforms, hit testing, and invalidation to be performed using the same primitives as painting. It also allows Servo to aggressively cull invisible and out-of-bounds painting elements, to reduce overdraw.

webrender_helpers

Structs

DisplayListBuildState

IndexableText

StackingContextCollectionFlags

StackingContextCollectionState

Enums

BorderPaintingMode
Describes how to paint the borders.

Traits

ToLayout