Module layout::display_list::background

Structs

BackgroundPlacement

Placment information for both image and gradient backgrounds.

Functions

clip

Compute a rounded clip rect for the background.

compute_background_image_size

For a given area and an image compute how big the image should be displayed on the background.

get_cyclic

Access element at index modulo the array length.

placement

Determines where to place an element background image or gradient.

tile_image

Tile an image

tile_image_axis

For either the x or the y axis ajust various values to account for tiling.

tile_image_round

tile_image_spaced