layout::display_list

Module stacking_context

Structs

ContainingBlock 🔒

DebugPrintItem

Refers to one of the child contents or stacking contexts of a StackingContext.

FragmentTextDecoration 🔒

The text decorations for a Fragment, collecting during StackingContextTree construction.

OverflowFrameData 🔒

ReferenceFrameData 🔒

ScrollFrameData 🔒

StackingContext

Either a stacking context or a stacking container, per the definitions in https://drafts.csswg.org/css-position-4/#painting-order.

StackingContextTree 🔒

Enums

DebugPrintField

Refers to one of the vecs of a StackingContext.

StackingContextBuildMode 🔒

StackingContextContent 🔒

A piece of content that directly belongs to a section of a stacking context.

StackingContextSection 🔒

StackingContextType 🔒

Functions

au_rect_to_length_rect 🔒

Type Aliases

ContainingBlockInfo 🔒