layout::display_list

Struct DisplayListBuilder

pub(crate) struct DisplayListBuilder<'a> {
    current_scroll_node_id: ScrollTreeNodeId,
    current_reference_frame_scroll_node_id: ScrollTreeNodeId,
    current_clip_id: ClipId,
    pub webrender_display_list_builder: &'a mut DisplayListBuilder,
    pub paint_info: &'a mut PaintDisplayListInfo,
    inspector_highlight: Option<InspectorHighlight>,
    paint_body_background: bool,
    clip_map: Vec<ClipChainId>,
    image_resolver: Arc<ImageResolver>,
    device_pixel_ratio: Scale<f32, CSSPixel, DevicePixel>,
    lcp_candidate_collector: Option<&'a mut LargestContentfulPaintCandidateCollector>,
}

Fields
current_scroll_node_id: ScrollTreeNodeId

The current ScrollTreeNodeId for this DisplayListBuilder. This allows only passing the builder instead passing the containing stacking_context::StackingContextContent::Fragment as an argument to display list building functions.

current_reference_frame_scroll_node_id: ScrollTreeNodeId

The current ScrollTreeNodeId for this DisplayListBuilder. This is necessary in addition to the Self::current_scroll_node_id, because some pieces of fragments as backgrounds with background-attachment: fixed need to not scroll while the rest of the fragment does.

current_clip_id: ClipId

The current ClipId for this DisplayListBuilder. This allows only passing the builder instead passing the containing stacking_context::StackingContextContent::Fragment as an argument to display list building functions.

webrender_display_list_builder: &'a mut DisplayListBuilder

The wr::DisplayListBuilder for this Servo DisplayListBuilder.

paint_info: &'a mut PaintDisplayListInfo

The PaintDisplayListInfo used to collect display list items and metadata.

inspector_highlight: Option<InspectorHighlight>

Data about the fragments that are highlighted by the inspector, if any.

This data is collected during the traversal of the fragment tree and used to paint the highlight at the very end.

paint_body_background: bool

Whether or not the <body> element should be painted. This is false if the root <html> element inherits the <body>’s background to paint the page canvas background. See https://drafts.csswg.org/css-backgrounds/#body-background.

clip_map: Vec<ClipChainId>

A mapping from ClipId To WebRender ClipChainId used when building this WebRender display list.

image_resolver: Arc<ImageResolver>

An ImageResolver to use during display list construction.

device_pixel_ratio: Scale<f32, CSSPixel, DevicePixel>

The device pixel ratio used for this Document’s display list.

lcp_candidate_collector: Option<&'a mut LargestContentfulPaintCandidateCollector>

The collector for calculating Largest Contentful Paint

Implementations

impl DisplayListBuilder<'_>

pub(crate) fn build( stacking_context_tree: &mut StackingContextTree, fragment_tree: &FragmentTree, image_resolver: Arc<ImageResolver>, device_pixel_ratio: Scale<f32, StyloCSSPixel, StyloDevicePixel>, highlighted_dom_node: Option<OpaqueNode>, debug: &DiagnosticsLogging, lcp_candidate_collector: Option<&mut LargestContentfulPaintCandidateCollector>, ) -> BuiltDisplayList

fn wr(&mut self) -> &mut DisplayListBuilder

fn pipeline_id(&mut self) -> PipelineId

fn mark_is_contentful(&mut self)

fn spatial_id(&self, id: ScrollTreeNodeId) -> SpatialId

fn clip_chain_id(&self, id: ClipId) -> ClipChainId

pub(crate) fn add_all_spatial_nodes(&mut self)

pub(crate) fn add_clip_to_display_list(&mut self, clip: &Clip) -> ClipChainId

Add the given Clip to the WebRender display list and create a mapping from its ClipId to a WebRender ClipChainId. This happens:

When WebRender display list construction starts: All clips created during the StackingContextTree construction are added in one batch. These clips are used for things such as overflow: scroll elements.

When a clip is added during WebRender display list construction for individual items. In that case, this is called by Self::maybe_create_clip.

fn maybe_create_clip( &mut self, radii: BorderRadius, rect: LayoutRect, force_clip_creation: bool, ) -> Option<ClipChainId>

Add a new clip to the WebRender display list being built. This only happens during WebRender display list building and these clips should be added after all clips from the StackingContextTree have already been processed.

fn common_properties( &self, clip_rect: LayoutRect, style: &ComputedValues, ) -> CommonItemProperties

fn paint_dom_inspector_highlight(&mut self)

Draw highlights around the node that is currently hovered in the devtools.

fn collect_lcp_candidate( &mut self, lcp_type: LargestContentfulPaintType, lcp_candidate_id: LCPCandidateID, clip_rect: LayoutRect, bounds: LayoutRect, )