Trait layout::display_list_builder::FragmentDisplayListBuilding

pub trait FragmentDisplayListBuilding {
    fn build_display_list_for_background_if_applicable(&self, style: &ComputedValues, display_list: &mut DisplayList, layout_context: &LayoutContext, level: StackingLevel, absolute_bounds: &Rect<Au>, clip: &ClippingRegion);
    fn compute_background_image_size(&self, style: &ComputedValues, bounds: &Rect<Au>, image: &Image) -> Size2D<Au>;
    fn build_display_list_for_background_image(&self, style: &ComputedValues, display_list: &mut DisplayList, layout_context: &LayoutContext, level: StackingLevel, absolute_bounds: &Rect<Au>, clip: &ClippingRegion, image_url: &Url);
    fn build_display_list_for_background_linear_gradient(&self, display_list: &mut DisplayList, level: StackingLevel, absolute_bounds: &Rect<Au>, clip: &ClippingRegion, gradient: &LinearGradient, style: &ComputedValues);
    fn build_display_list_for_borders_if_applicable(&self, style: &ComputedValues, display_list: &mut DisplayList, abs_bounds: &Rect<Au>, level: StackingLevel, clip: &ClippingRegion);
    fn build_display_list_for_outline_if_applicable(&self, style: &ComputedValues, display_list: &mut DisplayList, bounds: &Rect<Au>, clip: &ClippingRegion);
    fn build_display_list_for_box_shadow_if_applicable(&self, style: &ComputedValues, list: &mut DisplayList, layout_context: &LayoutContext, level: StackingLevel, absolute_bounds: &Rect<Au>, clip: &ClippingRegion);
    fn build_debug_borders_around_text_fragments(&self, style: &ComputedValues, display_list: &mut DisplayList, stacking_relative_border_box: &Rect<Au>, stacking_relative_content_box: &Rect<Au>, text_fragment: &ScannedTextFragmentInfo, clip: &ClippingRegion);
    fn build_debug_borders_around_fragment(&self, display_list: &mut DisplayList, stacking_relative_border_box: &Rect<Au>, clip: &ClippingRegion);
    fn build_display_list(&mut self, display_list: &mut DisplayList, layout_context: &LayoutContext, stacking_relative_flow_origin: &Point2D<Au>, relative_containing_block_size: &LogicalSize<Au>, relative_containing_block_mode: WritingMode, background_and_border_level: BackgroundAndBorderLevel, clip: &ClippingRegion);
    fn finalize_position_and_size_of_iframe(&self, iframe_fragment: &IframeFragmentInfo, offset: Point2D<Au>, layout_context: &LayoutContext);
    fn clipping_region_for_children(&self, current_clip: &ClippingRegion, stacking_relative_border_box: &Rect<Au>) -> ClippingRegion;
    fn calculate_style_specified_clip(&self, parent_clip: &ClippingRegion, stacking_relative_border_box: &Rect<Au>) -> ClippingRegion;
    fn build_display_list_for_text_fragment(&self, display_list: &mut DisplayList, text_fragment: &ScannedTextFragmentInfo, text_color: RGBA, stacking_relative_content_box: &Rect<Au>, shadow_blur_radius: Option<Au>, offset: &Point2D<Au>, clip: &ClippingRegion);
    fn build_display_list_for_text_decoration(&self, display_list: &mut DisplayList, color: &RGBA, stacking_relative_box: &LogicalRect<Au>, clip: &ClippingRegion, blur_radius: Au);
    fn build_fragment_type_specific_display_items(&mut self, display_list: &mut DisplayList, stacking_relative_border_box: &Rect<Au>, clip: &ClippingRegion);
}

Required Methods

fn build_display_list_for_background_if_applicable(&self, style: &ComputedValues, display_list: &mut DisplayList, layout_context: &LayoutContext, level: StackingLevel, absolute_bounds: &Rect<Au>, clip: &ClippingRegion)

Adds the display items necessary to paint the background of this fragment to the display list if necessary.

fn compute_background_image_size(&self, style: &ComputedValues, bounds: &Rect<Au>, image: &Image) -> Size2D<Au>

Computes the background size for an image with the given background area according to the rules in CSS-BACKGROUNDS  3.9.

fn build_display_list_for_background_image(&self, style: &ComputedValues, display_list: &mut DisplayList, layout_context: &LayoutContext, level: StackingLevel, absolute_bounds: &Rect<Au>, clip: &ClippingRegion, image_url: &Url)

Adds the display items necessary to paint the background image of this fragment to the display list at the appropriate stacking level.

fn build_display_list_for_background_linear_gradient(&self, display_list: &mut DisplayList, level: StackingLevel, absolute_bounds: &Rect<Au>, clip: &ClippingRegion, gradient: &LinearGradient, style: &ComputedValues)

Adds the display items necessary to paint the background linear gradient of this fragment to the display list at the appropriate stacking level.

fn build_display_list_for_borders_if_applicable(&self, style: &ComputedValues, display_list: &mut DisplayList, abs_bounds: &Rect<Au>, level: StackingLevel, clip: &ClippingRegion)

Adds the display items necessary to paint the borders of this fragment to a display list if necessary.

fn build_display_list_for_outline_if_applicable(&self, style: &ComputedValues, display_list: &mut DisplayList, bounds: &Rect<Au>, clip: &ClippingRegion)

Adds the display items necessary to paint the outline of this fragment to the display list if necessary.

fn build_display_list_for_box_shadow_if_applicable(&self, style: &ComputedValues, list: &mut DisplayList, layout_context: &LayoutContext, level: StackingLevel, absolute_bounds: &Rect<Au>, clip: &ClippingRegion)

Adds the display items necessary to paint the box shadow of this fragment to the display list if necessary.

fn build_debug_borders_around_text_fragments(&self, style: &ComputedValues, display_list: &mut DisplayList, stacking_relative_border_box: &Rect<Au>, stacking_relative_content_box: &Rect<Au>, text_fragment: &ScannedTextFragmentInfo, clip: &ClippingRegion)

Adds display items necessary to draw debug boxes around a scanned text fragment.

fn build_debug_borders_around_fragment(&self, display_list: &mut DisplayList, stacking_relative_border_box: &Rect<Au>, clip: &ClippingRegion)

Adds display items necessary to draw debug boxes around this fragment.

fn build_display_list(&mut self, display_list: &mut DisplayList, layout_context: &LayoutContext, stacking_relative_flow_origin: &Point2D<Au>, relative_containing_block_size: &LogicalSize<Au>, relative_containing_block_mode: WritingMode, background_and_border_level: BackgroundAndBorderLevel, clip: &ClippingRegion)

Adds the display items for this fragment to the given display list.

Arguments:

display_list: The display list to add display items to.

layout_context: The layout context.

dirty: The dirty rectangle in the coordinate system of the owning flow.

stacking_relative_flow_origin: Position of the origin of the owning flow with respect to its nearest ancestor stacking context.

relative_containing_block_size: The size of the containing block that position: relative makes use of.

clip: The region to clip the display items to.

fn finalize_position_and_size_of_iframe(&self, iframe_fragment: &IframeFragmentInfo, offset: Point2D<Au>, layout_context: &LayoutContext)

Sends the size and position of this iframe fragment to the constellation. This is out of line to guide inlining.

fn clipping_region_for_children(&self, current_clip: &ClippingRegion, stacking_relative_border_box: &Rect<Au>) -> ClippingRegion

Returns the appropriate clipping region for descendants of this flow.

fn calculate_style_specified_clip(&self, parent_clip: &ClippingRegion, stacking_relative_border_box: &Rect<Au>) -> ClippingRegion

Calculates the clipping rectangle for a fragment, taking the clip property into account per CSS 2.1  11.1.2.

fn build_display_list_for_text_fragment(&self, display_list: &mut DisplayList, text_fragment: &ScannedTextFragmentInfo, text_color: RGBA, stacking_relative_content_box: &Rect<Au>, shadow_blur_radius: Option<Au>, offset: &Point2D<Au>, clip: &ClippingRegion)

Creates the text display item for one text fragment. This can be called multiple times for one fragment if there are text shadows.

shadow_blur_radius will be Some if this is a shadow, even if the blur radius is zero.

fn build_display_list_for_text_decoration(&self, display_list: &mut DisplayList, color: &RGBA, stacking_relative_box: &LogicalRect<Au>, clip: &ClippingRegion, blur_radius: Au)

Creates the display item for a text decoration: underline, overline, or line-through.

fn build_fragment_type_specific_display_items(&mut self, display_list: &mut DisplayList, stacking_relative_border_box: &Rect<Au>, clip: &ClippingRegion)

A helper method that build_display_list calls to create per-fragment-type display items.