layout::dom

Trait NodeExt

pub(crate) trait NodeExt<'dom> {

    // Required methods
    fn as_image(&self) -> Option<(Option<Image>, PhysicalSize<f64>)>;
    fn as_canvas(&self) -> Option<(CanvasInfo, PhysicalSize<f64>)>;
    fn as_iframe(&self) -> Option<(PipelineId, BrowsingContextId)>;
    fn as_video(&self) -> Option<(Option<ImageKey>, Option<PhysicalSize<f64>>)>;
    fn as_svg(&self) -> Option<SVGElementData>;
    fn as_typeless_object_with_data_attribute(&self) -> Option<String>;
    fn ensure_inner_layout_data(&self) -> AtomicRefMut<'dom, InnerDOMLayoutData>;
    fn inner_layout_data(&self) -> Option<AtomicRef<'dom, InnerDOMLayoutData>>;
    fn box_slot(&self) -> BoxSlot<'dom>;
    fn unset_all_boxes(&self);
    fn unset_all_pseudo_boxes(&self);
    fn fragments_for_pseudo(
        &self,
        pseudo_element: Option<PseudoElement>,
    ) -> Vec<Fragment>;
    fn with_layout_box_base_including_pseudos(
        &self,
        callback: impl Fn(&LayoutBoxBase),
    );
    fn repair_style(&self, context: &SharedStyleContext<'_>);
    fn take_restyle_damage(&self) -> LayoutDamage;

}

Required Methods

fn as_image(&self) -> Option<(Option<Image>, PhysicalSize<f64>)>

Returns the image if it’s loaded, and its size in image pixels adjusted for image_density.

fn as_canvas(&self) -> Option<(CanvasInfo, PhysicalSize<f64>)>

fn as_iframe(&self) -> Option<(PipelineId, BrowsingContextId)>

fn as_video(&self) -> Option<(Option<ImageKey>, Option<PhysicalSize<f64>>)>

fn as_svg(&self) -> Option<SVGElementData>

fn as_typeless_object_with_data_attribute(&self) -> Option<String>

fn ensure_inner_layout_data(&self) -> AtomicRefMut<'dom, InnerDOMLayoutData>

fn inner_layout_data(&self) -> Option<AtomicRef<'dom, InnerDOMLayoutData>>

fn box_slot(&self) -> BoxSlot<'dom>

fn unset_all_boxes(&self)

Remove boxes for the element itself, and all of its pseudo-element boxes.

fn unset_all_pseudo_boxes(&self)

Remove all pseudo-element boxes for this element.

fn fragments_for_pseudo( &self, pseudo_element: Option<PseudoElement>, ) -> Vec<Fragment>

fn with_layout_box_base_including_pseudos( &self, callback: impl Fn(&LayoutBoxBase), )

fn repair_style(&self, context: &SharedStyleContext<'_>)

fn take_restyle_damage(&self) -> LayoutDamage

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.