layout

Module dom_traversal

Structs

NodeAndStyleInfo ğŸ”’

A data structure used to pass and store related layout information together to avoid having to repeat the same arguments in argument lists.

Enums

Contents ğŸ”’

NonReplacedContents ğŸ”’

PseudoElementContentItem ğŸ”’

Traits

TraversalHandler ğŸ”’

Functions

generate_pseudo_element_content ğŸ”’

https://www.w3.org/TR/CSS2/generate.html#propdef-content

get_quote_from_pair ğŸ”’

traverse_children_of ğŸ”’

traverse_eager_pseudo_element ğŸ”’

traverse_element ğŸ”’

traverse_pseudo_element_contents ğŸ”’