layout::flexbox::geom

Function rect_to_flow_relative

pub(super) fn rect_to_flow_relative<T>(
    flex_axis: FlexAxis,
    main_start_cross_start_sides_are: MainStartCrossStart,
    base_rect_size: FlexRelativeVec2<T>,
    rect: FlexRelativeRect<T>,
) -> LogicalRect<T>
where
    T: Copy + Add<Output = T> + Sub<Output = T>,

The start corner coordinates in both the input rectangle and output rectangle are relative to some “base rectangle” whose size is passed here.