Module layout::floats

Structs

Float

Information about a single float.

FloatList

Information about the floats next to a flow.

Floats

Encapsulates information about floats. This is optimized to avoid allocation if there are no floats, and to avoid copying when translating the list of floats downward.

PlacementInfo

All the information necessary to place a float.

SpeculatedFloatPlacement

The speculated inline sizes of floats flowing through or around a flow (depending on whether the flow is a block formatting context). These speculations are always upper bounds; the actual inline sizes might be less. Note that this implies that a speculated value of zero is a guarantee that there will be no floats on that side.

Enums

ClearType

The kind of clearance: left, right, or both.

FloatKind

The kind of float: left or right.

Functions

range_intersect