Module layout::flow

Servo's experimental layout system builds a tree of Flow and Fragment objects and solves layout constraints to obtain positions and display attributes of tree nodes. Positions are computed in several tree traversals driven by the fundamental data dependencies required by inline and block layout.

Flows are interior nodes in the layout tree and correspond closely to flow contexts in the CSS specification. Flows are responsible for positioning their child flow contexts and fragments. Flows have purpose-specific fields, such as auxiliary line structs, out-of-flow child lists, and so on.

Currently, the important types of flows are:

BlockFlow: A flow that establishes a block context. It has several child flows, each of which are positioned according to block formatting context rules (CSS block boxes). Block flows also contain a single box to represent their rendered borders, padding, etc. The BlockFlow at the root of the tree has special behavior: it stretches to the boundaries of the viewport.

InlineFlow: A flow that establishes an inline context. It has a flat list of child fragments/flows that are subject to inline layout and line breaking and structs to represent line breaks and mapping to CSS boxes, for the purpose of handling getClientRects() and similar methods.

Structs

AbsoluteDescendantInfo
Information about each absolutely-positioned descendant of the given flow.

AbsoluteDescendantIter

AbsoluteDescendants
Absolutely-positioned descendants of this flow.

BaseFlow
Data common to all flows.

ContainingBlockLink
A link to a flow's containing block.

EarlyAbsolutePositionInfo
Information needed to compute absolute (i.e. viewport-relative) flow positions (not to be confused with absolutely-positioned flows) that is computed during block-size assignment.

FlowFlags
Flags used in flows.

FragmentationContext

LateAbsolutePositionInfo
Information needed to compute absolute (i.e. viewport-relative) flow positions (not to be confused with absolutely-positioned flows) that is computed during final position assignment.

OpaqueFlow
A wrapper for the pointer address of a flow. These pointer addresses may only be compared for equality with other such pointer addresses, never dereferenced.

Enums

FlowClass

ForceNonfloatedFlag
Whether a base flow should be forced to be nonfloated. This can affect e.g. TableFlow, which is never floated because the table wrapper flow is the floated one.

Constants

AFFECTS_COUNTERS
Whether this flow has a fragment with counter-reset or counter-increment styles.

BLOCK_POSITION_IS_STATIC
Whether this flow behaves as though it had position: static for the purposes of positioning in the block direction. This is set for flows with position: static and position: relative as well as absolutely-positioned flows with unconstrained positions in the block direction.

CAN_BE_FRAGMENTED
Whether any ancestor is a fragmentation container

CLEARS_LEFT
Whether this flow clears to the left. This is checked all over layout, so a virtual call is too expensive.

CLEARS_RIGHT
Whether this flow clears to the right. This is checked all over layout, so a virtual call is too expensive.

CONTAINS_TEXT_OR_REPLACED_FRAGMENTS
Whether this flow contains any text and/or replaced fragments.

FLOATS_LEFT
Whether this flow is left-floated. This is checked all over layout, so a virtual call is too expensive.

FLOATS_RIGHT
Whether this flow is right-floated. This is checked all over layout, so a virtual call is too expensive.

HAS_COUNTER_AFFECTING_CHILDREN
Whether this flow's descendants have fragments that affect counter-reset or counter-increment styles.

INLINE_POSITION_IS_STATIC
Whether this flow behaves as though it had position: static for the purposes of positioning in the inline direction. This is set for flows with position: static and position: relative as well as absolutely-positioned flows with unconstrained positions in the inline direction.

IS_ABSOLUTELY_POSITIONED
Whether this flow is absolutely positioned. This is checked all over layout, so a virtual call is too expensive.

NEEDS_LAYER
Whether this flow must have its own layer. Even if this flag is not set, it might get its own layer if it's deemed to be likely to overlap flows with their own layer.

TEXT_ALIGN
Text alignment. NB: If you update this, update TEXT_ALIGN_SHIFT below.

Statics

TEXT_ALIGN_SHIFT
The number of bits we must shift off to handle the text alignment field.

Traits

Flow
Virtual methods that make up a float context.

ImmutableFlowUtils

InorderFlowTraversal
An in-order (sequential only) traversal.

MutableFlowUtils

MutableOwnedFlowUtils

PostorderFlowTraversal
A bottom-up traversal, with a optional in-order pass.

PreorderFlowTraversal
A top-down traversal.

Functions

base

child_iter
Iterates over the children of this immutable flow.

child_iter_mut
Iterates over the children of this flow.

mut_base

Type Definitions

AbsoluteDescendantOffsetIter