Module layout::flow

Servo's experimental layout system builds a tree of Flow and Fragment objects and solves layout constraints to obtain positions and display attributes of tree nodes. Positions are computed in several tree traversals driven by the fundamental data dependencies required by

Structs

AbsolutePositionInfo
Information needed to compute absolute (i.e. viewport-relative) flow positions (not to be confused with absolutely-positioned flows).

BaseFlow
Data common to all flows.

ContainingBlockLink
A link to a flow's containing block.

DescendantIter

Descendants
The Descendants of a flow.

FlowFlags
Flags used in flows.

Enums

FlowClass

ForceNonfloatedFlag
Whether a base flow should be forced to be nonfloated. This can affect e.g. TableFlow, which is never floated because the table wrapper flow is the floated one.

Constants

AFFECTS_COUNTERS
Whether this flow has a fragment with counter-reset or counter-increment styles.

CLEARS_LEFT
Whether this flow clears to the left. This is checked all over layout, so a virtual call is too expensive.

CLEARS_RIGHT
Whether this flow clears to the right. This is checked all over layout, so a virtual call is too expensive.

FLOATS_LEFT
Whether this flow is left-floated. This is checked all over layout, so a virtual call is too expensive.

FLOATS_RIGHT
Whether this flow is right-floated. This is checked all over layout, so a virtual call is too expensive.

HAS_COUNTER_AFFECTING_CHILDREN
Whether this flow's descendants have fragments that affect counter-reset or counter-increment styles.

HAS_LEFT_FLOATED_DESCENDANTS
Whether this flow has descendants that float left in the same block formatting context.

HAS_RIGHT_FLOATED_DESCENDANTS
Whether this flow has descendants that float right in the same block formatting context.

IMPACTED_BY_LEFT_FLOATS
Whether this flow is impacted by floats to the left in the same block formatting context (i.e. its height depends on some prior flows with float: left).

IMPACTED_BY_RIGHT_FLOATS
Whether this flow is impacted by floats to the right in the same block formatting context (i.e. its height depends on some prior flows with float: right).

IS_ABSOLUTELY_POSITIONED
Whether this flow is absolutely positioned. This is checked all over layout, so a virtual call is too expensive.

LAYERS_NEEDED_FOR_DESCENDANTS
Whether this flow contains a flow that has its own layer within the same absolute containing block.

NEEDS_LAYER
Whether this flow must have its own layer. Even if this flag is not set, it might get its own layer if it's deemed to be likely to overlap flows with their own layer.

TEXT_ALIGN
Text alignment. NB: If you update this, update TEXT_ALIGN_SHIFT below.

Traits

Flow
Virtual methods that make up a float context.

ImmutableFlowUtils

InorderFlowTraversal
An in-order (sequential only) traversal.

MutableFlowUtils

MutableOwnedFlowUtils

PostorderFlowTraversal
A bottom-up traversal, with a optional in-order pass.

PreorderFlowTraversal
A top-down traversal.

Functions

base

child_iter
Iterates over the children of this flow.

imm_child_iter
Iterates over the children of this immutable flow.

mut_base

Type Definitions

AbsDescendants

DescendantOffsetIter