layout::flow::inline

Module text_run

Structs

TextRun 🔒

A single TextRun for the box tree. These are all descendants of super::InlineBox or the root of the super::InlineFormattingContext. During box tree construction, text is split into TextRuns based on their font, script, etc. When these are created text is already shaped.

TextRunSegment 🔒

TwoCharsAtATimeIterator 🔒

Enums

SegmentStartSoftWrapPolicy 🔒

Constants

XI_LINE_BREAKING_CLASS_CM 🔒

XI_LINE_BREAKING_CLASS_GL 🔒

XI_LINE_BREAKING_CLASS_WJ 🔒

XI_LINE_BREAKING_CLASS_ZW 🔒

XI_LINE_BREAKING_CLASS_ZWJ 🔒

Functions

char_does_not_change_font 🔒

Whether or not this character should be able to change the font during segmentation. Certain character are not rendered at all, so it doesn’t matter what font we use to render them. They should just be added to the current segment.

get_font_for_first_font_for_style 🔒