Struct layout::flow::BaseFlow

pub struct BaseFlow {
    pub restyle_damage: RestyleDamage,
    pub children: FlowList,
    pub intrinsic_inline_sizes: IntrinsicISizes,
    pub position: LogicalRect<Au>,
    pub overflow: Overflow,
    pub parallel: FlowParallelInfo,
    pub floats: Floats,
    pub speculated_float_placement_in: SpeculatedFloatPlacement,
    pub speculated_float_placement_out: SpeculatedFloatPlacement,
    pub collapsible_margins: CollapsibleMargins,
    pub stacking_relative_position: Point2D<Au>,
    pub abs_descendants: AbsoluteDescendants,
    pub block_container_inline_size: Au,
    pub block_container_writing_mode: WritingMode,
    pub block_container_explicit_block_size: Option<Au>,
    pub absolute_cb: ContainingBlockLink,
    pub early_absolute_position_info: EarlyAbsolutePositionInfo,
    pub late_absolute_position_info: LateAbsolutePositionInfo,
    pub clip: ClippingRegion,
    pub stacking_relative_position_of_display_port: Rect<Au>,
    pub writing_mode: WritingMode,
    pub thread_id: u8,
    pub flags: FlowFlags,
    pub stacking_context_id: StackingContextId,
}

Data common to all flows.

Fields

restyle_damage

children

The children of this flow.

intrinsic_inline_sizes

Intrinsic inline sizes for this flow.

position

The upper left corner of the box representing this flow, relative to the box representing its parent flow.

For absolute flows, this represents the position with respect to its containing block.

This does not include margins in the block flow direction, because those can collapse. So for the block direction (usually vertical), this represents the border box. For the inline direction (usually horizontal), this represents the margin box.

overflow

The amount of overflow of this flow, relative to the containing block. Must include all the pixels of all the display list items for correct invalidation.

parallel

Data used during parallel traversals.

TODO(pcwalton): Group with other transient data to save space.

floats

The floats next to this flow.

speculated_float_placement_in

Metrics for floats in computed during the float metrics speculation phase.

speculated_float_placement_out

Metrics for floats out computed during the float metrics speculation phase.

collapsible_margins

The collapsible margins for this flow, if any.

stacking_relative_position

The position of this flow relative to the start of the nearest ancestor stacking context. This is computed during the top-down pass of display list construction.

abs_descendants

Details about descendants with position 'absolute' or 'fixed' for which we are the containing block. This is in tree order. This includes any direct children.

block_container_inline_size

The inline-size of the block container of this flow. Used for computing percentage and automatic values for width.

block_container_writing_mode

The writing mode of the block container of this flow.

FIXME (mbrubeck): Combine this and block_container_inline_size and maybe block_container_explicit_block_size into a struct, to guarantee they are set at the same time? Or just store a link to the containing block flow.

block_container_explicit_block_size

The block-size of the block container of this flow, if it is an explicit size (does not depend on content heights). Used for computing percentage values for height.

absolute_cb

Reference to the Containing Block, if this flow is absolutely positioned.

early_absolute_position_info

Information needed to compute absolute (i.e. viewport-relative) flow positions (not to be confused with absolutely-positioned flows) that is computed during block-size assignment.

late_absolute_position_info

Information needed to compute absolute (i.e. viewport-relative) flow positions (not to be confused with absolutely-positioned flows) that is computed during final position assignment.

clip

The clipping region for this flow and its descendants, in layer coordinates.

stacking_relative_position_of_display_port

The stacking-relative position of the display port.

FIXME(pcwalton): This might be faster as an Arc, since this varies only per-stacking-context.

writing_mode

The writing mode for this flow.

thread_id

For debugging and profiling, the identifier of the thread that laid out this fragment.

flags

Various flags for flows, tightly packed to save space.

stacking_context_id

The ID of the StackingContext that contains this flow. This is initialized to 0, but it assigned during the collect_stacking_contexts phase of display list construction.

Methods

impl BaseFlow

fn new(style: Option<&ServoComputedValues>, writing_mode: WritingMode, force_nonfloated: ForceNonfloatedFlag) -> BaseFlow

fn clone_with_children(&self, children: FlowList) -> BaseFlow

Return a new BaseFlow like this one but with the given children list

fn child_iter_mut(&mut self) -> MutFlowListIterator

fn debug_id(&self) -> usize

fn flow_id(&self) -> usize

fn collect_stacking_contexts_for_children(&mut self, parent_id: StackingContextId, contexts: &mut Vec<Box<StackingContext>>)

fn might_have_floats_in(&self) -> bool

fn might_have_floats_out(&self) -> bool