Struct layout::flow::BaseFlow

pub struct BaseFlow {
    pub restyle_damage: RestyleDamage,
    pub children: FlowList,
    pub intrinsic_inline_sizes: IntrinsicISizes,
    pub position: LogicalRect<Au>,
    pub overflow: Rect<Au>,
    pub parallel: FlowParallelInfo,
    pub floats: Floats,
    pub collapsible_margins: CollapsibleMargins,
    pub stacking_relative_position: Point2D<Au>,
    pub abs_descendants: AbsDescendants,
    pub block_container_inline_size: Au,
    pub block_container_writing_mode: WritingMode,
    pub block_container_explicit_block_size: Option<Au>,
    pub absolute_static_i_offset: Au,
    pub fixed_static_i_offset: Au,
    pub absolute_cb: ContainingBlockLink,
    pub absolute_position_info: AbsolutePositionInfo,
    pub clip: ClippingRegion,
    pub display_list_building_result: DisplayListBuildingResult,
    pub writing_mode: WritingMode,
    pub thread_id: u8,
    pub flags: FlowFlags,
    // some fields omitted
}

Data common to all flows.

Fields

restyle_damage

children

The children of this flow.

intrinsic_inline_sizes

Intrinsic inline sizes for this flow.

position

The upper left corner of the box representing this flow, relative to the box representing its parent flow.

For absolute flows, this represents the position with respect to its containing block.

This does not include margins in the block flow direction, because those can collapse. So for the block direction (usually vertical), this represents the border box. For the inline direction (usually horizontal), this represents the margin box.

overflow

The amount of overflow of this flow, relative to the containing block. Must include all the pixels of all the display list items for correct invalidation.

parallel

Data used during parallel traversals.

TODO(pcwalton): Group with other transient data to save space.

floats

The floats next to this flow.

collapsible_margins

The collapsible margins for this flow, if any.

stacking_relative_position

The position of this flow relative to the start of the nearest ancestor stacking context. This is computed during the top-down pass of display list construction.

abs_descendants

Details about descendants with position 'absolute' or 'fixed' for which we are the containing block. This is in tree order. This includes any direct children.

block_container_inline_size

The inline-size of the block container of this flow. Used for computing percentage and automatic values for width.

block_container_writing_mode

The writing mode of the block container of this flow.

FIXME (mbrubeck): Combine this and block_container_inline_size and maybe block_container_explicit_block_size into a struct, to guarantee they are set at the same time? Or just store a link to the containing block flow.

block_container_explicit_block_size

The block-size of the block container of this flow, if it is an explicit size (does not depend on content heights). Used for computing percentage values for height.

absolute_static_i_offset

Offset wrt the nearest positioned ancestor - aka the Containing Block for any absolutely positioned elements.

fixed_static_i_offset

Offset wrt the Initial Containing Block.

absolute_cb

Reference to the Containing Block, if this flow is absolutely positioned.

absolute_position_info

Information needed to compute absolute (i.e. viewport-relative) flow positions (not to be confused with absolutely-positioned flows).

FIXME(pcwalton): Merge with absolute_static_i_offset and fixed_static_i_offset above?

clip

The clipping region for this flow and its descendants, in layer coordinates.

display_list_building_result

The results of display list building for this flow.

writing_mode

The writing mode for this flow.

thread_id

For debugging and profiling, the identifier of the thread that laid out this fragment.

flags

Various flags for flows, tightly packed to save space.

Methods

impl BaseFlow

fn new(node: Option<ThreadSafeLayoutNode>, writing_mode: WritingMode, force_nonfloated: ForceNonfloatedFlag) -> BaseFlow

fn child_iter<'a>(&'a mut self) -> MutFlowListIterator<'a>

unsafe fn strong_ref_count<'a>(&'a self) -> &'a AtomicUsize

unsafe fn weak_ref_count<'a>(&'a self) -> &'a AtomicUsize

fn debug_id(&self) -> usize

fn validate_display_list_geometry(&self)

Ensures that all display list items generated by this flow are within the flow's overflow rect. This should only be used for debugging.