Struct layout::flow::ContainingBlockLink

pub struct ContainingBlockLink {
    link: Option<WeakFlowRef>,
}

A link to a flow's containing block.

This cannot safely be a Flow pointer because this is a pointer up the tree, not down the tree. A pointer up the tree is unsafe during layout because it can be used to access a node with an immutable reference while that same node is being laid out, causing possible iterator invalidation and use-after-free.

FIXME(pcwalton): I think this would be better with a borrow flag instead of unsafe.

Fields
link: Option<WeakFlowRef>

The pointer up to the containing block.

Methods

impl ContainingBlockLink

fn new() -> ContainingBlockLink

fn set(&mut self, link: FlowRef)

fn generated_containing_block_size(&self, for_flow: OpaqueFlow) -> LogicalSize<Au>

fn explicit_block_containing_size(&self, shared_context: &SharedStyleContext) -> Option<Au>