Trait layout::flow::Flow

pub trait Flow: Debug + Sync + Send + 'static {
    fn class(&self) -> FlowClass;
    fn collect_stacking_contexts(&mut self, _parent_id: StackingContextId, _: &mut Vec<Box<StackingContext>>) -> StackingContextId;
    fn build_display_list(&mut self, state: &mut DisplayListBuildState);
    fn compute_overflow(&self) -> Overflow;
    fn iterate_through_fragment_border_boxes(&self, iterator: &mut FragmentBorderBoxIterator, level: i32, stacking_context_position: &Point2D<Au>);
    fn mutate_fragments(&mut self, mutator: &mut FnMut(&mut Fragment));
    fn update_late_computed_inline_position_if_necessary(&mut self, inline_position: Au);
    fn update_late_computed_block_position_if_necessary(&mut self, block_position: Au);
    fn generated_containing_block_size(&self, _: OpaqueFlow) -> LogicalSize<Au>;
    fn repair_style(&mut self, new_style: &Arc<ServoComputedValues>);

    fn as_block(&self) -> &BlockFlow { ... }
    fn as_mut_block(&mut self) -> &mut BlockFlow { ... }
    fn as_inline(&self) -> &InlineFlow { ... }
    fn as_mut_inline(&mut self) -> &mut InlineFlow { ... }
    fn as_mut_table_wrapper(&mut self) -> &mut TableWrapperFlow { ... }
    fn as_table_wrapper(&self) -> &TableWrapperFlow { ... }
    fn as_mut_table(&mut self) -> &mut TableFlow { ... }
    fn as_table(&self) -> &TableFlow { ... }
    fn as_mut_table_colgroup(&mut self) -> &mut TableColGroupFlow { ... }
    fn as_mut_table_rowgroup(&mut self) -> &mut TableRowGroupFlow { ... }
    fn as_table_rowgroup(&self) -> &TableRowGroupFlow { ... }
    fn as_mut_table_row(&mut self) -> &mut TableRowFlow { ... }
    fn as_table_row(&self) -> &TableRowFlow { ... }
    fn as_mut_table_caption(&mut self) -> &mut TableCaptionFlow { ... }
    fn as_mut_table_cell(&mut self) -> &mut TableCellFlow { ... }
    fn as_mut_multicol(&mut self) -> &mut MulticolFlow { ... }
    fn as_table_cell(&self) -> &TableCellFlow { ... }
    fn column_intrinsic_inline_sizes(&mut self) -> &mut Vec<ColumnIntrinsicInlineSize> { ... }
    fn column_computed_inline_sizes(&mut self) -> &mut Vec<ColumnComputedInlineSize> { ... }
    fn bubble_inline_sizes(&mut self) { ... }
    fn assign_inline_sizes(&mut self, _ctx: &LayoutContext) { ... }
    fn assign_block_size<'a>(&mut self, _ctx: &'a LayoutContext<'a>) { ... }
    fn fragment(&mut self, layout_context: &LayoutContext, _fragmentation_context: Option<FragmentationContext>) -> Option<FlowRef> { ... }
    fn place_float_if_applicable<'a>(&mut self, _: &'a LayoutContext<'a>) { ... }
    fn assign_block_size_for_inorder_child_if_necessary<'a>(&mut self, layout_context: &'a LayoutContext<'a>, parent_thread_id: u8) -> bool { ... }
    fn store_overflow(&mut self, _: &LayoutContext) { ... }
    fn compute_absolute_position(&mut self, _: &LayoutContext) { ... }
    fn compute_collapsible_block_start_margin(&mut self, _layout_context: &mut LayoutContext, _margin_collapse_info: &mut MarginCollapseInfo) { ... }
    fn mark_as_root(&mut self) { ... }
    fn is_root(&self) -> bool { ... }
    fn positioning(&self) -> T { ... }
    fn is_fixed(&self) -> bool { ... }
    fn contains_positioned_fragments(&self) -> bool { ... }
    fn contains_relatively_positioned_fragments(&self) -> bool { ... }
    fn is_absolute_containing_block(&self) -> bool { ... }
    fn layer_id(&self) -> LayerId { ... }
    fn layer_id_for_overflow_scroll(&self) -> LayerId { ... }
    fn print_extra_flow_children(&self, _: &mut PrintTree) { ... }
}

Virtual methods that make up a float context.

Note that virtual methods have a cost; we should not overuse them in Servo. Consider adding methods to ImmutableFlowUtils or MutableFlowUtils before adding more methods here.

Required Methods

fn class(&self) -> FlowClass

Returns the class of flow that this is.

fn collect_stacking_contexts(&mut self, _parent_id: StackingContextId, _: &mut Vec<Box<StackingContext>>) -> StackingContextId

fn build_display_list(&mut self, state: &mut DisplayListBuildState)

Phase 5 of reflow: builds display lists.

fn compute_overflow(&self) -> Overflow

Returns the union of all overflow rects of all of this flow's fragments.

fn iterate_through_fragment_border_boxes(&self, iterator: &mut FragmentBorderBoxIterator, level: i32, stacking_context_position: &Point2D<Au>)

Iterates through border boxes of all of this flow's fragments. Level provides a zero based index indicating the current depth of the flow tree during fragment iteration.

fn mutate_fragments(&mut self, mutator: &mut FnMut(&mut Fragment))

Mutably iterates through fragments in this flow.

fn update_late_computed_inline_position_if_necessary(&mut self, inline_position: Au)

Updates the inline position of a child flow during the assign-height traversal. At present, this is only used for absolutely-positioned inline-blocks.

fn update_late_computed_block_position_if_necessary(&mut self, block_position: Au)

Updates the block position of a child flow during the assign-height traversal. At present, this is only used for absolutely-positioned inline-blocks.

fn generated_containing_block_size(&self, _: OpaqueFlow) -> LogicalSize<Au>

Return the size of the containing block generated by this flow for the absolutely- positioned descendant referenced by for_flow. For block flows, this is the padding box.

NB: Do not change this &self to &mut self under any circumstances! It has security implications because this can be called on parents concurrently from descendants!

fn repair_style(&mut self, new_style: &Arc<ServoComputedValues>)

Attempts to perform incremental fixup of this flow by replacing its fragment's style with the new style. This can only succeed if the flow has exactly one fragment.

Provided Methods

fn as_block(&self) -> &BlockFlow

If this is a block flow, returns the underlying object. Fails otherwise.

fn as_mut_block(&mut self) -> &mut BlockFlow

If this is a block flow, returns the underlying object, borrowed mutably. Fails otherwise.

fn as_inline(&self) -> &InlineFlow

If this is an inline flow, returns the underlying object. Fails otherwise.

fn as_mut_inline(&mut self) -> &mut InlineFlow

If this is an inline flow, returns the underlying object, borrowed mutably. Fails otherwise.

fn as_mut_table_wrapper(&mut self) -> &mut TableWrapperFlow

If this is a table wrapper flow, returns the underlying object, borrowed mutably. Fails otherwise.

fn as_table_wrapper(&self) -> &TableWrapperFlow

If this is a table wrapper flow, returns the underlying object. Fails otherwise.

fn as_mut_table(&mut self) -> &mut TableFlow

If this is a table flow, returns the underlying object, borrowed mutably. Fails otherwise.

fn as_table(&self) -> &TableFlow

If this is a table flow, returns the underlying object. Fails otherwise.

fn as_mut_table_colgroup(&mut self) -> &mut TableColGroupFlow

If this is a table colgroup flow, returns the underlying object, borrowed mutably. Fails otherwise.

fn as_mut_table_rowgroup(&mut self) -> &mut TableRowGroupFlow

If this is a table rowgroup flow, returns the underlying object, borrowed mutably. Fails otherwise.

fn as_table_rowgroup(&self) -> &TableRowGroupFlow

If this is a table rowgroup flow, returns the underlying object. Fails otherwise.

fn as_mut_table_row(&mut self) -> &mut TableRowFlow

If this is a table row flow, returns the underlying object, borrowed mutably. Fails otherwise.

fn as_table_row(&self) -> &TableRowFlow

If this is a table row flow, returns the underlying object. Fails otherwise.

fn as_mut_table_caption(&mut self) -> &mut TableCaptionFlow

If this is a table cell flow, returns the underlying object, borrowed mutably. Fails otherwise.

fn as_mut_table_cell(&mut self) -> &mut TableCellFlow

If this is a table cell flow, returns the underlying object, borrowed mutably. Fails otherwise.

fn as_mut_multicol(&mut self) -> &mut MulticolFlow

If this is a multicol flow, returns the underlying object, borrowed mutably. Fails otherwise.

fn as_table_cell(&self) -> &TableCellFlow

If this is a table cell flow, returns the underlying object. Fails otherwise.

fn column_intrinsic_inline_sizes(&mut self) -> &mut Vec<ColumnIntrinsicInlineSize>

If this is a table row, table rowgroup, or table flow, returns column intrinsic inline-sizes. Fails otherwise.

fn column_computed_inline_sizes(&mut self) -> &mut Vec<ColumnComputedInlineSize>

If this is a table row, table rowgroup, or table flow, returns column computed inline-sizes. Fails otherwise.

fn bubble_inline_sizes(&mut self)

Pass 1 of reflow: computes minimum and preferred inline-sizes.

Recursively (bottom-up) determine the flow's minimum and preferred inline-sizes. When called on this flow, all child flows have had their minimum and preferred inline-sizes set. This function must decide minimum/preferred inline-sizes based on its children's inline- sizes and the dimensions of any boxes it is responsible for flowing.

fn assign_inline_sizes(&mut self, _ctx: &LayoutContext)

Pass 2 of reflow: computes inline-size.

fn assign_block_size<'a>(&mut self, _ctx: &'a LayoutContext<'a>)

Pass 3a of reflow: computes block-size.

fn fragment(&mut self, layout_context: &LayoutContext, _fragmentation_context: Option<FragmentationContext>) -> Option<FlowRef>

Like assign_block_size, but is recurses explicitly into descendants. Fit as much content as possible within available_block_size. If that’s not all of it, truncate the contents of self and return a new flow similar to self with the rest of the content.

The default is to make a flow "atomic": it can not be fragmented.

fn place_float_if_applicable<'a>(&mut self, _: &'a LayoutContext<'a>)

If this is a float, places it. The default implementation does nothing.

fn assign_block_size_for_inorder_child_if_necessary<'a>(&mut self, layout_context: &'a LayoutContext<'a>, parent_thread_id: u8) -> bool

Assigns block-sizes in-order; or, if this is a float, places the float. The default implementation simply assigns block-sizes if this flow might have floats in. Returns true if it was determined that this child might have had floats in or false otherwise.

parent_thread_id is the thread ID of the parent. This is used for the layout tinting debug mode; if the block size of this flow was determined by its parent, we should treat it as laid out by its parent.

fn store_overflow(&mut self, _: &LayoutContext)

CSS Section 11.1 This is the union of rectangles of the flows for which we define the Containing Block.

FIXME(pcwalton): This should not be a virtual method, but currently is due to a compiler bug ("the trait Sized is not implemented for self").

Assumption: This is called in a bottom-up traversal, so kids' overflows have already been set. Assumption: Absolute descendants have had their overflow calculated.

fn compute_absolute_position(&mut self, _: &LayoutContext)

Phase 4 of reflow: computes absolute positions.

fn compute_collapsible_block_start_margin(&mut self, _layout_context: &mut LayoutContext, _margin_collapse_info: &mut MarginCollapseInfo)

fn mark_as_root(&mut self)

Marks this flow as the root flow. The default implementation is a no-op.

fn is_root(&self) -> bool

fn positioning(&self) -> T

The 'position' property of this flow.

fn is_fixed(&self) -> bool

Return true if this flow has position 'fixed'.

fn contains_positioned_fragments(&self) -> bool

fn contains_relatively_positioned_fragments(&self) -> bool

fn is_absolute_containing_block(&self) -> bool

Returns true if this is an absolute containing block.

fn layer_id(&self) -> LayerId

Returns a layer ID for the given fragment.

fn layer_id_for_overflow_scroll(&self) -> LayerId

Returns a layer ID for the given fragment.

fn print_extra_flow_children(&self, _: &mut PrintTree)

Print any extra children (such as fragments) contained in this Flow for debugging purposes. Any items inserted into the tree will become children of this flow.