Trait layout::flow::Flow

pub trait Flow: Debug + Sync {
    fn class(&self) -> FlowClass;
    fn build_display_list(&mut self, layout_context: &LayoutContext);
    fn compute_overflow(&self) -> Rect<Au>;
    fn iterate_through_fragment_border_boxes(&self, iterator: &mut FragmentBorderBoxIterator, stacking_context_position: &Point2D<Au>);
    fn mutate_fragments(&mut self, mutator: &mut FnMut(&mut Fragment));
    fn update_late_computed_inline_position_if_necessary(&mut self, inline_position: Au);
    fn update_late_computed_block_position_if_necessary(&mut self, block_position: Au);
    fn repair_style(&mut self, new_style: &Arc<ComputedValues>);

    fn as_immutable_block<'a>(&'a self) -> &'a BlockFlow { ... }
    fn as_block<'a>(&'a mut self) -> &'a mut BlockFlow { ... }
    fn as_immutable_inline<'a>(&'a self) -> &'a InlineFlow { ... }
    fn as_inline<'a>(&'a mut self) -> &'a mut InlineFlow { ... }
    fn as_table_wrapper<'a>(&'a mut self) -> &'a mut TableWrapperFlow { ... }
    fn as_immutable_table_wrapper<'a>(&'a self) -> &'a TableWrapperFlow { ... }
    fn as_table<'a>(&'a mut self) -> &'a mut TableFlow { ... }
    fn as_immutable_table<'a>(&'a self) -> &'a TableFlow { ... }
    fn as_table_colgroup<'a>(&'a mut self) -> &'a mut TableColGroupFlow { ... }
    fn as_table_rowgroup<'a>(&'a mut self) -> &'a mut TableRowGroupFlow { ... }
    fn as_immutable_table_rowgroup<'a>(&'a self) -> &'a TableRowGroupFlow { ... }
    fn as_table_row<'a>(&'a mut self) -> &'a mut TableRowFlow { ... }
    fn as_immutable_table_row<'a>(&'a self) -> &'a TableRowFlow { ... }
    fn as_table_caption<'a>(&'a mut self) -> &'a mut TableCaptionFlow { ... }
    fn as_table_cell<'a>(&'a mut self) -> &'a mut TableCellFlow { ... }
    fn as_immutable_table_cell<'a>(&'a self) -> &'a TableCellFlow { ... }
    fn column_intrinsic_inline_sizes<'a>(&'a mut self) -> &'a mut Vec<ColumnIntrinsicInlineSize> { ... }
    fn column_computed_inline_sizes<'a>(&'a mut self) -> &'a mut Vec<ColumnComputedInlineSize> { ... }
    fn bubble_inline_sizes(&mut self) { ... }
    fn assign_inline_sizes(&mut self, _ctx: &LayoutContext) { ... }
    fn assign_block_size<'a>(&mut self, _ctx: &'a LayoutContext<'a>) { ... }
    fn place_float_if_applicable<'a>(&mut self, _: &'a LayoutContext<'a>) { ... }
    fn assign_block_size_for_inorder_child_if_necessary<'a>(&mut self, layout_context: &'a LayoutContext<'a>, parent_thread_id: u8) -> bool { ... }
    fn compute_absolute_position(&mut self) { ... }
    fn compute_collapsible_block_start_margin(&mut self, _layout_context: &mut LayoutContext, _margin_collapse_info: &mut MarginCollapseInfo) { ... }
    fn mark_as_root(&mut self) { ... }
    fn is_store_overflow_delayed(&mut self) -> bool { ... }
    fn is_root(&self) -> bool { ... }
    fn positioning(&self) -> T { ... }
    fn is_fixed(&self) -> bool { ... }
    fn is_positioned(&self) -> bool { ... }
    fn is_relatively_positioned(&self) -> bool { ... }
    fn is_root_of_absolute_flow_tree(&self) -> bool { ... }
    fn is_absolute_containing_block(&self) -> bool { ... }
    fn generated_containing_block_rect(&self) -> LogicalRect<Au> { ... }
    fn layer_id(&self, fragment_id: u32) -> LayerId { ... }
    fn remove_compositor_layers(&self, _: ConstellationChan) { ... }
}

Virtual methods that make up a float context.

Note that virtual methods have a cost; we should not overuse them in Servo. Consider adding methods to ImmutableFlowUtils or MutableFlowUtils before adding more methods here.

Required Methods

fn class(&self) -> FlowClass

Returns the class of flow that this is.

fn build_display_list(&mut self, layout_context: &LayoutContext)

Phase 5 of reflow: builds display lists.

fn compute_overflow(&self) -> Rect<Au>

Returns the union of all overflow rects of all of this flow's fragments.

fn iterate_through_fragment_border_boxes(&self, iterator: &mut FragmentBorderBoxIterator, stacking_context_position: &Point2D<Au>)

Iterates through border boxes of all of this flow's fragments.

fn mutate_fragments(&mut self, mutator: &mut FnMut(&mut Fragment))

Mutably iterates through fragments in this flow.

fn update_late_computed_inline_position_if_necessary(&mut self, inline_position: Au)

Updates the inline position of a child flow during the assign-height traversal. At present, this is only used for absolutely-positioned inline-blocks.

fn update_late_computed_block_position_if_necessary(&mut self, block_position: Au)

Updates the block position of a child flow during the assign-height traversal. At present, this is only used for absolutely-positioned inline-blocks.

fn repair_style(&mut self, new_style: &Arc<ComputedValues>)

Attempts to perform incremental fixup of this flow by replacing its fragment's style with the new style. This can only succeed if the flow has exactly one fragment.

Provided Methods

fn as_immutable_block<'a>(&'a self) -> &'a BlockFlow

If this is a block flow, returns the underlying object, borrowed immutably. Fails otherwise.

fn as_block<'a>(&'a mut self) -> &'a mut BlockFlow

If this is a block flow, returns the underlying object. Fails otherwise.

fn as_immutable_inline<'a>(&'a self) -> &'a InlineFlow

If this is an inline flow, returns the underlying object, borrowed immutably. Fails otherwise.

fn as_inline<'a>(&'a mut self) -> &'a mut InlineFlow

If this is an inline flow, returns the underlying object. Fails otherwise.

fn as_table_wrapper<'a>(&'a mut self) -> &'a mut TableWrapperFlow

If this is a table wrapper flow, returns the underlying object. Fails otherwise.

fn as_immutable_table_wrapper<'a>(&'a self) -> &'a TableWrapperFlow

If this is a table wrapper flow, returns the underlying object, borrowed immutably. Fails otherwise.

fn as_table<'a>(&'a mut self) -> &'a mut TableFlow

If this is a table flow, returns the underlying object. Fails otherwise.

fn as_immutable_table<'a>(&'a self) -> &'a TableFlow

If this is a table flow, returns the underlying object, borrowed immutably. Fails otherwise.

fn as_table_colgroup<'a>(&'a mut self) -> &'a mut TableColGroupFlow

If this is a table colgroup flow, returns the underlying object. Fails otherwise.

fn as_table_rowgroup<'a>(&'a mut self) -> &'a mut TableRowGroupFlow

If this is a table rowgroup flow, returns the underlying object. Fails otherwise.

fn as_immutable_table_rowgroup<'a>(&'a self) -> &'a TableRowGroupFlow

If this is a table rowgroup flow, returns the underlying object, borrowed immutably. Fails otherwise.

fn as_table_row<'a>(&'a mut self) -> &'a mut TableRowFlow

If this is a table row flow, returns the underlying object. Fails otherwise.

fn as_immutable_table_row<'a>(&'a self) -> &'a TableRowFlow

If this is a table row flow, returns the underlying object, borrowed immutably. Fails otherwise.

fn as_table_caption<'a>(&'a mut self) -> &'a mut TableCaptionFlow

If this is a table cell flow, returns the underlying object. Fails otherwise.

fn as_table_cell<'a>(&'a mut self) -> &'a mut TableCellFlow

If this is a table cell flow, returns the underlying object. Fails otherwise.

fn as_immutable_table_cell<'a>(&'a self) -> &'a TableCellFlow

If this is a table cell flow, returns the underlying object, borrowed immutably. Fails otherwise.

fn column_intrinsic_inline_sizes<'a>(&'a mut self) -> &'a mut Vec<ColumnIntrinsicInlineSize>

If this is a table row, table rowgroup, or table flow, returns column intrinsic inline-sizes. Fails otherwise.

fn column_computed_inline_sizes<'a>(&'a mut self) -> &'a mut Vec<ColumnComputedInlineSize>

If this is a table row, table rowgroup, or table flow, returns column computed inline-sizes. Fails otherwise.

fn bubble_inline_sizes(&mut self)

Pass 1 of reflow: computes minimum and preferred inline-sizes.

Recursively (bottom-up) determine the flow's minimum and preferred inline-sizes. When called on this flow, all child flows have had their minimum and preferred inline-sizes set. This function must decide minimum/preferred inline-sizes based on its children's inline- sizes and the dimensions of any boxes it is responsible for flowing.

fn assign_inline_sizes(&mut self, _ctx: &LayoutContext)

Pass 2 of reflow: computes inline-size.

fn assign_block_size<'a>(&mut self, _ctx: &'a LayoutContext<'a>)

Pass 3a of reflow: computes block-size.

fn place_float_if_applicable<'a>(&mut self, _: &'a LayoutContext<'a>)

If this is a float, places it. The default implementation does nothing.

fn assign_block_size_for_inorder_child_if_necessary<'a>(&mut self, layout_context: &'a LayoutContext<'a>, parent_thread_id: u8) -> bool

Assigns block-sizes in-order; or, if this is a float, places the float. The default implementation simply assigns block-sizes if this flow is impacted by floats. Returns true if this child was impacted by floats or false otherwise.

parent_thread_id is the thread ID of the parent. This is used for the layout tinting debug mode; if the block size of this flow was determined by its parent, we should treat it as laid out by its parent.

fn compute_absolute_position(&mut self)

Phase 4 of reflow: computes absolute positions.

fn compute_collapsible_block_start_margin(&mut self, _layout_context: &mut LayoutContext, _margin_collapse_info: &mut MarginCollapseInfo)

fn mark_as_root(&mut self)

Marks this flow as the root flow. The default implementation is a no-op.

fn is_store_overflow_delayed(&mut self) -> bool

Return true if store overflow is delayed for this flow.

Currently happens only for absolutely positioned flows.

fn is_root(&self) -> bool

fn positioning(&self) -> T

The 'position' property of this flow.

fn is_fixed(&self) -> bool

Return true if this flow has position 'fixed'.

fn is_positioned(&self) -> bool

fn is_relatively_positioned(&self) -> bool

fn is_root_of_absolute_flow_tree(&self) -> bool

Return true if this is the root of an absolute flow tree.

fn is_absolute_containing_block(&self) -> bool

Returns true if this is an absolute containing block.

fn generated_containing_block_rect(&self) -> LogicalRect<Au>

Return the dimensions of the containing block generated by this flow for absolutely- positioned descendants. For block flows, this is the padding box.

NB: Do not change this &self to &mut self under any circumstances! It has security implications because this can be called on parents concurrently from descendants!

fn layer_id(&self, fragment_id: u32) -> LayerId

Returns a layer ID for the given fragment.

fn remove_compositor_layers(&self, _: ConstellationChan)

Remove any compositor layers associated with this flow