Trait layout::flow::HasBaseFlow

pub unsafe trait HasBaseFlow { }

This marker trait indicates that a type is a struct with #[repr(C)] whose first field is of type BaseFlow or some type that also implements this trait.

In other words, the memory representation of BaseFlow must be a prefix of the memory representation of types implementing HasBaseFlow.