Trait layout::flow::ImmutableFlowUtils

pub trait ImmutableFlowUtils {
    fn is_block_like(self) -> bool;
    fn is_table(self) -> bool;
    fn is_table_caption(self) -> bool;
    fn is_proper_table_child(self) -> bool;
    fn is_table_row(self) -> bool;
    fn is_table_cell(self) -> bool;
    fn is_table_colgroup(self) -> bool;
    fn is_table_rowgroup(self) -> bool;
    fn is_table_kind(self) -> bool;
    fn need_anonymous_flow(self, child: &Flow) -> bool;
    fn generate_missing_child_flow<N: ThreadSafeLayoutNode>(self, node: &N, ctx: &LayoutContext) -> FlowRef;
    fn contains_roots_of_absolute_flow_tree(&self) -> bool;
    fn is_leaf(self) -> bool;
    fn child_count(self) -> usize;
    fn is_block_container(self) -> bool;
    fn is_block_flow(self) -> bool;
    fn is_inline_flow(self) -> bool;
    fn print(self, title: String);
    fn print_with_tree(self, print_tree: &mut PrintTree);
    fn floats_might_flow_through(self) -> bool;
    fn baseline_offset_of_last_line_box_in_flow(self) -> Option<Au>;
}

Required Methods

fn is_block_like(self) -> bool

Returns true if this flow is a block flow or subclass thereof.

fn is_table(self) -> bool

Returns true if this flow is a table flow.

fn is_table_caption(self) -> bool

Returns true if this flow is a table caption flow.

fn is_proper_table_child(self) -> bool

Returns true if this flow is a proper table child.

fn is_table_row(self) -> bool

Returns true if this flow is a table row flow.

fn is_table_cell(self) -> bool

Returns true if this flow is a table cell flow.

fn is_table_colgroup(self) -> bool

Returns true if this flow is a table colgroup flow.

fn is_table_rowgroup(self) -> bool

Returns true if this flow is a table rowgroup flow.

fn is_table_kind(self) -> bool

Returns true if this flow is one of table-related flows.

fn need_anonymous_flow(self, child: &Flow) -> bool

Returns true if anonymous flow is needed between this flow and child flow.

fn generate_missing_child_flow<N: ThreadSafeLayoutNode>(self, node: &N, ctx: &LayoutContext) -> FlowRef

Generates missing child flow of this flow.

fn contains_roots_of_absolute_flow_tree(&self) -> bool

Returns true if this flow contains fragments that are roots of an absolute flow tree.

fn is_leaf(self) -> bool

Returns true if this flow has no children.

fn child_count(self) -> usize

Returns the number of children that this flow possesses.

fn is_block_container(self) -> bool

Return true if this flow is a Block Container.

fn is_block_flow(self) -> bool

Returns true if this flow is a block flow.

fn is_inline_flow(self) -> bool

Returns true if this flow is an inline flow.

fn print(self, title: String)

Dumps the flow tree for debugging.

fn print_with_tree(self, print_tree: &mut PrintTree)

Dumps the flow tree for debugging into the given PrintTree.

fn floats_might_flow_through(self) -> bool

Returns true if floats might flow through this flow, as determined by the float placement speculation pass.

fn baseline_offset_of_last_line_box_in_flow(self) -> Option<Au>