Trait layout::flow::MutableFlowUtils

pub trait MutableFlowUtils {
    fn traverse_preorder<T: PreorderFlowTraversal>(self, traversal: &T);
    fn traverse_postorder<T: PostorderFlowTraversal>(self, traversal: &T);
    fn traverse_preorder_absolute_flows<T>(&mut self, traversal: &mut T) where T: PreorderFlowTraversal;
    fn traverse_postorder_absolute_flows<T>(&mut self, traversal: &mut T) where T: PostorderFlowTraversal;
    fn repair_style_and_bubble_inline_sizes(self, style: &Arc<ServoComputedValues>);
}

Required Methods

fn traverse_preorder<T: PreorderFlowTraversal>(self, traversal: &T)

Traverses the tree in preorder.

fn traverse_postorder<T: PostorderFlowTraversal>(self, traversal: &T)

Traverses the tree in postorder.

fn traverse_preorder_absolute_flows<T>(&mut self, traversal: &mut T) where T: PreorderFlowTraversal

Traverse the Absolute flow tree in preorder.

Traverse all your direct absolute descendants, who will then traverse their direct absolute descendants.

Return true if the traversal is to continue or false to stop.

fn traverse_postorder_absolute_flows<T>(&mut self, traversal: &mut T) where T: PostorderFlowTraversal

Traverse the Absolute flow tree in postorder.

Return true if the traversal is to continue or false to stop.

fn repair_style_and_bubble_inline_sizes(self, style: &Arc<ServoComputedValues>)

Calls repair_style and bubble_inline_sizes. You should use this method instead of calling them individually, since there is no reason not to perform both operations.