Struct layout::flow_list::FlowList

pub struct FlowList {
    // some fields omitted
}

Methods

impl FlowList

fn front<'a>(&'a self) -> Option<&'a Flow>

Provide a reference to the front element, or None if the list is empty

unsafe fn front_mut<'a>(&'a mut self) -> Option<&'a mut Flow>

Provide a mutable reference to the front element, or None if the list is empty

fn back<'a>(&'a self) -> Option<&'a Flow>

Provide a reference to the back element, or None if the list is empty

unsafe fn back_mut<'a>(&'a mut self) -> Option<&'a mut Flow>

Provide a mutable reference to the back element, or None if the list is empty

fn push_front(&mut self, new_head: FlowRef)

Add an element first in the list

O(1)

fn pop_front(&mut self) -> Option<FlowRef>

Remove the first element and return it, or None if the list is empty

O(1)

fn push_back(&mut self, new_tail: FlowRef)

Add an element last in the list

O(1)

fn new() -> FlowList

Create an empty list

fn iter<'a>(&'a self) -> FlowListIterator<'a>

Provide a forward iterator

fn iter_mut<'a>(&'a mut self) -> MutFlowListIterator<'a>

Provide a forward iterator with mutable references

fn is_empty(&self) -> bool

O(1)

fn len(&self) -> usize

O(1)