Module layout::fragment

The Fragment type, which represents the leaves of the layout tree.

Structs

CanvasFragmentInfo

Fragment
Fragments (struct Fragment) are the leaves of the layout tree. They cannot position themselves. In general, fragments do not have a simple correspondence with CSS fragments in the specification:

IframeFragmentInfo
A fragment that represents an inline frame (iframe). This stores the pipeline ID so that the size of this iframe can be communicated via the constellation to the iframe's own layout task.

ImageFragmentInfo
A fragment that represents a replaced content image and its accompanying borders, shadows, etc.

InlineAbsoluteHypotheticalFragmentInfo
A hypothetical box (see CSS 2.1  10.3.7) for an absolutely-positioned block that was declared with display: inline;.

InlineBlockFragmentInfo
A fragment that represents an inline-block element.

QuantitiesIncludedInIntrinsicInlineSizes

ReplacedImageFragmentInfo

ScannedTextFragmentInfo
A scanned text fragment represents a single run of text with a distinct style. A TextFragment may be split into two or more fragments across line breaks. Several TextFragments may correspond to a single DOM text node. Split text fragments are implemented by referring to subsets of a single TextRun object.

SplitInfo
Describes how to split a fragment. This is used during line breaking as part of the return value of find_split_info_for_inline_size().

SplitOptions

SplitResult
Describes how to split a fragment into two. This contains up to two SplitInfos.

TableColumnFragmentInfo
A fragment that represents a table column.

TruncationResult
Describes how a fragment should be truncated.

UnscannedTextFragmentInfo
Data for an unscanned text fragment. Unscanned text fragments are the results of flow construction that have not yet had their inline-size determined.

Enums

CoordinateSystem
The coordinate system used in stacking_relative_border_box(). See the documentation of that method for details.

GeneratedContentInfo
Information for generated content.

SpecificFragmentInfo
Info specific to the kind of fragment.

Constants

INTRINSIC_INLINE_SIZE_INCLUDES_BORDER

INTRINSIC_INLINE_SIZE_INCLUDES_MARGINS

INTRINSIC_INLINE_SIZE_INCLUDES_PADDING

INTRINSIC_INLINE_SIZE_INCLUDES_SPECIFIED

RETRY_AT_CHARACTER_BOUNDARIES
True if we should attempt to split at character boundaries if this split fails. This is used to implement overflow-wrap: break-word.

STARTS_LINE
True if this is the first fragment on the line.

Traits

FragmentBorderBoxIterator
A top-down fragment border box iteration handler.