Struct layout::fragment::Fragment

pub struct Fragment {
    pub node: OpaqueNode,
    pub style: Arc<ServoComputedValues>,
    pub selected_style: Arc<ServoComputedValues>,
    pub border_box: LogicalRect<Au>,
    pub border_padding: LogicalMargin<Au>,
    pub margin: LogicalMargin<Au>,
    pub specific: SpecificFragmentInfo,
    pub inline_context: Option<InlineFragmentContext>,
    pub restyle_damage: RestyleDamage,
    pub pseudo: PseudoElementType<()>,
    pub flags: FragmentFlags,
    debug_id: DebugId,
    pub stacking_context_id: StackingContextId,
}

Fragments (struct Fragment) are the leaves of the layout tree. They cannot position themselves. In general, fragments do not have a simple correspondence with CSS fragments in the specification:

Several fragments may correspond to the same CSS box or DOM node. For example, a CSS text box broken across two lines is represented by two fragments.

Some CSS fragments are not created at all, such as some anonymous block fragments induced by inline fragments with block-level sibling fragments. In that case, Servo uses an InlineFlow with BlockFlow siblings; the InlineFlow is block-level, but not a block container. It is positioned as if it were a block fragment, but its children are positioned according to inline flow.

A SpecificFragmentInfo::Generic is an empty fragment that contributes only borders, margins, padding, and backgrounds. It is analogous to a CSS nonreplaced content box.

A fragment's type influences how its styles are interpreted during layout. For example, replaced content such as images are resized differently from tables, text, or other content. Different types of fragments may also contain custom data; for example, text fragments contain text.

Do not add fields to this structure unless they're really really mega necessary! Fragments get moved around a lot and thus their size impacts performance of layout quite a bit.

FIXME(#2260, pcwalton): This can be slimmed down some by (at least) moving inline_context to be on InlineFlow only.

Fields
node: OpaqueNode

An opaque reference to the DOM node that this Fragment originates from.

style: Arc<ServoComputedValues>

The CSS style of this fragment.

selected_style: Arc<ServoComputedValues>

The CSS style of this fragment when it's selected

border_box: LogicalRect<Au>

The position of this fragment relative to its owning flow. The size includes padding and border, but not margin.

NB: This does not account for relative positioning. NB: Collapsed borders are not included in this.

border_padding: LogicalMargin<Au>

The sum of border and padding; i.e. the distance from the edge of the border box to the content edge of the fragment.

margin: LogicalMargin<Au>

The margin of the content box.

specific: SpecificFragmentInfo

Info specific to the kind of fragment. Keep this enum small.

inline_context: Option<InlineFragmentContext>

Holds the style context information for fragments that are part of an inline formatting context.

restyle_damage: RestyleDamage

How damaged this fragment is since last reflow.

pseudo: PseudoElementType<()>

The pseudo-element that this fragment represents.

flags: FragmentFlags

Various flags for this fragment.

debug_id: DebugId

A debug ID that is consistent for the life of this fragment (via transform etc). This ID should not be considered stable across multiple layouts or fragment manipulations.

stacking_context_id: StackingContextId

The ID of the StackingContext that contains this fragment. This is initialized to 0, but it assigned during the collect_stacking_contexts phase of display list construction.

Methods

impl Fragment

fn new<N: ThreadSafeLayoutNode>(node: &N, specific: SpecificFragmentInfo, ctx: &LayoutContext) -> Fragment

Constructs a new Fragment instance.

fn from_opaque_node_and_style(node: OpaqueNode, pseudo: PseudoElementType<()>, style: Arc<ServoComputedValues>, selected_style: Arc<ServoComputedValues>, restyle_damage: RestyleDamage, specific: SpecificFragmentInfo) -> Fragment

Constructs a new Fragment instance from an opaque node.

fn transform(&self, size: LogicalSize<Au>, info: SpecificFragmentInfo) -> Fragment

Transforms this fragment into another fragment of the given type, with the given size, preserving all the other data.

fn transform_with_split_info(&self, split: &SplitInfo, text_run: Arc<TextRun>) -> Fragment

Transforms this fragment using the given SplitInfo, preserving all the other data.

fn transform_into_ellipsis(&self, layout_context: &LayoutContext) -> Fragment

Transforms this fragment into an ellipsis fragment, preserving all the other data.

fn restyle_damage(&self) -> RestyleDamage

fn contains_node(&self, node_address: OpaqueNode) -> bool

fn add_inline_context_style(&mut self, node_info: InlineFragmentNodeInfo)

Adds a style to the inline context for this fragment. If the inline context doesn't exist yet, it will be created.

fn quantities_included_in_intrinsic_inline_size(&self) -> QuantitiesIncludedInIntrinsicInlineSizes

Determines which quantities (border/padding/margin/specified) should be included in the intrinsic inline size of this fragment.

fn surrounding_intrinsic_inline_size(&self) -> Au

Returns the portion of the intrinsic inline-size that consists of borders, padding, and/or margins.

FIXME(#2261, pcwalton): This won't work well for inlines: is this OK?

fn style_specified_intrinsic_inline_size(&self) -> IntrinsicISizesContribution

Uses the style only to estimate the intrinsic inline-sizes. These may be modified for text or replaced elements.

fn guess_inline_content_edge_offsets(&self) -> SpeculatedInlineContentEdgeOffsets

Returns a guess as to the distances from the margin edge of this fragment to its content in the inline direction. This will generally be correct unless percentages are involved.

This is used for the float placement speculation logic.

fn calculate_line_height(&self, layout_context: &LayoutContext) -> Au

fn border_width(&self) -> LogicalMargin<Au>

Returns the sum of the inline-sizes of all the borders of this fragment. Note that this can be expensive to compute, so if possible use the border_padding field instead.

fn compute_inline_direction_margins(&mut self, containing_block_inline_size: Au)

Computes the margins in the inline direction from the containing block inline-size and the style. After this call, the inline direction of the margin field will be correct.

Do not use this method if the inline direction margins are to be computed some other way (for example, via constraint solving for blocks).

fn compute_block_direction_margins(&mut self, containing_block_inline_size: Au)

Computes the margins in the block direction from the containing block inline-size and the style. After this call, the block direction of the margin field will be correct.

Do not use this method if the block direction margins are to be computed some other way (for example, via constraint solving for absolutely-positioned flows).

fn compute_border_and_padding(&mut self, containing_block_inline_size: Au, border_collapse: T)

Computes the border and padding in both inline and block directions from the containing block inline-size and the style. After this call, the border_padding field will be correct.

TODO(pcwalton): Remove border_collapse; we can figure it out from our style and specific fragment info.

fn relative_position(&self, containing_block_size: &LogicalSize<Au>) -> LogicalSize<Au>

fn clear(&self) -> Option<ClearType>

Always inline for SCCP.

FIXME(pcwalton): Just replace with the clear type from the style module for speed?

fn style(&self) -> &ServoComputedValues

fn selected_style(&self) -> &ServoComputedValues

fn white_space(&self) -> T

fn text_decoration(&self) -> T

Returns the text decoration of this fragment, according to the style of the nearest ancestor element.

NB: This may not be the actual text decoration, because of the override rules specified in CSS 2.1  16.3.1. Unfortunately, computing this properly doesn't really fit into Servo's model. Therefore, this is a best lower bound approximation, but the end result may actually have the various decoration flags turned on afterward.

fn inline_start_offset(&self) -> Au

Returns the inline-start offset from margin edge to content edge.

FIXME(#2262, pcwalton): I think this method is pretty bogus, because it won't work for inlines.

fn can_split(&self) -> bool

Returns true if this element can be split. This is true for text fragments, unless white-space: pre or white-space: nowrap is set.

fn is_unscanned_generated_content(&self) -> bool

Returns true if and only if this fragment is a generated content fragment.

fn is_scanned_text_fragment(&self) -> bool

Returns true if and only if this is a scanned text fragment.

fn compute_intrinsic_inline_sizes(&mut self) -> IntrinsicISizesContribution

Computes the intrinsic inline-sizes of this fragment.

fn minimum_splittable_inline_size(&self) -> Au

Returns the narrowest inline-size that the first splittable part of this fragment could possibly be split to. (In most cases, this returns the inline-size of the first word in this fragment.)

fn content_inline_size(&self) -> Au

TODO: What exactly does this function return? Why is it Au(0) for SpecificFragmentInfo::Generic?

fn content_box(&self) -> LogicalRect<Au>

Returns the dimensions of the content box.

This is marked #[inline] because it is frequently called when only one or two of the values are needed and that will save computation.

fn calculate_split_position(&self, max_inline_size: Au, starts_line: bool) -> Option<SplitResult>

Attempts to find the split positions of a text fragment so that its inline-size is no more than max_inline_size.

A return value of None indicates that the fragment could not be split. Otherwise the information pertaining to the split is returned. The inline-start and inline-end split information are both optional due to the possibility of them being whitespace.

fn truncate_to_inline_size(&self, max_inline_size: Au) -> Option<TruncationResult>

Truncates this fragment to the given max_inline_size, using a character-based breaking strategy. If no characters could fit, returns None.

fn calculate_split_position_using_breaking_strategy<'a, I>(&self, slice_iterator: I, max_inline_size: Au, flags: SplitOptions) -> Option<SplitResult> where I: Iterator<Item=TextRunSlice<'a>>

A helper method that uses the breaking strategy described by slice_iterator (at present, either natural word breaking or character breaking) to split this fragment.

fn merge_with(&mut self, next_fragment: Fragment)

The opposite of calculate_split_position_using_breaking_strategy: merges this fragment with the next one.

fn reset_text_range_and_inline_size(&mut self)

Restore any whitespace that was stripped from a text fragment, and recompute inline metrics if necessary.

fn assign_replaced_inline_size_if_necessary(&mut self, container_inline_size: Au)

Assigns replaced inline-size, padding, and margins for this fragment only if it is replaced content per CSS 2.1  10.3.2.

fn assign_replaced_block_size_if_necessary(&mut self, containing_block_block_size: Option<Au>)

Assign block-size for this fragment if it is replaced content. The inline-size must have been assigned first.

Ideally, this should follow CSS 2.1  10.6.2.

fn inline_metrics(&self, layout_context: &LayoutContext) -> InlineMetrics

Calculates block-size above baseline, depth below baseline, and ascent for this fragment when used in an inline formatting context. See CSS 2.1  10.8.1.

fn is_hypothetical(&self) -> bool

Returns true if this fragment is a hypothetical box. See CSS 2.1  10.3.7.

fn can_merge_with_fragment(&self, other: &Fragment) -> bool

Returns true if this fragment can merge with another immediately-following fragment or false otherwise.

fn is_primary_fragment(&self) -> bool

Returns true if and only if this is the primary fragment for the fragment's style object (conceptually, though style sharing makes this not really true, of course). The primary fragment is the one that draws backgrounds, borders, etc., and takes borders, padding and margins into account. Every style object has at most one primary fragment.

At present, all fragments are primary fragments except for inline-block and table wrapper fragments. Inline-block fragments are not primary fragments because the corresponding block flow is the primary fragment, while table wrapper fragments are not primary fragments because the corresponding table flow is the primary fragment.

fn update_late_computed_replaced_inline_size_if_necessary(&mut self)

Determines the inline sizes of inline-block fragments. These cannot be fully computed until inline size assignment has run for the child flow: thus it is computed "late", during block size assignment.

fn update_late_computed_inline_position_if_necessary(&mut self)

fn update_late_computed_block_position_if_necessary(&mut self)

fn repair_style(&mut self, new_style: &Arc<ServoComputedValues>)

fn stacking_relative_border_box(&self, stacking_relative_flow_origin: &Point2D<Au>, relative_containing_block_size: &LogicalSize<Au>, relative_containing_block_mode: WritingMode, coordinate_system: CoordinateSystem) -> Rect<Au>

Given the stacking-context-relative position of the containing flow, returns the border box of this fragment relative to the parent stacking context. This takes position: relative into account.

If coordinate_system is Parent, this returns the border box in the parent stacking context's coordinate system. Otherwise, if coordinate_system is Own and this fragment establishes a stacking context itself, this returns a border box anchored at (0, 0). (If this fragment does not establish a stacking context, then it always belongs to its parent stacking context and thus coordinate_system is ignored.)

This is the method you should use for display list construction as well as getBoundingClientRect() and so forth.

fn stacking_relative_content_box(&self, stacking_relative_border_box: &Rect<Au>) -> Rect<Au>

Given the stacking-context-relative border box, returns the stacking-context-relative content box.

fn establishes_stacking_context(&self) -> bool

Returns true if this fragment establishes a new stacking context and false otherwise.

fn effective_z_index(&self) -> i32

fn compute_overflow(&self, flow_size: &Size2D<Au>, relative_containing_block_size: &LogicalSize<Au>) -> Overflow

Computes the overflow rect of this fragment relative to the start of the flow.

fn requires_line_break_afterward_if_wrapping_on_newlines(&self) -> bool

fn strip_leading_whitespace_if_necessary(&mut self) -> WhitespaceStrippingResult

fn strip_trailing_whitespace_if_necessary(&mut self) -> WhitespaceStrippingResult

Returns true if the entire fragment was stripped.

fn inline_styles(&self) -> InlineStyleIterator

fn margin_box_inline_size(&self) -> Au

Returns the inline-size of this fragment's margin box.

fn is_positioned(&self) -> bool

Returns true if this node or any of the nodes within its inline fragment context have non-static position.

fn is_absolutely_positioned(&self) -> bool

Returns true if this node is absolutely positioned.

fn is_inline_absolute(&self) -> bool

fn meld_with_next_inline_fragment(&mut self, next_fragment: &Fragment)

fn meld_with_prev_inline_fragment(&mut self, prev_fragment: &Fragment)

fn fragment_id(&self) -> usize

fn fragment_type(&self) -> FragmentType

fn layer_id(&self) -> LayerId

fn layer_id_for_overflow_scroll(&self) -> LayerId

fn is_vertically_aligned_to_top_or_bottom(&self) -> bool

Returns true if any of the inline styles associated with this fragment have vertical-align set to top or bottom.

fn is_text_or_replaced(&self) -> bool