Struct layout::fragment::IframeFragmentInfo

pub struct IframeFragmentInfo {
    pub pipeline_id: PipelineId,
}

A fragment that represents an inline frame (iframe). This stores the pipeline ID so that the size of this iframe can be communicated via the constellation to the iframe's own layout thread.

Fields

pipeline_id

The pipeline ID of this iframe.

Methods

impl IframeFragmentInfo

fn new<N: ThreadSafeLayoutNode>(node: &N) -> IframeFragmentInfo

Creates the information specific to an iframe fragment.

fn calculate_replaced_inline_size(&self, style: &ServoComputedValues, containing_size: Au) -> Au

fn calculate_replaced_block_size(&self, style: &ServoComputedValues, containing_size: Option<Au>) -> Au

fn calculate_replaced_size(content_size: LengthOrPercentageOrAuto, style_min_size: LengthOrPercentage, style_max_size: LengthOrPercentageOrNone, containing_size: Option<Au>, default_size: Au) -> Au