Struct layout::fragment::ImageFragmentInfo

pub struct ImageFragmentInfo {
    pub replaced_image_fragment_info: ReplacedImageFragmentInfo,
    pub image: ImageHolder<UntrustedNodeAddress>,
}

A fragment that represents a replaced content image and its accompanying borders, shadows, etc.

Fields

replaced_image_fragment_info

The image held within this fragment.

image

Methods

impl ImageFragmentInfo

fn new(node: &ThreadSafeLayoutNode, image_url: Url, local_image_cache: Arc<Mutex<LocalImageCache<UntrustedNodeAddress>>>) -> ImageFragmentInfo

Creates a new image fragment from the given URL and local image cache.

FIXME(pcwalton): The fact that image fragments store the cache in the fragment makes little sense to me.

fn image_inline_size(&mut self) -> Au

Returns the original inline-size of the image.

fn image_block_size(&mut self) -> Au

Returns the original block-size of the image.

fn tile_image(position: &mut Au, size: &mut Au, virtual_position: Au, image_size: u32)

Tile an image