Struct layout::fragment::ScannedTextFragmentInfo

pub struct ScannedTextFragmentInfo {
    pub run: Arc<TextRun>,
    pub content_size: LogicalSize<Au>,
    pub insertion_point: Option<ByteIndex>,
    pub range: Range<ByteIndex>,
    pub range_end_including_stripped_whitespace: ByteIndex,
    pub flags: ScannedTextFlags,
}

A scanned text fragment represents a single run of text with a distinct style. A TextFragment may be split into two or more fragments across line breaks. Several TextFragments may correspond to a single DOM text node. Split text fragments are implemented by referring to subsets of a single TextRun object.

Fields
run: Arc<TextRun>

The text run that this represents.

content_size: LogicalSize<Au>

The intrinsic size of the text fragment.

insertion_point: Option<ByteIndex>

The byte offset of the insertion point, if any.

range: Range<ByteIndex>

The range within the above text run that this represents.

range_end_including_stripped_whitespace: ByteIndex

The endpoint of the above range, including whitespace that was stripped out. This exists so that we can restore the range to its original value (before line breaking occurred) when performing incremental reflow.

flags: ScannedTextFlags
Methods

impl ScannedTextFragmentInfo

fn new(run: Arc<TextRun>, range: Range<ByteIndex>, content_size: LogicalSize<Au>, insertion_point: Option<ByteIndex>, flags: ScannedTextFlags) -> ScannedTextFragmentInfo

Creates the information specific to a scanned text fragment from a range and a text run.

fn text(&self) -> &str

fn requires_line_break_afterward_if_wrapping_on_newlines(&self) -> bool

fn selected(&self) -> bool