Trait layout::fragment::FragmentBorderBoxIterator

pub trait FragmentBorderBoxIterator {
    fn process(&mut self, fragment: &Fragment, level: i32, overflow: &Rect<Au>);

    fn should_process(&mut self, fragment: &Fragment) -> bool;
}

A top-down fragment border box iteration handler.

Required methods

fn process(&mut self, fragment: &Fragment, level: i32, overflow: &Rect<Au>)

The operation to perform.

fn should_process(&mut self, fragment: &Fragment) -> bool

Returns true if this fragment must be processed in-order. If this returns false, we skip the operation for this fragment, but continue processing siblings.