Module layout::generated_content

The generated content assignment phase.

This phase handles CSS counters, quotes, and ordered lists per CSS  12.3-12.5. It cannot be done in parallel and is therefore a sequential pass that runs on as little of the flow tree as possible.

Structs

Counter
A counter per CSS 2.1  12.4.

CounterValue
The value of a counter at a given level.

ResolveGeneratedContent
The generated content resolution traversal.

ResolveGeneratedContentFragmentMutator
The object that mutates the generated content fragments.

Enums

RenderingMode
How a counter value is to be rendered.

Statics

ARABIC_INDIC

BENGALI

CAMBODIAN

CJK_DECIMAL

CJK_EARTHLY_BRANCH

CJK_HEAVENLY_STEM

DECIMAL

DEVANAGARI

GUJARATI

GURMUKHI

HIRAGANA

HIRAGANA_IROHA

KANNADA

KATAKANA

KATAKANA_IROHA

LAO

LOWER_ALPHA

LOWER_GREEK

MALAYALAM

MONGOLIAN

MYANMAR

ORIYA

PERSIAN

TELUGU

THAI

TIBETAN

UPPER_ALPHA

Functions

handle_negative_value
If the system uses a negative sign, handle negative values per CSS-COUNTER-STYLES  2.

push_alphabetic_representation
Pushes the string that represents the value rendered using the given alphabetic system onto the accumulator per CSS-COUNTER-STYLES  3.1.4.

push_numeric_representation
Pushes the string that represents the value rendered using the given numeric system onto the accumulator per CSS-COUNTER-STYLES  3.1.5.

push_representation
Appends string that represents the value rendered using the system appropriate for the given list-style-type onto the given string.

render_text
Creates fragment info for a literal string.

static_representation
Returns the static character that represents the value rendered using the given list-style, if possible.