Struct layout::inline::InlineFlow

pub struct InlineFlow {
    pub base: BaseFlow,
    pub fragments: InlineFragments,
    pub lines: Vec<Line>,
    pub minimum_block_size_above_baseline: Au,
    pub minimum_depth_below_baseline: Au,
    pub first_line_indentation: Au,
}

Flows for inline layout.

Fields
base: BaseFlow

Data common to all flows.

fragments: InlineFragments

A vector of all inline fragments. Several fragments may correspond to one node/element.

lines: Vec<Line>

A vector of ranges into fragments that represents line positions. These ranges are disjoint and are the result of inline layout. This also includes some metadata used for positioning lines.

minimum_block_size_above_baseline: Au

The minimum block-size above the baseline for each line, as specified by the line height and font style.

minimum_depth_below_baseline: Au

The minimum depth below the baseline for each line, as specified by the line height and font style.

first_line_indentation: Au

The amount of indentation to use on the first line. This is determined by our block parent (because percentages are relative to the containing block, and we aren't in a position to compute things relative to our parent's containing block).

Methods

impl InlineFlow

fn from_fragments(fragments: InlineFragments, writing_mode: WritingMode) -> InlineFlow

fn set_inline_fragment_positions(fragments: &mut InlineFragments, line: &Line, line_align: T, indentation: Au, is_last_line: bool)

Sets fragment positions in the inline direction based on alignment for one line. This performs text justification if mandated by the style.

fn justify_inline_fragments(fragments: &mut InlineFragments, line: &Line, slack_inline_size: Au)

Justifies the given set of inline fragments, distributing the slack_inline_size among all of them according to the value of text-justify.

fn set_block_fragment_positions(fragments: &mut InlineFragments, line: &Line, minimum_block_size_above_baseline: Au, minimum_depth_below_baseline: Au, layout_context: &LayoutContext)

Sets final fragment positions in the block direction for one line.

fn compute_minimum_ascent_and_descent(&self, font_context: &mut FontContext, style: &ServoComputedValues) -> (Au, Au)

Computes the minimum ascent and descent for each line. This is done during flow construction.

style is the style of the block.

fn update_restyle_damage(&mut self)

fn containing_block_range_for_flow_surrounding_fragment_at_index(&self, fragment_index: FragmentIndex) -> Range<FragmentIndex>

fn containing_block_range_for_flow(&self, opaque_flow: OpaqueFlow) -> Range<FragmentIndex>

fn baseline_offset_of_last_line(&self) -> Option<Au>