Struct layout::inline::InlineFragments

pub struct InlineFragments {
    pub fragments: Vec<Fragment>,
}

Represents a list of inline fragments, including element ranges.

Fields
fragments: Vec<Fragment>

The fragments themselves.

Methods

impl InlineFragments

fn new() -> InlineFragments

Creates an empty set of inline fragments.

fn len(&self) -> usize

Returns the number of inline fragments.

fn is_empty(&self) -> bool

Returns true if this list contains no fragments and false if it contains at least one fragment.

fn get(&self, index: usize) -> &Fragment

A convenience function to return the fragment at a given index.

fn get_mut(&mut self, index: usize) -> &mut Fragment

A convenience function to return a mutable reference to the fragment at a given index.