Struct layout::inline::InlineMetrics

pub struct InlineMetrics {
    pub block_size_above_baseline: Au,
    pub depth_below_baseline: Au,
    pub ascent: Au,
}

Block-size above the baseline, depth below the baseline, and ascent for a fragment. See CSS 2.1  10.8.1.

Fields

block_size_above_baseline

depth_below_baseline

ascent

Methods

impl InlineMetrics

fn from_font_metrics(font_metrics: &FontMetrics, line_height: Au) -> InlineMetrics

Calculates inline metrics from font metrics and line block-size per CSS 2.1  10.8.1.

fn from_block_height(font_metrics: &FontMetrics, block_height: Au) -> InlineMetrics

Calculates inline metrics from font metrics and line block-size per CSS 2.1  10.8.1.