Module layout::model

Borders, padding, and margins.

Structs

AdjoiningMargins
A collapsible margin. See CSS 2.1  8.3.1.

IntrinsicISizes
Intrinsic inline-sizes, which consist of minimum and preferred.

IntrinsicISizesContribution
The temporary result of the computation of intrinsic inline-sizes.

MarginCollapseInfo

MinMaxConstraint
A min or max constraint

Enums

CollapsibleMargins
Represents the block-start and block-end margins of a flow with collapsible margins. See CSS 2.1  8.3.1.

FinalMarginState

MarginCollapseState

MaybeAuto
Useful helper data type when computing values for blocks and positioned elements.

Traits

ToGfxMatrix

Functions

padding_from_style

specified

specified_border_radius

specified_margin_from_style
Returns the explicitly-specified margin lengths from the given style. Percentage and auto margins are returned as zero.

specified_or_none