Struct layout::model::IntrinsicISizesContribution

pub struct IntrinsicISizesContribution {
    pub content_intrinsic_sizes: IntrinsicISizes,
    pub surrounding_size: Au,
}

The temporary result of the computation of intrinsic inline-sizes.

Fields

content_intrinsic_sizes

Intrinsic sizes for the content only (not counting borders, padding, or margins).

surrounding_size

The inline size of borders and padding, as well as margins if appropriate.

Methods

impl IntrinsicISizesContribution

fn new() -> IntrinsicISizesContribution

Creates and initializes an inline size computation with all sizes set to zero.

fn finish(self) -> IntrinsicISizes

Adds the content intrinsic sizes and the surrounding size together to yield the final intrinsic size computation.

fn union_inline(&mut self, sizes: &IntrinsicISizes)

Updates the computation so that the minimum is the maximum of the current minimum and the given minimum and the preferred is the sum of the current preferred and the given preferred. This is used when laying out fragments in the inline direction.

FIXME(pcwalton): This is incorrect when the inline fragment contains forced line breaks (e.g. <br> or white-space: pre).

fn union_nonbreaking_inline(&mut self, sizes: &IntrinsicISizes)

Updates the computation so that the minimum is the sum of the current minimum and the given minimum and the preferred is the sum of the current preferred and the given preferred. This is used when laying out fragments in the inline direction when white-space is pre or nowrap.

fn union_block(&mut self, sizes: &IntrinsicISizes)

Updates the computation so that the minimum is the maximum of the current minimum and the given minimum and the preferred is the maximum of the current preferred and the given preferred. This can be useful when laying out fragments in the block direction (but note that it does not take floats into account, so BlockFlow does not use it).

This is used when contributing the intrinsic sizes for individual fragments.