Module layout::query

Utilities for querying the layout, as needed by the layout thread.

Structs

CollectingFragmentBorderBoxIterator

FragmentLocatingFragmentIterator

LayoutRPCImpl

MarginRetrievingFragmentBorderBoxIterator

ParentBorderBoxInfo

ParentOffsetBorderBoxIterator

PositionRetrievingFragmentBorderBoxIterator

UnioningFragmentBorderBoxIterator

UnioningFragmentScrollAreaIterator

Enums

MarginPadding

OverflowDirection

PositionProperty

Side

Functions

overflow_direction

process_content_box_request

process_content_boxes_request

process_margin_style_query

process_node_geometry_request

process_node_layer_id_request

process_node_overflow_request

process_node_scroll_area_request

process_offset_parent_query

process_resolved_style_request
Return the resolved value of property for a given (pseudo)element. https://drafts.csswg.org/cssom/#resolved-value