Module layout::query

Utilities for querying the layout, as needed by the layout thread.

Structs

CollectingFragmentBorderBoxIterator

FragmentClientRectQueryIterator

LayoutRPCImpl

LayoutThreadData

Mutable data belonging to the LayoutThread.

MarginRetrievingFragmentBorderBoxIterator

NodeOffsetBoxInfo

ParentBorderBoxInfo

ParentOffsetBorderBoxIterator

PositionRetrievingFragmentBorderBoxIterator

UnioningFragmentBorderBoxIterator

UnioningFragmentScrollAreaIterator

Enums

InnerTextItem

MarginPadding

OverflowDirection

PositionProperty

Side

Functions

create_font_declaration

inner_text_collection_steps

is_last_table_cell

is_last_table_row

overflow_direction

process_client_rect_query

process_content_box_request

process_content_boxes_request

process_element_inner_text_query

process_node_scroll_area_request

https://drafts.csswg.org/cssom-view/#scrolling-area

process_node_scroll_id_request

process_offset_parent_query

process_resolved_font_style_request

process_resolved_style_request

Return the resolved value of property for a given (pseudo)element. https://drafts.csswg.org/cssom/#resolved-value

process_resolved_style_request_internal

The primary resolution logic, which assumes that the element is styled.

resolve_for_declarations