layout

Module sizing

https://drafts.csswg.org/css-sizing/

Structs

ContentSizes 🔒

InlineContentSizesResult 🔒

LazySize 🔒

Represents a size that can’t be fully resolved until the intrinsic size is known. This is useful in the block axis, since the intrinsic size depends on layout, but the other inputs are known beforehand.

LazySizeData 🔒

Sizes 🔒

Enums

IntrinsicSizingMode 🔒

Size 🔒

The possible values accepted by the sizing properties. https://drafts.csswg.org/css-sizing/#sizing-properties

SizeConstraint 🔒

Represents the sizing constraint that the preferred, min and max sizing properties impose on one axis.

Traits

ComputeInlineContentSizes 🔒

Functions

outer_inline 🔒