Module layout::table

CSS table formatting contexts.

Structs

ChildInlineSizeInfo
Inline-size information that we need to push down to table children.

ColumnComputedInlineSize
The actual inline size for each column.

ColumnIntrinsicInlineSize
Information about the intrinsic inline sizes of columns within a table.

InternalTable
Table, TableRowGroup, TableRow, TableCell types. Their inline-sizes are calculated in the same way and do not have margins.

TableFlow
A table flow corresponded to the table's internal table fragment under a table wrapper flow. The properties position, float, and margin-* are used on the table wrapper fragment, not table fragment per CSS 2.1  10.5.

Traits

TableLikeFlow
Encapsulates functionality shared among all table-like flows: for now, tables and table rowgroups.