Module layout::table

CSS table formatting contexts.

Structs

ColumnComputedInlineSize

The actual inline size for each column.

ColumnIntrinsicInlineSize

Information about the intrinsic inline sizes of columns within a table.

ColumnStyle

InternalTable

Table, TableRowGroup, TableRow, TableCell types. Their inline-sizes are calculated in the same way and do not have margins.

MutTableRowAndGroupIterator

Iterator over all the rows of a table, which also provides the Fragment for rowgroups if any

TableCellColumnIndexData

TableCellStyleInfo

TableCellStyleIterator

An iterator over table cells, yielding all relevant style objects for each cell

TableCellStyleIteratorRowInfo

TableFlow

A table flow corresponded to the table’s internal table fragment under a table wrapper flow. The properties position, float, and margin-* are used on the table wrapper fragment, not table fragment per CSS 2.1  10.5.

TableInlineCollapsedBorders

Inline collapsed borders for the table itself.

TableRowAndGroupIterator

Iterator over all the rows of a table, which also provides the Fragment for rowgroups if any

TableRowIterator

Iterator over all the rows of a table

Enums

NextBlockCollapsedBorders

PreviousBlockCollapsedBorders

Traits

TableLikeFlow

Encapsulates functionality shared among all table-like flows: for now, tables and table rowgroups.

VecExt

Functions

perform_border_collapse_for_row

Updates the border styles in the block direction for a single row. This function should only be called if border collapsing is on. It is factored out into a separate function because we process children of rowgroups too.