Function layout::table::perform_border_collapse_for_row

fn perform_border_collapse_for_row(
    child_table_row: &mut TableRowFlow, 
    table_inline_borders: &TableInlineCollapsedBorders, 
    previous_block_borders: PreviousBlockCollapsedBorders, 
    next_block_borders: NextBlockCollapsedBorders<'_>, 
    inline_spacing: &mut Vec<Au>, 
    block_spacing: &mut Vec<Au>
)

Updates the border styles in the block direction for a single row. This function should only be called if border collapsing is on. It is factored out into a separate function because we process children of rowgroups too.