layout::table

Module layout

Structs

CellLayout 🔒

A result of a final or speculative layout of a single cell in the table. Note that this is only done for slots that are not covered by spans or empty.

CellOrColumnOuterSizes 🔒

CellOrTrackMeasure 🔒

ColspanToDistribute 🔒

The inline size constraints provided by a cell that span multiple columns (colspan > 1). These constraints are distributed to the individual columns that make up this cell’s span.

ColumnLayout 🔒

Information stored during the layout of columns.

RowFragmentLayout 🔒

RowGroupFragmentLayout 🔒

RowLayout 🔒

Information stored during the layout of rows.

RowspanToDistribute 🔒

TableAndTrackDimensions 🔒

TableLayout 🔒

A helper struct that performs the layout of the box tree version of a table into the fragment tree version. This implements https://drafts.csswg.org/css-tables/#table-layout-algorithm

Functions

get_size_percentage_contribution 🔒

max_two_optional_percentages 🔒

Type Aliases

CollapsedBorders 🔒