Struct layout::table::ColumnIntrinsicInlineSize

pub struct ColumnIntrinsicInlineSize {
    pub preferred: Au,
    pub minimum_length: Au,
    pub percentage: CSSFloat,
    pub constrained: bool,
}

Information about the intrinsic inline sizes of columns within a table.

During table inline-size bubbling, we might need to store both a percentage constraint and a specific width constraint. For instance, one cell might say that it wants to be 100 pixels wide in the inline direction and another cell might say that it wants to take up 20% of the inline- size of the table. Now because we bubble up these constraints during the bubble-inline-sizes phase of layout, we don't know yet how wide the table is ultimately going to be in the inline direction. As we need to pick the maximum width of all cells for a column (in this case, the maximum of 100 pixels and 20% of the table), the preceding constraint means that we must potentially store both a specified width and a specified percentage, so that the inline-size assignment phase of layout will know which one to pick.

Fields
preferred: Au

The preferred intrinsic inline size.

minimum_length: Au

The largest specified size of this column as a length.

percentage: CSSFloat

The largest specified size of this column as a percentage (width property).

constrained: bool

Whether the column inline size is constrained per INTRINSIC  4.1.

Methods

impl ColumnIntrinsicInlineSize

fn new() -> ColumnIntrinsicInlineSize

Returns a newly-initialized ColumnIntrinsicInlineSize with all fields blank.

fn greatest_percentage(&self, other: &ColumnIntrinsicInlineSize) -> CSSFloat

Returns the higher of the two percentages specified in self and other.