Module layout::table_row

CSS table formatting contexts.

Structs

BorderCollapseInfoForChildTableCell

CellIntrinsicInlineSize

Information about the column inline size and span for each cell.

CollapsedBorder

All aspects of a border that can collapse with adjacent borders. See CSS 2.1  17.6.2.1.

CollapsedBorderSpacingForRow

CollapsedBordersForRow

TableRowFlow

A single row of a table.

TableRowSizeData

Enums

CollapsedBorderProvenance

Where a border style comes from.

Functions

get_spanned_border_size

Given an array of (_, cumulative_border_size), the index of the current row, and the >1 row_span of the cell, calculate the amount of border-spacing spanned by the row. In case the rowspan was larger than required, this will fix it up.

perform_inline_direction_border_collapse_for_row

Performs border-collapse in the inline direction for all the cells’ inside borders in the inline-direction cells and propagates the outside borders (the far left and right) up to the table row. This is done eagerly here so that at least the inline inside border collapse computations can be parallelized across all the rows of the table.

propagate_column_inline_sizes_to_child

Pushes column inline size, incoming rowspan, and border collapse info down to a child.

set_inline_position_of_child_flow

Lay out table cells inline according to the computer column sizes.