Module layout::table_wrapper

CSS tables.

This follows the “More Precise Definitions of Inline Layout and Table Layout” proposal written by L. David Baron (Mozilla) here:

http://dbaron.org/css/intrinsic/

Hereafter this document is referred to as INTRINSIC.

Structs

AbsoluteTable

AutoLayoutCandidateGuess

The layout “guesses” defined in INTRINSIC  4.3.

ExcessInlineSizeDistributionInfo

FloatedTable

IntermediateColumnInlineSize

An intermediate column size assignment.

Table

TableWrapperFlow

A table wrapper flow based on a block formatting context.

Enums

SelectedAutoLayoutCandidateGuess

The CSSFloat member specifies the weight of the smaller of the two guesses, on a scale from 0.0 to 1.0.

TableLayout

Functions

initial_computed_inline_size

Returns the computed inline size of the table wrapper represented by block.

interp

Linearly interpolates between two guesses, as specified by INTRINSIC  4.3.

weight

Computes the weight needed to linearly interpolate middle between two guesses low and high as specified by INTRINSIC  4.3.