Function layout::text::font_metrics_for_style

pub fn font_metrics_for_style(font_context: &mut FontContext, font_style: Arc<FontStyle>) -> FontMetrics

Returns the metrics of the font represented by the given FontStyle, respectively.

#[inline] because often the caller only needs a few fields from the font metrics.