Struct layout::traversal::RecalcStyleAndConstructFlows

pub struct RecalcStyleAndConstructFlows {
    context: LayoutContext,
    driver: TraversalDriver,
}

Fields
context: LayoutContext driver: TraversalDriver
Methods

impl RecalcStyleAndConstructFlows

fn layout_context(&self) -> &LayoutContext

impl RecalcStyleAndConstructFlows

fn new(context: LayoutContext, driver: TraversalDriver) -> Self

Creates a traversal context, taking ownership of the shared layout context.

fn destroy(self) -> LayoutContext

Consumes this traversal context, returning ownership of the shared layout context to the caller.