Struct layout::traversal::RecalcStyleAndConstructFlows

pub struct RecalcStyleAndConstructFlows<'a> {
    context: LayoutContext<'a>,
}

Fields
context: LayoutContext<'a>
Implementations

impl<'a> RecalcStyleAndConstructFlows<'a>

pub fn new(context: LayoutContext<'a>) -> Self

Creates a traversal context, taking ownership of the shared layout context.

pub fn context(&self) -> &LayoutContext<'a>

pub fn destroy(self) -> LayoutContext<'a>

Consumes this traversal context, returning ownership of the shared layout context to the caller.